Raw Simple And Extended Percentage Agreement

Computes raw simple and extended percentage agreement among raters.

Agree(x, tolerance = 0, na.rm = FALSE)

Arguments

x

a data.frame or a \(k \times m\) matrix, k subjects (in rows) m raters (in columns).

tolerance

number of successive rating categories that should be regarded as rater agreement (see details).

na.rm

logical, indicating whether NA values should be stripped before the computation proceeds. If set to TRUE only the complete cases of the ratings will be used. Defaults to FALSE.

Details

Using extended percentage agreement (tolerance != 0) is only possible for numerical values. If tolerance equals 1, for example, raters differing by one scale degree are interpreted as agreeing.

Value

numeric value of coefficient of interrater reliability

The number of finally (potentially after omitting missing values) used subjects and raters are returned as attributes:

subjects

the number of subjects examined.

raters

the number of raters.

Author

Matthias Gamer <m.gamer@uke.uni-hamburg.de>,
some editorial amendments Andri Signorell <andri@signorell.net>

See also

Examples

categ <- c("V", "N", "P")
lvls  <- factor(categ, levels=categ)
rtr1  <- rep(lvls, c(60, 30, 10))
rtr2  <- rep(rep(lvls, nlevels(lvls)), c(53,5,2, 11,14,5, 1,6,3))
rtr3  <- rep(rep(lvls, nlevels(lvls)), c(48,8,3, 15,10,7, 3,4,2))

Agree(cbind(rtr1, rtr2))       # Simple percentage Agreement
#> [1] 0.7
#> attr(,"subjects")
#> [1] 100
#> attr(,"raters")
#> [1] 2
Agree(data.frame(rtr1, rtr2))  # can be a data.frame
#> [1] 0.7
#> attr(,"subjects")
#> [1] 100
#> attr(,"raters")
#> [1] 2
Agree(cbind(rtr1, rtr2, rtr3)) # Simple percentage Agreement
#> [1] 0.6
#> attr(,"subjects")
#> [1] 100
#> attr(,"raters")
#> [1] 3

Agree(cbind(rtr1, rtr2), 1)    # Extended percentage Agreement
#> [1] 0.97
#> attr(,"subjects")
#> [1] 100
#> attr(,"raters")
#> [1] 2