Anderson-Darling Test of Goodness-of-Fit

Performs the Anderson-Darling test of goodness-of-fit to a specified continuous univariate probability distribution.

AndersonDarlingTest(x, null = "punif", ..., nullname)

Arguments

x
numeric vector of data values.

null
a function, or a character string giving the name of a function, to compute the cumulative distribution function for the null distribution.

...
additional arguments for the cumulative distribution function.

nullname
optional character string describing the null distribution.
The default is "uniform distribution".

Details

This command performs the Anderson-Darling test of goodness-of-fit to the distribution specified by the argument null. It is assumed that the values in x are independent and identically distributed random values, with some cumulative distribution function \(F\). The null hypothesis is that \(F\) is the function specified by the argument null, while the alternative hypothesis is that \(F\) is some other function.

Value

An object of class "htest" representing the result of the hypothesis test.

References

Anderson, T.W. and Darling, D.A. (1952) Asymptotic theory of certain 'goodness-of-fit' criteria based on stochastic processes. Annals of Mathematical Statistics 23, 193--212.

Anderson, T.W. and Darling, D.A. (1954) A test of goodness of fit. Journal of the American Statistical Association 49, 765--769.

Marsaglia, G. and Marsaglia, J. (2004) Evaluating the Anderson-Darling Distribution. Journal of Statistical Software 9 (2), 1--5. February 2004. http://www.jstatsoft.org/v09/i02

See also

Examples

x <- rnorm(10, mean=2, sd=1)
AndersonDarlingTest(x, "pnorm", mean=2, sd=1)

#> 
#> 	Anderson-Darling test of goodness-of-fit
#> 	Null hypothesis: Normal distribution
#> 	with parameters mean = 2.000e+00, sd = 1.000e+00
#> 
#> data:  x
#> An = 1.0832, p-value = 0.3147
#>