Atkinson Index - A Measure of Inequality.

The Atkinson index is an inequality measure and is useful in determining which end of the distribution contributed most to the observed inequality.

Atkinson(x, n = rep(1, length(x)), parameter = 0.5, na.rm = FALSE)

Arguments

x
a vector containing at least non-negative elements.

n
a vector of frequencies, must be same length as x.

parameter
parameter of the inequality measure (if set to NULL the default parameter of the respective measure is used).

na.rm
logical. Should missing values be removed? Defaults to FALSE.

Value

the value of the Akinson Index.

Note

This function was previously published as ineq() in the ineq package and has been integrated here without logical changes, but with some extensions for NA-handling and the use of weights.

References

Cowell, F. A. (2000) Measurement of Inequality in Atkinson, A. B. / Bourguignon, F. (Eds): Handbook of Income Distribution. Amsterdam.

Cowell, F. A. (1995) Measuring Inequality Harvester Wheatshef: Prentice Hall.

Marshall, Olkin (1979) Inequalities: Theory of Majorization and Its Applications. New York: Academic Press.

Author

Achim Zeileis <Achim.Zeileis@R-project.org>

See also

Examples

# generate vector (of incomes)
x <- c(541, 1463, 2445, 3438, 4437, 5401, 6392, 8304, 11904, 22261)

# compute Atkinson coefficient with parameter=0.5
Atkinson(x, parameter=0.5)

#> [1] 0.1796591