Function to write a table

CatTable helps printing a table, if is has to be broken into multiple rows. Rowlabels will be repeated after every new break.

CatTable(tab, wcol, nrepchars, width = getOption("width"))

Arguments

tab

the rows of a table to be printed, pasted together in one string with constant columnwidth.

wcol

integer, the width of the columns. All columns must have the same width.

nrepchars

integer, the number of characters to be repeated with every break. This is typically the maximum width of the rowlabels.

width

integer, the width of the whole table. Default is the width of the current command window (getOption("width")).

Author

Andri Signorell <andri@signorell.net>

See also

Examples

options(scipen=8)

# used in bivariate description functions
Desc(temperature ~ cut(delivery_min, breaks=40), data=d.pizza)
#> Error in .Call("rskew", as.numeric(x), as.numeric(mean(x)), PACKAGE = "DescTools"): "rskew" not available for .Call() for package "DescTools"

txt <- c( 
   paste(sample(letters, 500, replace=TRUE), collapse="")
 , paste(sample(letters, 500, replace=TRUE), collapse="")
 , paste(sample(letters, 500, replace=TRUE), collapse="")
)
txt <- paste(c("aaa","bbb","ccc"), txt, sep="")

CatTable(txt, nrepchars=3, wcol=5)
#> aaaaumjvmhrwbxfyunfeknittaaofujbkyrlkseybfbewmzteiwfovtibkxlxesgtfeodwmbjdrtjb
#> bbbnsejzsvejfvkflazpcaeiibpaxydxbgstpncdwdvujcbiyyrcywkvvwkafpksvbrtbxygavitcw
#> cccrxhgwkexjxndsaemqxvditfybvglhggzfhpvvmrpkaucrwfutnotsqnsgztbvnttdtklnkuwfqx
#> 
#> aaaqdvjxsmnimwoszjmafuvcosaqrooudnnqjpwvjqnfvvogueufmkvthpsrplwohdmnisadxnjxmm
#> bbbexvtewewraxuzqvevqihcyzpditaarpyxkrsyqhgqzmhmcgybiabxmyxptuqmsrsamivwmcylql
#> cccpiittupaevnuiqjihganwpjqiwskjmkfsvvnkmqyzapaswzrcxvomzkibxawohunddcztevzhxb
#> 
#> aaajskouokdvqnnrvhmtkcvkjjrzcdouvboxroufhisxwhlcoqctktgdmpqkudorgbevsozigplxuk
#> bbbesfgrjoihpvhpqkgudqvsrskurpfxbvwksuxkghbcxwnfngffkukexlvvzefoshhocdzamjxowz
#> ccceskbtxdekamakdipycrvxhjkemzwgyuoydxbafrcegfsieggwwluzutrgvjiphetvcbfjqljdsl
#> 
#> aaazyygxzwgxefafmjnxtfbtjwoxaaqtspqfumykqofiacftryyuvvqrbaukzzgwettknvgznnggwm
#> bbbhvudhtajepidmdpveyggzpzdbaqeeubkbhojstnopijgixsprymvrbwinhvssbgdrpsmfobqpxg
#> cccfkrtkrnzazvhjvdiwjjdertufvzveqnuzontuwceqojugydopmwjrkhczouobrcbvioinlorywt
#> 
#> aaagmacovjsamxqhozpjbanlsobgnoahsosgxucvvaiggkfccvgklulytibywnhwwevdmgmltmniek
#> bbbxkhlqxheanxhlrlerlbjbganuhbthhaclqnspjzrrdspglufkrbgynqlfazftwxheipnyuwidic
#> cccqllnijifvedzdkckmouuuzttqsiruuxcgulxabkdgoqqaeuvoupmfxsbruobsvemmnntrhqrqmg
#> 
#> aaavybzeywagidwpzwqyebzfwrduujnufpldfcaapkkcqvfrqatpjekywgpabwhllqgvhsvzkvahnr
#> bbbudqkstjyiucevdeplecsdjbtclehvsciadvqlyuslptverkrcxuzcngisgfcmreucspfvmibsjw
#> ccccyqweehgmmopgrnmyytghjuitogkmnnalrbhptbnlrywkgyamuliodjrlvthxvytlepcjudvmht
#>