Find the Closest Value

Find the closest value(s) of a number in a vector x. Multiple values will be reported, if the differences are the same.

Closest(x, a, which = FALSE, na.rm = FALSE)

Arguments

x

the vector to be searched in

a

the reference value

which

a logical value defining if the index position or the value should be returned. By default will the value be returned.

na.rm

a logical value indicating whether NA values should be stripped before the computation proceeds.

Value

the value or index in x which is closest to a

Author

Andri Signorell <andri@signorell.net>

See also

Examples

set.seed(8)
x <- sample(10, size=10, replace=TRUE)

Closest(x, 6)
#> [1] 7 7 7
Closest(x, 6, which=TRUE)
#> [1] 2 4 6

Closest(c(2, 3, 4, 5), 3.5)
#> [1] 3 4