Column Wrap

Wraps text in a character matrix so, that it's displayed over more than one line.

ColumnWrap(x, width = NULL)

Arguments

x
the matrix with one row

width
integer, the width of the columns in characters

Details

A data.frame containing character columns with long texts is often wrapped by columns. This can lead to a loss of overview. ColumnWrap wraps the lines within the columns.

Value

a character matrix

Author

Andri Signorell <andri@signorell.net>

See also

Examples

Abstract(d.pizza)

#> ------------------------------------------------------------------------------ 
#> d.pizza
#> 
#> data frame:	1209 obs. of  16 variables
#> 		917 complete cases (75.8%)
#> 
#>   Nr  ColName         Class            NAs          Levels               
#>   1   index           integer            .                               
#>   2   date            Date              32 (2.6%)                        
#>   3   week            numeric           32 (2.6%)                        
#>   4   weekday         numeric           32 (2.6%)                        
#>   5   area            factor            10 (0.8%)   (3): 1-Brent,        
#>                                                     2-Camden,            
#>                                                     3-Westminster        
#>   6   count           integer           12 (1.0%)                        
#>   7   rabate          logical           12 (1.0%)                        
#>   8   price           numeric           12 (1.0%)                        
#>   9   operator        factor             8 (0.7%)   (3): 1-Allanah,      
#>                                                     2-Maria, 3-Rhonda    
#>   10  driver          factor             5 (0.4%)   (7): 1-Butcher,      
#>                                                     2-Carpenter,         
#>                                                     3-Carter, 4-Farmer,  
#>                                                     5-Hunter, ...        
#>   11  delivery_min    numeric            .                               
#>   12  temperature     numeric           39 (3.2%)                        
#>   13  wine_ordered    integer           12 (1.0%)                        
#>   14  wine_delivered  integer           12 (1.0%)                        
#>   15  wrongpizza      logical            4 (0.3%)                        
#>   16  quality         ordered, factor  201 (16.6%)  (3): 1-low, 2-medium,
#>                                                     3-high               
#>