Get HWZ Datasets

Wrapper for didactical datasets used in statistic courses.

CourseData(name, url = NULL, header = TRUE, sep = ";", ...)

Arguments

name
the name of the file, usually without extension.

url
a url where the data reside, should have the form "http://www.mysite.net/data/". Defaults to the data folder on my site.

header
a logical value indicating whether the file contains the names of the variables as its first line. If missing, the value is determined from the file format: header is set to TRUE if and only if the first row contains one fewer field than the number of columns.

sep
the field separator character. Values on each line of the file are separated by this character. Default is - unlike in read.table - ";" instead of 'white space'.

...
the dots are sent to read.table.

Value

A data.frame containing a representation of the data in the file.

See also

Examples

if (FALSE) {
d.farm <- CourseData("farmer")
}