Cramer-von Mises Test for Normality

Performs the Cramer-von Mises test for the composite hypothesis of normality, see e.g. Thode (2002, Sec. 5.1.3).

CramerVonMisesTest(x)

Arguments

x
a numeric vector of data values, the number of which must be greater than 7. Missing values are allowed.

Details

The Cramer-von Mises test is an EDF omnibus test for the composite hypothesis of normality. The test statistic is $$ W = \frac{1}{12 n} + \sum_{i=1}^{n} \left (p_{(i)} - \frac{2i-1}{2n} \right), $$ where \(p_{(i)} = \Phi([x_{(i)} - \overline{x}]/s)\). Here, \(\Phi\) is the cumulative distribution function of the standard normal distribution, and \(\overline{x}\) and \(s\) are mean and standard deviation of the data values. The p-value is computed from the modified statistic \(Z=W (1.0 + 0.5/n)\) according to Table 4.9 in Stephens (1986).

Value

A list of class htest, containing the following components:

statistic

the value of the Cramer-von Mises statistic.

p.value

the p-value for the test.

method

the character string “Cramer-von Mises normality test”.

data.name

a character string giving the name(s) of the data.

References

Stephens, M.A. (1986) Tests based on EDF statistics In: D'Agostino, R.B. and Stephens, M.A., eds.: Goodness-of-Fit Techniques. Marcel Dekker, New York.

Thode Jr., H.C. (2002) Testing for Normality Marcel Dekker, New York.

Author

Juergen Gross <gross@statistik.uni-dortmund.de>

See also

Examples

CramerVonMisesTest(rnorm(100, mean = 5, sd = 3))

#> 
#> 	Cramer-von Mises normality test
#> 
#> data:  rnorm(100, mean = 5, sd = 3)
#> W = 0.045832, p-value = 0.5716
#> 

CramerVonMisesTest(runif(100, min = 2, max = 4))

#> 
#> 	Cramer-von Mises normality test
#> 
#> data:  runif(100, min = 2, max = 4)
#> W = 0.28912, p-value = 0.0004285
#>