Datasets for Probabilistic Simulation

For performing elementary probability calculations in introductory statistic courses, we might want to simulate random games. The dataset roulette contains the standard sample space for one spin on a roulette wheel. cards contains the standard set of 52 playing cards in four colours (without Jokers). tarot does the same with a classic tarot deck.

cards
tarot
roulette

Value

cards is a data.frame with two columns named rank and suit

tarot is a data.frame with three columns named rank, suit and desc

roulette is a data.frame with two columns named num and col

Examples

head(cards)

#>   rank suit
#> 1    2 club
#> 2    3 club
#> 3    4 club
#> 4    5 club
#> 5    6 club
#> 6    7 club

head(tarot)

#>   rank  suit desc
#> 1    1 wands <NA>
#> 2    2 wands <NA>
#> 3    3 wands <NA>
#> 4    4 wands <NA>
#> 5    5 wands <NA>
#> 6    6 wands <NA>

head(roulette)

#>   num   col
#> 1   0 white
#> 2  26 black
#> 3   3   red
#> 4  35 black
#> 5  12   red
#> 6  28 black

# drawing 5 cards
sample(do.call(paste, cards), 5)

#> [1] "Q spade"   "9 diamond" "3 heart"   "Q diamond" "4 club"   

# spin the wheel
sample(do.call(paste, roulette), 1)

#> [1] "7 red"

# simulate the evening in Las Vegas
sample(do.call(paste, roulette), 32, replace=TRUE)

#>  [1] "31 black" "0 white"  "26 black" "9 red"    "35 black" "1 red"   
#>  [7] "16 red"   "16 red"   "21 red"   "6 black"  "12 red"   "23 red"  
#> [13] "11 black" "26 black" "2 black"  "2 black"  "34 red"   "4 black" 
#> [19] "12 red"   "14 red"   "24 black" "6 black"  "17 black" "7 red"   
#> [25] "3 red"    "36 red"   "16 red"   "30 red"   "32 red"   "12 red"  
#> [31] "27 red"   "6 black"