Convert a Text to a Table

Convert a text to a table by using complete columns of spaces (or any other separator) as delimiting point.

FixToTable(txt, sep = " ", delim = "\t", trim = TRUE, header = TRUE)

Arguments

txt
the text to be partitioned. Works best, if txt is a matrix.

sep
the separator to use. Will frequently be " ".

delim
the new delimiter to insert. (default tab)

trim
logical. Should the separated text be trimmed from whitespace? Defaults to TRUE.

header
logical. Should the first line be interpreted as header?

Details

Only a complete appearance of the separator character in the same position over all rows will be accepted as column delimiter.

Value

a matrix of the separated text.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

# let's get some tabbed text
txt <- matrix(capture.output(Titanic[,,2,1]))
FixToTable(txt[-1,])

#>      Class  Male  Female
#> [1,] "1st"  "118" "4"   
#> [2,] "2nd"  "154" "13"  
#> [3,] "3rd"  "387" "89"  
#> [4,] "Crew" "670" "3"