The Frechet Distribution

Density function, distribution function, quantile function and random generation for the Frechet distribution with location, scale and shape parameters.

dFrechet(x, loc=0, scale=1, shape=1, log = FALSE)
pFrechet(q, loc=0, scale=1, shape=1, lower.tail = TRUE)
qFrechet(p, loc=0, scale=1, shape=1, lower.tail = TRUE)
rFrechet(n, loc=0, scale=1, shape=1)

Arguments

x, q

Vector of quantiles.

p

Vector of probabilities.

n

Number of observations.

loc, scale, shape

Location, scale and shape parameters (can be given as vectors).

log

Logical; if TRUE, the log density is returned.

lower.tail

Logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]

Details

The Frechet distribution function with parameters \(loc = a\), \(scale = b\) and \(shape = s\) is $$G(z) = \exp\left\{-\left(\frac{z-a}{b}\right)^{-s} \right\}$$ for \(z > a\) and zero otherwise, where \(b > 0\) and \(s > 0\).

Value

dFrechet gives the density function, pFrechet gives the distribution function, qFrechet gives the quantile function, and rFrechet generates random deviates.

See also

Author

Alec Stephenson <alec_stephenson@hotmail.com>

Examples

dFrechet(2:4, 1, 0.5, 0.8)
#> [1] 0.25871959 0.09487423 0.05010381
pFrechet(2:4, 1, 0.5, 0.8)
#> [1] 0.5630712 0.7190122 0.7878127
qFrechet(seq(0.9, 0.6, -0.1), 2, 0.5, 0.8)
#> [1] 10.329571  5.260165  3.813966  3.157788
rFrechet(6, 1, 0.5, 0.8)
#> [1] 2.832629 1.585455 1.360407 1.155260 1.067372 1.258905
p <- (1:9)/10
pFrechet(qFrechet(p, 1, 2, 0.8), 1, 2, 0.8)
#> [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9
## [1] 0.1 0.2 0.3 0.4 0.5 0.6 0.7 0.8 0.9