Frequency Table for a Single Variable

Calculates absolute and relative frequencies of a vector x. Continuous (numeric) variables will be cut using the same logic as used by the function hist. Categorical variables will be aggregated by table. The result will contain single and cumulative frequencies for both, absolute values and percentages.

Freq(x, breaks = hist(x, plot = FALSE)$breaks, include.lowest = TRUE,
     ord = c("level", "desc", "asc", "name"),
     useNA = c("no", "ifany", "always"), ...)

# S3 method for Freq
print(x, digits = NULL, ...)

Arguments

x
the variable to be described, can be any atomic type.

breaks
either a numeric vector of two or more cut points or a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut. Default taken from the function hist(). This is ignored if x is not of numeric type.

include.lowest
logical, indicating if an x[i] equal to the lowest (or highest, for right = FALSE) "breaks" value should be included. Ignored if x is not of numeric type.

ord
how should the result be ordered? Default is "level", other choices are 'by frequency' ("descending" or "ascending") or 'by name of the levels' ("name"). The argument can be abbreviated. This is ignored if x is numeric.

useNA
one out of "no", "ifany", "always". Defines whether to include extra NA levels in the table. Defaults to "no" which is the table() default too.

digits
integer, determining the number of digits used to format the relative frequencies.

...
further arguments are passed to the function cut(). Use dig.lab to control the format of numeric group names. Use the argument right to define if the intervals should be closed on the right (and open on the left) or vice versa.
In print.Freq the dots are not used.

Details

If breaks is specified as a single number, the range of the data is divided into breaks pieces of equal length, and then the outer limits are moved away by 0.1% of the range to ensure that the extreme values both fall within the break intervals. (If x is a constant vector, equal-length intervals are created that cover the single value.)

Value

an object of type "Freq", which is basically a data.frame with 5 columns (earning a specific print routine), containing the following components:

level

factor. The levels of the grouping variable.

freq

integer. The absolute frequencies.

perc

numeric. The relative frequencies (percent).

cumfreq

integer. The cumulative sum of the absolute frequencies.

cumperc

numeric. The cumulative sum of the relative frequencies.

See also

Examples

data(d.pizza)

# result is a data.frame
d.freq <- Freq(d.pizza$price)
d.freq

#>         level  freq  perc  cumfreq  cumperc
#> 1      [0,10]     4  .003        4     .003
#> 2     (10,20]    96  .080      100     .084
#> 3     (20,30]   183  .153      283     .236
#> 4     (30,40]   147  .123      430     .359
#> 5     (40,50]   263  .220      693     .579
#> 6     (50,60]   169  .141      862     .720
#> 7     (60,70]   119  .099      981     .820
#> 8     (70,80]   109  .091    1'090     .911
#> 9     (80,90]    68  .057    1'158     .967
#> 10   (90,100]    22  .018    1'180     .986
#> 11  (100,110]     7  .006    1'187     .992
#> 12  (110,120]     6  .005    1'193     .997
#> 13  (120,130]     3  .003    1'196     .999
#> 14  (130,140]     1  .001    1'197    1.000

# it is printed by default with 3 digits for the percent values,
# but the number of digits can be defined in the print function
print(d.freq, digits=5)

#>         level  freq    perc  cumfreq  cumperc
#> 1      [0,10]     4  .00334        4   .00334
#> 2     (10,20]    96  .08020      100   .08354
#> 3     (20,30]   183  .15288      283   .23642
#> 4     (30,40]   147  .12281      430   .35923
#> 5     (40,50]   263  .21972      693   .57895
#> 6     (50,60]   169  .14119      862   .72013
#> 7     (60,70]   119  .09942      981   .81955
#> 8     (70,80]   109  .09106    1'090   .91061
#> 9     (80,90]    68  .05681    1'158   .96742
#> 10   (90,100]    22  .01838    1'180   .98580
#> 11  (100,110]     7  .00585    1'187   .99165
#> 12  (110,120]     6  .00501    1'193   .99666
#> 13  (120,130]     3  .00251    1'196   .99916
#> 14  (130,140]     1  .00084    1'197  1.00000

# sorted by frequency
Freq(d.pizza$driver, ord="desc")

#>        level  freq  perc  cumfreq  cumperc
#> 1  Carpenter   272  .226      272     .226
#> 2     Carter   234  .194      506     .420
#> 3     Taylor   204  .169      710     .590
#> 4     Hunter   156  .130      866     .719
#> 5     Miller   125  .104      991     .823
#> 6     Farmer   117  .097    1'108     .920
#> 7    Butcher    96  .080    1'204    1.000

# sorted by name, including NAs
Freq(d.pizza$driver, ord="name", useNA="ifany")

#>        level  freq  perc  cumfreq  cumperc
#> 1    Butcher    96  .079       96     .079
#> 2  Carpenter   272  .225      368     .304
#> 3     Carter   234  .194      602     .498
#> 4     Farmer   117  .097      719     .595
#> 5     Hunter   156  .129      875     .724
#> 6     Miller   125  .103    1'000     .827
#> 7     Taylor   204  .169    1'204     .996
#> 8       <NA>     5  .004    1'209    1.000

# percentages and cumulative frequencies for a vector of count data
Freq(as.table(c(2,4,12,8)))

#>    level  freq  perc  cumfreq  cumperc
#> 1      A     2  .077        2     .077
#> 2      B     4  .154        6     .231
#> 3      C    12  .462       18     .692
#> 4      D     8  .308       26    1.000