Create a New Excel Instance

Start a new instance of Excel and return its handle. This is needed to address the Excel application and objects afterwards.

GetNewXL(visible = TRUE, newdoc = TRUE)

Arguments

visible
logical, should Excel made visible? Defaults to TRUE.

newdoc
logical, determining if a new workbook should be created. Defaults to TRUE.

Details

Here's a list of some frequently used commands.
Let's assume:

xl <- GetNewXL()

workbooks	xl$workbooks()$count()

quit excel	xl$quit()

Author

Andri Signorell <andri@signorell.net>

See also

Examples

if (FALSE)  # Windows-specific example
# get a handle to a new excel instance
xl <- GetNewXL()