(Robust) Jarque Bera Test

This function performs the Jarque-Bera tests of normality either the robust or the classical way.

JarqueBeraTest(x, robust = TRUE, method = c("chisq", "mc"),
               N = 0, na.rm = FALSE)

Arguments

x

a numeric vector of data values.

robust

defines, whether the robust version should be used. Default is TRUE.

method

a character string out of chisq or mc, specifying how the critical values should be obtained. Default is approximated by the chisq-distribution or empirically via Monte Carlo.

N

number of Monte Carlo simulations for the empirical critical values

na.rm

defines if NAs should be omitted. Default is FALSE.

Details

The test is based on a joint statistic using skewness and kurtosis coefficients. The robust Jarque-Bera (RJB) version of utilizes the robust standard deviation (namely the mean absolute deviation from the median, as provided e. g. by MeanAD(x, FUN=median)) to estimate sample kurtosis and skewness. For more details see Gel and Gastwirth (2006).
Setting robust to FALSE will perform the original Jarque-Bera test (see Jarque, C. and Bera, A (1980)).

Value

A list with class htest containing the following components:

statistic

the value of the test statistic.

parameter

the degrees of freedom.

p.value

the p-value of the test.

method

type of test was performed.

data.name

a character string giving the name of the data.

Note

This function is melted from the jarque.bera.test (in tseries package) and the rjb.test from the package lawstat.

References

Gastwirth, J. L.(1982) Statistical Properties of A Measure of Tax Assessment Uniformity, Journal of Statistical Planning and Inference 6, 1-12.

Gel, Y. R. and Gastwirth, J. L. (2008) A robust modification of the Jarque-Bera test of normality, Economics Letters 99, 30-32.

Jarque, C. and Bera, A. (1980) Efficient tests for normality, homoscedasticity and serial independence of regression residuals, Economics Letters 6, 255-259.

Author

W. Wallace Hui, Yulia R. Gel, Joseph L. Gastwirth, Weiwen Miao

See also

Examples

x <- rnorm(100)    # null hypothesis
JarqueBeraTest(x)
#> 
#> 	Robust Jarque Bera Test
#> 
#> data:  x
#> X-squared = 16.785, df = 2, p-value = 0.0002266
#> 

x <- runif(100)    # alternative hypothesis
JarqueBeraTest(x, robust=TRUE)
#> 
#> 	Robust Jarque Bera Test
#> 
#> data:  x
#> X-squared = 2.7459, df = 2, p-value = 0.2534
#>