Kendall's \(\tau_{a}\)

Calculate Kendall's tau-a statistic, a measure of association for ordinal factors in a two-way table.
The function has interfaces for a table (matrix) and for single vectors.

KendallTauA(x, y = NULL, direction = c("row", "column"), conf.level = NA, ...)

Arguments

x

a numeric vector or a table. A matrix will be treated as table.

y

NULL (default) or a vector with compatible dimensions to x. If y is provided, table(x, y, ...) is calculated.

direction

direction of the calculation. Can be "row" (default) or "column", where "row" calculates Kendall's tau-a (R|C) ("column dependent").

conf.level

confidence level of the interval. If set to NA (which is the default) no confidence interval will be calculated.

...

further arguments are passed to the function table, allowing i.e. to set useNA. This refers only to the vector interface.

Details

Kendall's tau coefficient (sometimes called "Kendall rank correlation coefficient"), is a statistic used to measure the association between two measured quantities. It is a measure of rank correlation: the similarity of the orderings of the data when ranked by each of the quantities.
Kendall's tau-a is computed as $$ \tau_a(C|R) = \frac{P-Q}{\frac{1}{2} \cdot n \cdot (n-1)}$$ where P equals twice the number of concordances and Q twice the number of discordances. Its range is [-1, 1].
(Note that Kendall tau-a does not take into consideration any ties, which makes it unpractical. Consider using KendallTauB (Tau-b) when ties are present.)

Value

a single numeric value if no confidence intervals are requested,

and otherwise a numeric vector with 3 elements for the estimate, the lower and the upper confidence interval

References

Agresti, A. (2002) Categorical Data Analysis. John Wiley & Sons, pp. 57-59.

Hollander, M, Wolfe, D. A., Chicken, E. (2014) Nonparametric Statistical Methods, Third edition, Wiley,

Liebetrau, A. M. (1983) Measures of Association, Sage University Papers Series on Quantitative Applications in the Social Sciences, 07-004. Newbury Park, CA: Sage, pp. 49-56

Author

Andri Signorell <andri@signorell.net>

See also

Examples

# example in:
# http://support.sas.com/documentation/cdl/en/statugfreq/63124/PDF/default/statugfreq.pdf
# pp. S. 1821

tab <- as.table(rbind(c(26,26,23,18,9),c(6,7,9,14,23)))

# Kendall's tau-a C|R
KendallTauA(tab, direction="column", conf.level=0.95)
#>     tau_a    lwr.ci    upr.ci 
#> 0.2068323 0.1771300 0.2365346 
# Kendall's tau-a R|C
KendallTauA(tab, direction="row", conf.level=0.95)
#>     tau_a    lwr.ci    upr.ci 
#> 0.2068323 0.1771300 0.2365346 

# http://support.sas.com/documentation/cdl/en/statugfreq/63124/PDF/default/statugfreq.pdf
# pp. 1814 (143)
tab <- as.table(cbind(c(11,2),c(4,6)))

KendallTauA(tab, direction="row", conf.level=0.95)
#>      tau_a     lwr.ci     upr.ci 
#> 0.22924901 0.07441035 0.38408768 
KendallTauA(tab, direction="column", conf.level=0.95)
#>      tau_a     lwr.ci     upr.ci 
#> 0.22924901 0.07441035 0.38408768 

# Liebetrau, pp. 52
x <- c(1,2,2,3,3,3,4,5)
y <- c(1,3,2,1,5,3,4,5)

ConDisPairs(table(x, y))
#> $pi.c
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    6    5    3    2    0
#> [2,]    4    5    4    3    1
#> [3,]    2    3    4    4    3
#> [4,]    1    3    4    6    5
#> [5,]    0    2    3    5    6
#> 
#> $pi.d
#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    0    1    2    4    5
#> [2,]    0    1    1    2    3
#> [3,]    2    1    0    0    1
#> [4,]    4    3    1    1    0
#> [5,]    5    4    2    1    0
#> 
#> $C
#> [1] 18
#> 
#> $D
#> [1] 3
#> 
KendallTauA(x, y, conf.level=0.95)
#>     tau_a    lwr.ci    upr.ci 
#> 0.5357143 0.1324073 0.9390213