List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [25] "AllIdentical"           "AndersonDarlingTest"    "Append"                
#>  [28] "Arrow"                  "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AxisBreak"             
#>  [34] "BarText"                "BarnardTest"            "BartelsRankTest"       
#>  [37] "Bg"                     "BhapkarTest"            "BinToDec"              
#>  [40] "BinTree"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "CCC"                    "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "CmToPts"               
#>  [64] "CmyToCmyk"              "CmykToCmy"              "CmykToRgb"             
#>  [67] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [70] "CoefVar"                "CoefVarCI"              "CohenD"                
#>  [73] "CohenKappa"             "ColToGray"              "ColToGrey"             
#>  [76] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [79] "ColToRgb"               "CollapseTable"          "ColorLegend"           
#>  [82] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [85] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [88] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [91] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [94] "Cor"                    "CorCI"                  "CorPart"               
#>  [97] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [100] "CourseData"             "Cov"                    "CramerV"               
#> [103] "CramerVonMisesTest"     "CronbachAlpha"          "Cross"                 
#> [106] "CrossN"                 "Cstat"                  "CutQ"                  
#> [109] "DB"                     "Day"                    "Day<-"                 
#> [112] "DecToBin"               "DecToHex"               "DecToOct"              
#> [115] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [118] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [121] "DivCoefMax"             "Divisors"               "DoBy"                  
#> [124] "DoCall"                 "Dot"                    "DrawArc"               
#> [127] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [130] "DrawEllipse"            "DrawRegPolygon"         "Dummy"                 
#> [133] "DunnTest"               "DunnettTest"            "DurbinWatsonTest"      
#> [136] "EX"                     "Entropy"                "Eps"                   
#> [139] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [142] "Factorize"              "Fade"                   "FctArgs"               
#> [145] "Fibonacci"              "FindColor"              "FindCorr"              
#> [148] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [151] "FixToTable"             "Flags"                  "Fmt"                   
#> [154] "Format"                 "Frac"                   "Freq"                  
#> [157] "Freq2D"                 "GCD"                    "GTest"                 
#> [160] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [163] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [166] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [169] "GetNewXL"               "Gini"                   "GiniDeltas"            
#> [172] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [175] "GoodmanKruskalTau"      "Gsd"                    "Herfindahl"            
#> [178] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [181] "HighLow"                "Hmean"                  "HmsToMinute"           
#> [184] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [187] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [190] "HuberM"                 "HunterGaston"           "ICC"                   
#> [193] "IPMT"                   "IQRw"                   "IRR"                   
#> [196] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [199] "InDots"                 "Interval"               "IsDate"                
#> [202] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [205] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [208] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [211] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [214] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [217] "KendallW"               "Keywords"               "KrippAlpha"            
#> [220] "Kurt"                   "LCM"                    "LOCF"                  
#> [223] "LOF"                    "Label"                  "Label<-"               
#> [226] "Labels"                 "Labels<-"               "Lambda"                
#> [229] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [232] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [235] "LinScale"               "LineToUser"             "LogSt"                 
#> [238] "LogStInv"               "Logit"                  "LogitInv"              
#> [241] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [244] "MAD"                    "MADCI"                  "MAE"                   
#> [247] "MAPE"                   "MHChisqTest"            "MSE"                   
#> [250] "Mar"                    "Margins"                "MaxDigits"             
#> [253] "Mean"                   "MeanAD"                 "MeanCI"                
#> [256] "MeanCIn"                "MeanDiffCI"             "MeanSE"                
#> [259] "Median"                 "MedianCI"               "Mgp"                   
#> [262] "Mgsub"                  "Midx"                   "Minute"                
#> [265] "MixColor"               "ModSummary"             "Mode"                  
#> [268] "Month"                  "MonthDays"              "MosesTest"             
#> [271] "MoveAvg"                "MultMerge"              "MultinomCI"            
#> [274] "MutInf"                 "N"                      "NAIf"                  
#> [277] "NAIfBlank"              "NAIfZero"               "NMAE"                  
#> [280] "NMSE"                   "NPV"                    "NPVFixBond"            
#> [283] "Ndec"                   "NemenyiTest"            "Nf"                    
#> [286] "Now"                    "OPR"                    "ORToRelRisk"           
#> [289] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [292] "Outlier"                "Overlap"                "PDFManual"             
#> [295] "PMT"                    "PPMT"                   "PageTest"              
#> [298] "PairApply"              "Pal"                    "ParseFormula"          
#> [301] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [304] "PearsonTest"            "PercTable"              "PercentRank"           
#> [307] "Permn"                  "Phi"                    "Phrase"                
#> [310] "PlotACF"                "PlotArea"               "PlotBag"               
#> [313] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [316] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [319] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [322] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [325] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [328] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [331] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [334] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [337] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [340] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [343] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [346] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [349] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [352] "Prec"                   "Primes"                 "PseudoR2"              
#> [355] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [358] "QuantileCI"             "Quarter"                "Quot"                  
#> [361] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [364] "RTempdirAlive"          "RadToDeg"               "Range"                 
#> [367] "Rank"                   "Recode"                 "Recycle"               
#> [370] "RelRisk"                "Rename"                 "Rev"                   
#> [373] "RevCode"                "RgbToCmy"               "RgbToCol"              
#> [376] "RgbToHex"               "RgbToLong"              "RndPairs"              
#> [379] "RndWord"                "RobScale"               "RomanToInt"            
#> [382] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [385] "RunsTest"               "SD"                     "SDN"                   
#> [388] "SLN"                    "SMAPE"                  "SYD"                   
#> [391] "Sample"                 "SampleTwins"            "SaveAs"                
#> [394] "ScheffeTest"            "SecToHms"               "Second"                
#> [397] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [400] "SetAttr"                "SetNames"               "Shade"                 
#> [403] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [406] "SignTest"               "Skew"                   "Small"                 
#> [409] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [412] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [415] "Spec"                   "SphToCart"              "SplitAt"               
#> [418] "SplitPath"              "SplitToCol"             "SplitToDummy"          
#> [421] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [424] "Str"                    "StrAbbr"                "StrAlign"              
#> [427] "StrCap"                 "StrChop"                "StrCountW"             
#> [430] "StrDist"                "StrExtract"             "StrExtractBetween"     
#> [433] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [436] "StrPos"                 "StrRev"                 "StrRight"              
#> [439] "StrSpell"               "StrSplit"               "StrTrim"               
#> [442] "StrTrunc"               "StrVal"                 "Strata"                
#> [445] "StripAttr"              "StuartMaxwellTest"      "StuartTauC"            
#> [448] "SysInfo"                "TMod"                   "TOne"                  
#> [451] "TTestA"                 "TextContrastColor"      "TextToTable"           
#> [454] "TheilU"                 "Timezone"               "TitleRect"             
#> [457] "ToLong"                 "ToWide"                 "ToWrd"                 
#> [460] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [463] "Today"                  "Trim"                   "TschuprowT"            
#> [466] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [469] "UnirootAll"             "Unit"                   "Unit<-"                
#> [472] "Untable"                "Unwhich"                "VIF"                   
#> [475] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [478] "VarN"                   "VarTest"                "VarX"                  
#> [481] "VecRot"                 "VecShift"               "Vigenere"              
#> [484] "VonNeumannTest"         "Week"                   "Weekday"               
#> [487] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [490] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [493] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [496] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [499] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [502] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [505] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [508] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [511] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [514] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [517] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [520] "YTM"                    "Year"                   "YearDay"               
#> [523] "YearDays"               "YearMonth"              "YuenTTest"             
#> [526] "YuleQ"                  "YuleY"                  "ZTest"                 
#> [529] "ZeroIfNA"               "Zodiac"                 "aovlDetails"           
#> [532] "aovlErrorTerms"         "as.CDateFmt"            "as.fmt"                
#> [535] "as.matrix.xtabs"        "axTicks.Date"           "axTicks.POSIXct"       
#> [538] "createCOMReference"     "dBenf"                  "dExtrVal"              
#> [541] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [544] "dGompertz"              "dGumbel"                "dNegWeibull"           
#> [547] "dOrder"                 "dRevGumbel"             "dRevWeibull"           
#> [550] "dTri"                   "identify.formula"       "lines.Lc"              
#> [553] "lines.SmoothSpline"     "lines.lm"               "lines.loess"           
#> [556] "lines.smooth.spline"    "pBenf"                  "pExtrVal"              
#> [559] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [562] "pGompertz"              "pGumbel"                "pNegWeibull"           
#> [565] "pOrder"                 "pRevGumbel"             "pRevWeibull"           
#> [568] "pTri"                   "plot.Conf"              "plot.Lc"               
#> [571] "plot.Lclist"            "plot.PostHocTest"       "plot.bagplot"          
#> [574] "plot.palette"           "power.chisq.test"       "print.Assocs"          
#> [577] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [580] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [583] "print.PercTable"        "print.PostHocTest"      "print.mtest"           
#> [586] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [589] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [592] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [595] "qRevGumbelExp"          "qRevWeibull"            "qTri"                  
#> [598] "rBenf"                  "rExtrVal"               "rFrechet"              
#> [601] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [604] "rGumbel"                "rNegWeibull"            "rOrder"                
#> [607] "rRevGumbel"             "rRevWeibull"            "rSum21"                
#> [610] "rTri"                   "reorder.factor"         "split.formula"