Confidence Intervals for Median Absolute Deviations

A function for the median absolute deviation is included in base R, mad, but there's no function for calculating confidence intervals. Arachchige/Prendergast introduce interval estimators of the MAD to make reliable inferences for dispersion for a single population and ratios and differences of MADs for comparing two populations.

MADCI(x, y = NULL, two.samp.diff = TRUE, gld.est = "TM",
      conf.level = 0.95, sides = c("two.sided","left","right"),
      na.rm = FALSE, ...)

Arguments

x
a (non-empty) numeric vector of data values.

y
a second (non-empty) numeric vector of data values.

two.samp.diff
logical, defining if the confidence intervals for a difference of x and y should be calculated.

gld.est
type of gld estimator.

conf.level
confidence level of the interval

sides
a character string specifying the side of the confidence interval, must be one of "two.sided" (default), "left" or "right". You can specify just the initial letter. "left" would be analogue to a hypothesis of "greater" in a t.test.

na.rm
logical. Should missing values be removed? Defaults to FALSE.

...
further arguments, not used here

Value

a numeric vector with 3 elements:

mad

median absolute deviation

lwr.ci

lower bound of the confidence interval

upr.ci

upper bound of the confidence interval

References

Arachchige Chandima N. P. G., Prendergast Luke A. (2019) Confidence intervals for median absolute deviations, arXiv:1910.00229 [math.ST]

See also

Examples

x <- rlnorm(100)
y <- rlnorm(200, meanlog=1.2)

MADCI(x)                           # single sample

#>       mad    lwr.ci    upr.ci 
#> 0.7844891 0.3406550 1.2283233 

MADCI(x, y)                        # two sample difference

#>       mad    lwr.ci    upr.ci 
#> -2.045040 -2.697248 -1.392833 

MADCI(x, y, two.samp.diff = FALSE) # two sample squared ratio 

#>        mad     lwr.ci     upr.ci 
#> 0.07686797 0.02359939 0.25037452