Mean Absolute Deviation From a Center Point

Calculates the mean absolute deviation from a center point, typically the sample mean or the median.

MeanAD(x, FUN = mean, na.rm = FALSE)

Arguments

x
a vector containing the observations.

FUN
the name of a function to be used as center. Can as well be a self defined function. Default is mean.

na.rm
a logical value indicating whether or not missing values should be removed. Defaults to FALSE.

Details

The MeanAD function calculates the mean absolute deviation from the mean value (or from another supplied center point) of x, after having removed NA values (if requested): $$\frac{1}{n} \cdot \sum_{i=1}^{n}\left | x_{i}-c \right | \; \; \; \textup{where} \; c=mean(x) \; \textup{or} \; c=med(x)$$
It exists primarily to simplify the discussion of descriptive statistics during an introductory stats class.

Value

Numeric value.

See also

Examples

x <- runif(100)
MeanAD(x)

#> [1] 0.2688156

speed <- c(58, 88, 40, 60, 72, 66, 80, 48, NA)
MeanAD(speed)

#> [1] NA

MeanAD(speed, na.rm=TRUE)

#> [1] 12.5

# using the median as centerpoint
x <- c(2,3,5,3,1,15,23)

MeanAD(x, FUN=mean)

#> [1] 6.612245

MeanAD(x, FUN=median)

#> [1] 5.285714

# define a fix center
MeanAD(x, 4)

#> [1] 5.428571