Confidence Interval for the Median

Calculates the confidence interval for the median.

MedianCI(x, conf.level = 0.95, sides = c("two.sided", "left", "right"),
         na.rm = FALSE, method = c("exact", "boot"), R = 999)

Arguments

x

a (non-empty) numeric vector of data values.

conf.level

confidence level of the interval

sides

a character string specifying the side of the confidence interval, must be one of "two.sided" (default), "left" or "right". You can specify just the initial letter. "left" would be analogue to a hypothesis of "greater" in a t.test.

na.rm

logical. Should missing values be removed? Defaults to FALSE.

method

defining the type of interval that should be calculated (one out of "exact", "boot"). Default is "exact". See Details.

R

The number of bootstrap replicates. Usually this will be a single positive integer. See boot.ci for details.

Details

The "exact" method is the way SAS is said to calculate the confidence interval. This is implemented in SignTest and is extracted therefrom. The boot confidence interval type is calculated by means of boot.ci with default type "basic".
Use sapply, resp.apply, to get the confidence intervals from a data.frame or from a matrix.

Value

a numeric vector with 3 elements:

median

median

lwr.ci

lower bound of the confidence interval

upr.ci

upper bound of the confidence interval

Author

Andri Signorell <andri@signorell.net>

See also

Examples

MedianCI(d.pizza$price, na.rm=TRUE)
#> median lwr.ci upr.ci 
#> 46.764 44.970 47.664 
#> attr(,"conf.level")
#> [1] 0.9506831
MedianCI(d.pizza$price, conf.level=0.99, na.rm=TRUE)
#> median lwr.ci upr.ci 
#> 46.764 44.970 48.070 
#> attr(,"conf.level")
#> [1] 0.990742

t(round(sapply(d.pizza[,c("delivery_min","temperature","price")], MedianCI, na.rm=TRUE), 3))
#>              median lwr.ci upr.ci
#> delivery_min 24.400  23.80 25.200
#> temperature  50.000  49.50 50.700
#> price        46.764  44.97 47.664

MedianCI(d.pizza$price, na.rm=TRUE, method="exact")
#> median lwr.ci upr.ci 
#> 46.764 44.970 47.664 
#> attr(,"conf.level")
#> [1] 0.9506831
MedianCI(d.pizza$price, na.rm=TRUE, method="boot")
#> median lwr.ci upr.ci 
#> 46.764 45.864 48.558