Compute the Convex Combination of Two Colors

This function can be used to compute the result of color mixing (it assumes additive mixing).

MixColor(col1, col2, amount1 = 0.5)

Arguments

col1

the first color.

col2

the second color.

amount1

the amount of color1. The amount of color2 results in (1-amount1).

Value

The mixed color as hexstring

Author

Andri Signorell <andri@signorell.net>

See also

Examples

# a mix between red and yellow with rates 3:7
MixColor("red", "yellow", 0.3)
#> [1] "#FFB200FF"