Moses Test of Extreme Reactions

Perform Moses test of extreme reactions, which is a distribution-free non-parametric test for the difference between two independent groups in the extremity of scores (in both directions) that the groups contain. Scores from both groups are pooled and converted to ranks, and the test statistic is the span of scores (the range plus 1) in one of the groups chosen arbitrarily. An exact probability is computed for the span and then recomputed after dropping a specified number of extreme scores from each end of its range. The exact one-tailed probability is calculated.

MosesTest(x, ...)

# S3 method for default
MosesTest(x, y, extreme = NULL, ...)

# S3 method for formula
MosesTest(formula, data, subset, na.action, ...)

Arguments

x
numeric vector of data values. x will be treated as control group. Non-finite (e.g. infinite or missing) values will be omitted.

y
numeric vector of data values. y will be treated as experiment group. Non-finite (e.g. infinite or missing) values will be omitted.

formula
a formula of the form lhs ~ rhs where lhs gives the data values and rhs the corresponding groups.

data
an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).

subset
an optional vector specifying a subset of observations to be used.

na.action
a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").

extreme
integer, defines the number of extreme values to be dropped from the control group before calculating the span. Default (NULL) is the integer part of 0.05 * length(x) or 1, whichever is greater. If extreme is too large, it will be cut down to floor(length(x)-2)/2.

...
further arguments to be passed to or from methods.

Details

For two independent samples from a continuous field, this tests whether extreme values are equally likely in both populations or if they are more likely to occur in the population from which the sample with the larger range was drawn.

Note that the ranks are calculated in decreasing mode.

Value

A list with class “htest” containing the following components:

statistic

the value of the Moses Test statistic.

p.value

the p-value for the test.

method

the character string “Moses Test of Extreme Reactions”.

data.name

a character string giving the name(s) of the data.

References

Moses, L.E. (1952) A Two-Sample Test, Psychometrika, 17, 239-247.

http://www-01.ibm.com/support/knowledgecenter/SSLVMB_20.0.0/com.ibm.spss.statistics.help/alg_npar_tests_moses.htm

See also

Examples

x <- c(0.80, 0.83, 1.89, 1.04, 1.45, 1.38, 1.91, 1.64, 0.73, 1.46)
y <- c(1.15, 0.88, 0.90, 0.74, 1.21)

MosesTest(x, y)

#> 
#> 	Moses Test of Extreme Reactions
#> 
#> data:  x and y
#> S = 12, p-value = 0.8462
#> alternative hypothesis: extreme values are more likely in x than in y
#> 

set.seed(1479)
x <- sample(1:20, 10, replace=TRUE)
y <- sample(5:25, 6, replace=TRUE)

MosesTest(x, y)

#> 
#> 	Moses Test of Extreme Reactions
#> 
#> data:  x and y
#> S = 11, p-value = 0.3916
#> alternative hypothesis: extreme values are more likely in x than in y
#>