Percentage Table

Prints a 2-way contingency table along with percentages, marginal, and conditional distributions. All the frequencies are nested into one single table.

# S3 method for default
PercTable(x, y = NULL, ...)

# S3 method for table
PercTable(tab, row.vars = NULL, col.vars = NULL, justify = "right",
          freq = TRUE, rfrq = "100", expected = FALSE, residuals = FALSE,
          stdres = FALSE, margins = NULL, digits = NULL, ...)

# S3 method for formula
PercTable(formula, data, subset, na.action, ...)

# S3 method for PercTable
print(x, vsep = NULL, ...)

Margins(tab, ...)

Arguments

x, y
objects which can be interpreted as factors (including character strings). x and y will be tabulated via table(x, y).
If x is a matrix, it will be coerced to a table via as.table(x).

tab
a r x c-contingency table

row.vars
a vector of row variables (see Details).

col.vars
a vector of column variables (see Details). If this is left to NULL the table structure will be preserved.

justify
either "left" or "right" for defining the alignment of the table cells.

freq
boolean. Should absolute frequencies be included? Defaults to TRUE.

rfrq
a string with 3 characters, each of them being 1 or 0. The first position means total percentages, the second means row percentages and the third column percentages. "011" produces a table output with row and column percentages.

expected
the expected counts under the null hypothesis.

residuals
the Pearson residuals, (observed - expected) / sqrt(expected).

stdres
standardized residuals, (observed - expected) / sqrt(V), where V is the residual cell variance (for the case where x is a matrix, n * p * (1 - p) otherwise).

margins
a vector, consisting out of 1 and/or 2. Defines the margin sums to be included. 1 stands for row margins, 2 for column margins, c(1,2) for both. Default is NULL (none).

digits
integer. With how many digits shoud the relative frequencies be formatted? Default can be set by DescToolsOptions(digits=x).

formula
a formula of the form lhs ~ rhs where lhs will be tabled versus rhs (table(lhs, rhs)).

data
an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula formula. By default the variables are taken from environment(formula).

subset
an optional vector specifying a subset of observations to be used.

na.action
a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").

vsep
logical, defining if an empty row should be introduced between the table rows. Default is FALSE, if only a table with one single description (either frequencies or percents) should be returned and TRUE in any other case.

...
the dots are passed from PercTable.default() to PercTable.table() and from Margins to the function Freq.

Details

PercTable prints a 2-dimensional table. The absolute and relative frequencies are nested into one flat table by means of ftable. row.vars, resp. col.vars can be used to define the structure of the table. row.vars can either be the names of the dimensions (included percentages are named "idx") or numbers (1:3, where 1 is the first dimension of the table, 2 the second and 3 the percentages).
Use Sort() if you want to have your table sorted by rows.

The style in which numbers are formatted is selected by Fmt() from the DescTools options. Absolute frequencies will use Fmt("abs") and Fmt("per") will do it for the percentages. The options can be changed with Fmt(abs=as.fmt(...)) which is basically a "fmt"-object containing any format information used in Format.

Margins() returns a list containing all the one dimensional margin tables of a n-dimensional table along the given dimensions. It uses margin.table() for all the dimensions and adds the appropriate percentages.

Value

Returns an object of class "ftable".

References

Agresti, Alan (2007) Introduction to categorical data analysis. NY: John Wiley and Sons, Section 2.4.5

See also

Examples

tab <- table(driver=d.pizza$driver, area=d.pizza$area)

PercTable(tab=tab, col.vars=2)

#>                                                
#>             area   Brent   Camden   Westminster
#> driver                                         
#>                                                
#> Butcher     freq      72        1            22
#>             perc    .060     .001          .018
#>                                                
#> Carpenter   freq      29       19           221
#>             perc    .024     .016          .185
#>                                                
#> Carter      freq     177       47             5
#>             perc    .148     .039          .004
#>                                                
#> Farmer      freq      19       87            11
#>             perc    .016     .073          .009
#>                                                
#> Hunter      freq     128        4            24
#>             perc    .107     .003          .020
#>                                                
#> Miller      freq       6       41            77
#>             perc    .005     .034          .064
#>                                                
#> Taylor      freq      42      142            20
#>             perc    .035     .119          .017
#>                                                

PercTable(tab=tab, col.vars=2, margins=c(1,2))

#>                                                      
#>             area   Brent   Camden   Westminster   Sum
#> driver                                               
#>                                                      
#> Butcher     freq      72        1            22    95
#>             perc    .060     .001          .018  .080
#>                                                      
#> Carpenter   freq      29       19           221   269
#>             perc    .024     .016          .185  .225
#>                                                      
#> Carter      freq     177       47             5   229
#>             perc    .148     .039          .004  .192
#>                                                      
#> Farmer      freq      19       87            11   117
#>             perc    .016     .073          .009  .098
#>                                                      
#> Hunter      freq     128        4            24   156
#>             perc    .107     .003          .020  .131
#>                                                      
#> Miller      freq       6       41            77   124
#>             perc    .005     .034          .064  .104
#>                                                      
#> Taylor      freq      42      142            20   204
#>             perc    .035     .119          .017  .171
#>                                                      
#> Sum         freq     473      341           380 1'194
#>             perc    .396     .286          .318 1.000
#>                                                      

PercTable(tab=tab, col.vars=2, margins=2)

#>                                                
#>             area   Brent   Camden   Westminster
#> driver                                         
#>                                                
#> Butcher     freq      72        1            22
#>             perc    .060     .001          .018
#>                                                
#> Carpenter   freq      29       19           221
#>             perc    .024     .016          .185
#>                                                
#> Carter      freq     177       47             5
#>             perc    .148     .039          .004
#>                                                
#> Farmer      freq      19       87            11
#>             perc    .016     .073          .009
#>                                                
#> Hunter      freq     128        4            24
#>             perc    .107     .003          .020
#>                                                
#> Miller      freq       6       41            77
#>             perc    .005     .034          .064
#>                                                
#> Taylor      freq      42      142            20
#>             perc    .035     .119          .017
#>                                                
#> Sum         freq     473      341           380
#>             perc    .396     .286          .318
#>                                                

PercTable(tab=tab, col.vars=2, margins=1)

#>                                                      
#>             area   Brent   Camden   Westminster   Sum
#> driver                                               
#>                                                      
#> Butcher     freq      72        1            22    95
#>             perc    .060     .001          .018  .080
#>                                                      
#> Carpenter   freq      29       19           221   269
#>             perc    .024     .016          .185  .225
#>                                                      
#> Carter      freq     177       47             5   229
#>             perc    .148     .039          .004  .192
#>                                                      
#> Farmer      freq      19       87            11   117
#>             perc    .016     .073          .009  .098
#>                                                      
#> Hunter      freq     128        4            24   156
#>             perc    .107     .003          .020  .131
#>                                                      
#> Miller      freq       6       41            77   124
#>             perc    .005     .034          .064  .104
#>                                                      
#> Taylor      freq      42      142            20   204
#>             perc    .035     .119          .017  .171
#>                                                      

PercTable(tab=tab, col.vars=2, margins=NULL)

#>                                                
#>             area   Brent   Camden   Westminster
#> driver                                         
#>                                                
#> Butcher     freq      72        1            22
#>             perc    .060     .001          .018
#>                                                
#> Carpenter   freq      29       19           221
#>             perc    .024     .016          .185
#>                                                
#> Carter      freq     177       47             5
#>             perc    .148     .039          .004
#>                                                
#> Farmer      freq      19       87            11
#>             perc    .016     .073          .009
#>                                                
#> Hunter      freq     128        4            24
#>             perc    .107     .003          .020
#>                                                
#> Miller      freq       6       41            77
#>             perc    .005     .034          .064
#>                                                
#> Taylor      freq      42      142            20
#>             perc    .035     .119          .017
#>                                                

PercTable(tab=tab, col.vars=2, rfrq="000")

#>                                         
#> area        Brent   Camden   Westminster
#> driver                                  
#> Butcher        72        1            22
#> Carpenter      29       19           221
#> Carter        177       47             5
#> Farmer         19       87            11
#> Hunter        128        4            24
#> Miller          6       41            77
#> Taylor         42      142            20

# just the percentages without absolute values
PercTable(tab=tab, col.vars=2, rfrq="110", freq=FALSE)

#>                                                 
#>             area    Brent   Camden   Westminster
#> driver                                          
#>                                                 
#> Butcher     perc     .060     .001          .018
#>             p.row    .758     .011          .232
#>                                                 
#> Carpenter   perc     .024     .016          .185
#>             p.row    .108     .071          .822
#>                                                 
#> Carter      perc     .148     .039          .004
#>             p.row    .773     .205          .022
#>                                                 
#> Farmer      perc     .016     .073          .009
#>             p.row    .162     .744          .094
#>                                                 
#> Hunter      perc     .107     .003          .020
#>             p.row    .821     .026          .154
#>                                                 
#> Miller      perc     .005     .034          .064
#>             p.row    .048     .331          .621
#>                                                 
#> Taylor      perc     .035     .119          .017
#>             p.row    .206     .696          .098
#>                                                 

# just the row percentages in percent format (pfmt = TRUE)
PercTable(tab, freq= FALSE, rfrq="010", pfmt=TRUE, digits=1)

#>                                         
#> area        Brent   Camden   Westminster
#> driver                                  
#> Butcher        .8       .0            .2
#> Carpenter      .1       .1            .8
#> Carter         .8       .2            .0
#> Farmer         .2       .7            .1
#> Hunter         .8       .0            .2
#> Miller         .0       .3            .6
#> Taylor         .2       .7            .1

# just the expected frequencies and the standard residuals
PercTable(tab=tab, rfrq="000", expected = TRUE, stdres = TRUE)

#>                                                  
#>             area     Brent   Camden   Westminster
#> driver                                           
#>                                                  
#> Butcher     freq        72        1            22
#>             exp         38       27            30
#>             stdres       8       -6            -2
#>                                                  
#> Carpenter   freq        29       19           221
#>             exp        107       77            86
#>             stdres     -11       -9            20
#>                                                  
#> Carter      freq       177       47             5
#>             exp         91       65            73
#>             stdres      13       -3           -11
#>                                                  
#> Farmer      freq        19       87            11
#>             exp         46       33            37
#>             stdres      -5       12            -5
#>                                                  
#> Hunter      freq       128        4            24
#>             exp         62       45            50
#>             stdres      12       -8            -5
#>                                                  
#> Miller      freq         6       41            77
#>             exp         49       35            39
#>             stdres      -8        1             8
#>                                                  
#> Taylor      freq        42      142            20
#>             exp         81       58            65
#>             stdres      -6       14            -7
#>                                                  

# rearrange output such that freq are inserted as columns instead of rows
PercTable(tab=tab, col.vars=c(3,2), rfrq="111")

#>                                                                                
#>              freq                           perc                          p.row
#> area        Brent   Camden   Westminster   Brent   Camden   Westminster   Brent
#> driver                                                                         
#> Butcher        72        1            22    .060     .001          .018    .758
#> Carpenter      29       19           221    .024     .016          .185    .108
#> Carter        177       47             5    .148     .039          .004    .773
#> Farmer         19       87            11    .016     .073          .009    .162
#> Hunter        128        4            24    .107     .003          .020    .821
#> Miller          6       41            77    .005     .034          .064    .048
#> Taylor         42      142            20    .035     .119          .017    .206
#>                                                                
#>                                    p.col                       
#> area        Camden   Westminster   Brent   Camden   Westminster
#> driver                                                         
#> Butcher       .011          .232    .152     .003          .058
#> Carpenter     .071          .822    .061     .056          .582
#> Carter        .205          .022    .374     .138          .013
#> Farmer        .744          .094    .040     .255          .029
#> Hunter        .026          .154    .271     .012          .063
#> Miller        .331          .621    .013     .120          .203
#> Taylor        .696          .098    .089     .416          .053

# putting the areas in rows
PercTable(tab=tab, col.vars=c(3,1), rfrq="100", margins=c(1,2))

#>                                                                               
#>                  freq                                                         
#> driver        Butcher   Carpenter   Carter   Farmer   Hunter   Miller   Taylor
#> area                                                                          
#> Brent              72          29      177       19      128        6       42
#> Camden              1          19       47       87        4       41      142
#> Westminster        22         221        5       11       24       77       20
#> Sum                95         269      229      117      156      124      204
#>                                                                            
#>                        perc                                                
#> driver        Sum   Butcher   Carpenter   Carter   Farmer   Hunter   Miller
#> area                                                                       
#> Brent         473      .060        .024     .148     .016     .107     .005
#> Camden        341      .001        .016     .039     .073     .003     .034
#> Westminster   380      .018        .185     .004     .009     .020     .064
#> Sum         1'194      .080        .225     .192     .098     .131     .104
#>                           
#>                           
#> driver        Taylor   Sum
#> area                      
#> Brent           .035  .396
#> Camden          .119  .286
#> Westminster     .017  .318
#> Sum             .171 1.000

# formula interface with subset
PercTable(driver ~ area, data=d.pizza, subset=wine_delivered==0)

#>                                                
#>             area   Brent   Camden   Westminster
#> driver                                         
#>                                                
#> Butcher     freq      65        1            18
#>             perc    .064     .001          .018
#>                                                
#> Carpenter   freq      27       14           170
#>             perc    .026     .014          .167
#>                                                
#> Carter      freq     161       42             4
#>             perc    .158     .041          .004
#>                                                
#> Farmer      freq      19       72             9
#>             perc    .019     .071          .009
#>                                                
#> Hunter      freq     113        4            22
#>             perc    .111     .004          .022
#>                                                
#> Miller      freq       6       35            67
#>             perc    .006     .034          .066
#>                                                
#> Taylor      freq      36      118            18
#>             perc    .035     .116          .018
#>                                                

# sort the table by rows, order first column (Zurich), then third, then row.names (0)
PercTable(tab=Sort(tab, ord=c(1,3,0)))

#>                                                
#>             area   Brent   Camden   Westminster
#> driver                                         
#>                                                
#> Miller      freq       6       41            77
#>             perc    .005     .034          .064
#>                                                
#> Farmer      freq      19       87            11
#>             perc    .016     .073          .009
#>                                                
#> Carpenter   freq      29       19           221
#>             perc    .024     .016          .185
#>                                                
#> Taylor      freq      42      142            20
#>             perc    .035     .119          .017
#>                                                
#> Butcher     freq      72        1            22
#>             perc    .060     .001          .018
#>                                                
#> Hunter      freq     128        4            24
#>             perc    .107     .003          .020
#>                                                
#> Carter      freq     177       47             5
#>             perc    .148     .039          .004
#>                                                

# reverse the row variables, so that absolute frequencies and percents
# are not nested together
PercTable(tab, row.vars=c(3, 1))

#>                                                
#>        area        Brent   Camden   Westminster
#>        driver                                  
#> freq   Butcher        72        1            22
#>        Carpenter      29       19           221
#>        Carter        177       47             5
#>        Farmer         19       87            11
#>        Hunter        128        4            24
#>        Miller          6       41            77
#>        Taylor         42      142            20
#>                                                
#> perc   Butcher      .060     .001          .018
#>        Carpenter    .024     .016          .185
#>        Carter       .148     .039          .004
#>        Farmer       .016     .073          .009
#>        Hunter       .107     .003          .020
#>        Miller       .005     .034          .064
#>        Taylor       .035     .119          .017
#>                                                

# the vector interface
PercTable(x=d.pizza$driver, y=d.pizza$area)

#>                                                
#>                    Brent   Camden   Westminster
#>                                                
#> Butcher     freq      72        1            22
#>             perc    .060     .001          .018
#>                                                
#> Carpenter   freq      29       19           221
#>             perc    .024     .016          .185
#>                                                
#> Carter      freq     177       47             5
#>             perc    .148     .039          .004
#>                                                
#> Farmer      freq      19       87            11
#>             perc    .016     .073          .009
#>                                                
#> Hunter      freq     128        4            24
#>             perc    .107     .003          .020
#>                                                
#> Miller      freq       6       41            77
#>             perc    .005     .034          .064
#>                                                
#> Taylor      freq      42      142            20
#>             perc    .035     .119          .017
#>                                                

PercTable(x=d.pizza$driver, y=d.pizza$area, margins=c(1,2), rfrq="000", useNA="ifany")

#>                                                     
#>             Brent   Camden   Westminster    NA   Sum
#>                                                     
#> Butcher        72        1            22     1    96
#> Carpenter      29       19           221     3   272
#> Carter        177       47             5     5   234
#> Farmer         19       87            11     0   117
#> Hunter        128        4            24     0   156
#> Miller          6       41            77     1   125
#> Taylor         42      142            20     0   204
#> NA              1        3             1     0     5
#> Sum           474      344           381    10 1'209

# one dimensional x falls back to the function Freq()
PercTable(x=d.pizza$driver)

#>                        
#>             freq   perc
#>                        
#> Butcher       96   .080
#> Carpenter    272   .226
#> Carter       234   .194
#> Farmer       117   .097
#> Hunter       156   .130
#> Miller       125   .104
#> Taylor       204   .169

# the margin tables
Margins(Titanic)

#> $Class
#>    level  freq  perc  cumfreq  cumperc
#> 1    1st   325  .148      325     .148
#> 2    2nd   285  .129      610     .277
#> 3    3rd   706  .321    1'316     .598
#> 4   Crew   885  .402    2'201    1.000
#> 
#> $Sex
#>     level   freq  perc  cumfreq  cumperc
#> 1    Male  1'731  .786    1'731     .786
#> 2  Female    470  .214    2'201    1.000
#> 
#> $Age
#>    level   freq  perc  cumfreq  cumperc
#> 1  Child    109  .050      109     .050
#> 2  Adult  2'092  .950    2'201    1.000
#> 
#> $Survived
#>    level   freq  perc  cumfreq  cumperc
#> 1     No  1'490  .677    1'490     .677
#> 2    Yes    711  .323    2'201    1.000
#>