(Weighted) Standard Deviation

This function computes the standard deviation of the values in x. If na.rm is TRUE then missing values are removed before computation proceeds.

SD(x, weights = NULL, na.rm = FALSE, ...)

Arguments

x
a numeric vector or an R object which is coercible to one by as.double(x).

weights
a numerical vector of weights the same length as x giving the weights to use for elements of x.

na.rm
logical. Should missing values be removed?

...
further arguments passed to or from other methods.

Details

Like var this uses denominator \(n - 1\).

The standard deviation of a zero-length vector (after removal of NAs if na.rm = TRUE) is not defined and gives an error. The standard deviation of a length-one vector is NA.

See also

Examples

SD(1:2)^2

#> [1] 0.5