Send a Mail Using Outlook as Mail Client

Sending emails in R can be required in some reporting tasks. As we already have RDCOMClient available we wrap the send code in a function.

SendOutlookMail(to, cc = NULL, bcc = NULL, subject, body, attachment = NULL)

Arguments

to
a vector of recipients

cc
a vector of recipients receiving a carbon copy

bcc
a vector of recipients receiving a blind carbon copy

subject
the subject of the mail

body
the body text of the mail

attachment
a vector of paths to attachments

Value

Nothing is returned

Author

Andri Signorell <andri@signorell.net> strongly based on code of Franziska Mueller

See also

Examples

if (FALSE) {
SendOutlookMail(to=c("me@microsoft.com", "you@rstudio.com"), subject = "Some Info", 
                body = "Hi all\r Find the files attached\r Regards, Andri", 
                attachment = c("C:/temp/fileA.txt", 
                               "C:/temp/fileB.txt"))
}