Set the Names in an Object

This is a convenience function that sets the names on an object and returns the object. It is most useful at the end of a function definition where one is creating the object to be returned and would prefer not to store it under a name just so the names can be assigned. In addition to the function setNames the user can decide, whether rownames, colnames or simply the names are to be set.

SetNames(x, ...)

Arguments

x
an object for which a names attribute will be meaningful

...
the names to be assigned to the object. This should be a character vector of names named rownames, colnames or names. Setting rownames=NULL would remove existing rownames. All kind of names can be changed at the same time.

Value

An object of the same sort as object with the new names assigned.

See also

Examples

tab <- table(d.pizza$driver, d.pizza$wine_delivered)
SetNames(BinomCI(tab[,1], rowSums(tab)), rownames=rownames(tab))

#>                 est    lwr.ci    upr.ci
#> Butcher   0.9042553 0.8279588 0.9488080
#> Carpenter 0.7925926 0.7402915 0.8366847
#> Carter    0.9177489 0.8751179 0.9467131
#> Farmer    0.8547009 0.7795898 0.9072606
#> Hunter    0.9025974 0.8455167 0.9400818
#> Miller    0.8720000 0.8021644 0.9196530
#> Taylor    0.8557214 0.8004723 0.8976285