Return Some Randomly Chosen Elements of an Object

For displaying the first and last elements of an object there are the functions head and tail. Sometimes one might want to see more randomly scattered elements. This function returns some random parts of a vector, matrix or a data frame. The order of the elements within the object will be preserved.

Some(x, n = 6L, ...)
# S3 method for default
Some(x, n = 6L, ...)
# S3 method for data.frame
Some(x, n = 6L, ...)
# S3 method for matrix
Some(x, n = 6L, addrownums = TRUE, ...)

Arguments

x
an object

n
a single integer. If positive, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

addrownums
if there are no row names, create them from the row numbers.

...
arguments to be passed to or from other methods.

Details

For matrices, 2-dim tables and data frames, Some() returns some n rows when n > 0 or all but the some n rows when n < 0. Some.matrix() is not exported (unlike head.matrix).

If a matrix has no row names, then Some() will add row names of the form "[n,]" to the result, so that it looks similar to the last lines of x when printed. Setting addrownums = FALSE suppresses this behaviour.

I desisted from implementing interfaces for tables, ftables and functions, as this would not make much sense.

Value

An object (usually) like x but generally smaller.

See also

Examples

Some(letters)

#> [1] "j" "n" "s" "u" "v" "y"

Some(letters, n = -6L)

#>  [1] "c" "d" "f" "g" "h" "i" "j" "k" "l" "n" "p" "q" "r" "s" "t" "v" "w" "x" "y"
#> [20] "z"

Some(freeny.x, n = 10L)

#>       lag quarterly revenue price index income level market potential
#>  [1,]               8.79636     4.70997      5.82110          12.9699
#>  [2,]               8.79236     4.70217      5.82558          12.9733
#>  [3,]               8.79137     4.68944      5.83112          12.9774
#>  [4,]               8.81486     4.68558      5.84046          12.9806
#>  [5,]               8.81301     4.64019      5.85036          12.9831
#>  [6,]               8.90751     4.62553      5.86464          12.9854
#>  [7,]               8.93673     4.61991      5.87769          12.9900
#>  [8,]               8.96161     4.61654      5.89763          12.9943
#>  [9,]               8.96044     4.61407      5.92574          12.9992
#> [10,]               9.00868     4.60766      5.94232          13.0033
#> [11,]               9.03049     4.60227      5.95365          13.0099
#> [12,]               9.06906     4.58960      5.96120          13.0159
#> [13,]               9.05871     4.57592      5.97805          13.0212
#> [14,]               9.10698     4.58661      6.00377          13.0265
#> [15,]               9.12685     4.57997      6.02829          13.0351
#> [16,]               9.17096     4.57176      6.03475          13.0429
#> [17,]               9.18665     4.56104      6.03906          13.0497
#> [18,]               9.23823     4.54906      6.05046          13.0551
#> [19,]               9.26487     4.53957      6.05563          13.0634
#> [20,]               9.28436     4.51018      6.06093          13.0693
#> [21,]               9.31378     4.50352      6.07103          13.0737
#> [22,]               9.35025     4.49360      6.08018          13.0770
#> [23,]               9.35835     4.46505      6.08858          13.0849
#> [24,]               9.39767     4.44924      6.10199          13.0918
#> [25,]               9.42150     4.43966      6.11207          13.0950
#> [26,]               9.44223     4.42025      6.11596          13.0984
#> [27,]               9.48721     4.41060      6.12129          13.1089
#> [28,]               9.52374     4.41151      6.12200          13.1169
#> [29,]               9.53980     4.39810      6.13119          13.1222
#> [30,]               9.58123     4.38513      6.14705          13.1266
#> [31,]               9.60048     4.37320      6.15336          13.1356
#> [32,]               9.64496     4.32770      6.15627          13.1415
#> [33,]               9.64390     4.32023      6.16274          13.1444
#> [34,]               9.69405     4.30909      6.17369          13.1459
#> [35,]               9.69958     4.30909      6.16135          13.1520
#> [36,]               9.68683     4.30552      6.18231          13.1593
#> [37,]               9.71774     4.29627      6.18768          13.1579
#> [38,]               9.74924     4.27839      6.19377          13.1625
#> [39,]               9.77536     4.27789      6.20030          13.1664

Some(freeny.y)

#> [1] 8.79137 9.10698 9.23823 9.39767 9.53980 9.64390