Compactly Display the Structure of any R Object

Basically a wrapper for str(), extended with an enumeration for the variables of a data.frame.

Str(x, ...)

Arguments

x

any R object about which you want to have some information.

...

dots are passed to str.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

Str(d.pizza)
#> 'data.frame':	1209 obs. of  16 variables:
#>   1 $ index         : int  1 2 3 4 5 6 7 8 9 10 ...
#>   2 $ date          : Date, format: "2014-03-01" "2014-03-01" "2014-03-01" "2014-03-01" ...
#>   3 $ week          : num  9 9 9 9 9 9 9 9 9 9 ...
#>   4 $ weekday       : num  6 6 6 6 6 6 6 6 6 6 ...
#>   5 $ area          : Factor w/ 3 levels "Brent","Camden",..: 2 3 3 1 1 2 2 1 3 1 ...
#>   6 $ count         : int  5 2 3 2 5 1 4 NA 3 6 ...
#>   7 $ rabate        : logi  TRUE FALSE FALSE FALSE TRUE FALSE ...
#>   8 $ price         : num  65.7 27 41 26 57.6 ...
#>   9 $ operator      : Factor w/ 3 levels "Allanah","Maria",..: 3 3 1 1 3 1 3 1 1 3 ...
#>  10 $ driver        : Factor w/ 7 levels "Butcher","Carpenter",..: 7 1 1 7 3 7 7 7 7 3 ...
#>  11 $ delivery_min  : num  20 19.6 17.8 37.3 21.8 48.7 49.3 25.6 26.4 24.3 ...
#>  12 $ temperature   : num  53 56.4 36.5 NA 50 27 33.9 54.8 48 54.4 ...
#>  13 $ wine_ordered  : int  0 0 0 0 0 0 1 NA 0 1 ...
#>  14 $ wine_delivered: int  0 0 0 0 0 0 1 NA 0 1 ...
#>  15 $ wrongpizza    : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  16 $ quality       : Ord.factor w/ 3 levels "low"<"medium"<..: 2 3 NA NA 2 1 1 3 3 2 ...