Stratified Sampling

Stratified sampling with equal/unequal probabilities.

Strata(x, stratanames = NULL, size,
       method = c("srswor", "srswr", "poisson", "systematic"),
       pik, description = FALSE)

Arguments

x
a data frame or a matrix; its number of rows is n, the population size.

stratanames
vector of stratification variables.

size
vector of stratum sample sizes (in the order in which the strata are given in the input data set).

method
method to select units; implemented are: a) simple random sampling without replacement ("srswor"), b) simple random sampling with replacement ("srswr"), c) Poisson sampling ("poisson"), d) systematic sampling ("systematic") (default is "srswor").

pik
vector of inclusion probabilities or auxiliary information used to compute them; this argument is only used for unequal probability sampling (Poisson and systematic). If an auxiliary information is provided, the function uses the inclusionprobabilities function for computing these probabilities. If the method is "srswr" and the sample size is larger than the population size, this vector is normalized to one.

description
a message is printed if its value is TRUE; the message gives the number of selected units and the number of the units in the population. By default, the value is FALSE.

Value

The function produces an object, which contains the following information:

id

the identifier of the selected units.

stratum

the unit stratum.

prob

the final unit inclusion probability.

See also

Examples

# Example from An and Watts (New SAS procedures for Analysis of Sample Survey Data)
# generates artificial data (a 235X3 matrix with 3 columns: state, region, income).
# the variable "state" has 2 categories ('nc' and 'sc').
# the variable "region" has 3 categories (1, 2 and 3).
# the sampling frame is stratified by region within state.
# the income variable is randomly generated

m <- rbind(matrix(rep("nc",165), 165, 1, byrow=TRUE),
           matrix(rep("sc", 70), 70, 1, byrow=TRUE))
m <- cbind.data.frame(m, c(rep(1, 100), rep(2,50), rep(3,15),
                      rep(1, 30), rep(2, 40)), 1000 * runif(235))
names(m) <- c("state", "region", "income")

# computes the population stratum sizes
table(m$region, m$state)

#>    
#>      nc  sc
#>   1 100  30
#>   2  50  40
#>   3  15   0

# not run
#     nc  sc
#  1 100  30
#  2  50  40
#  3  15   0
# there are 5 cells with non-zero values
# one draws 5 samples (1 sample in each stratum)
# the sample stratum sizes are 10,5,10,4,6, respectively
# the method is 'srswor' (equal probability, without replacement)

s <- Strata(m, c("region", "state"), size=c(10, 5, 10, 4, 6), method="srswor")

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>          income state region  income.1 stratum size  id
#> 1.52   23.10093    nc      1  23.10093       1   10  52
#> 1.62  761.42889    nc      1 761.42889       1   10  62
#> 1.84  592.34629    nc      1 592.34629       1   10  84
#> 1.9   408.09932    nc      1 408.09932       1   10   9
#> 1.94  552.14334    nc      1 552.14334       1   10  94
#> 1.14  980.51893    nc      1 980.51893       1   10  14
#> 1.22  861.17353    nc      1 861.17353       1   10  22
#> 1.71  859.76367    nc      1 859.76367       1   10  71
#> 1.66  213.89583    nc      1 213.89583       1   10  66
#> 1.41   44.13647    nc      1  44.13647       1   10  41
#> 2.127 398.03868    nc      2 398.03868       2    5 127
#> 2.126  80.87036    nc      2  80.87036       2    5 126
#> 2.102 982.44097    nc      2 982.44097       2    5 102
#> 2.105 267.47048    nc      2 267.47048       2    5 105
#> 2.136 737.76557    nc      2 737.76557       2    5 136
#> 3.157 633.58222    nc      3 633.58222       3   10 157
#> 3.154 385.42599    nc      3 385.42599       3   10 154
#> 3.156 794.92344    nc      3 794.92344       3   10 156
#> 3.163 135.61091    nc      3 135.61091       3   10 163
#> 3.158 427.83172    nc      3 427.83172       3   10 158
#> 3.152 490.38595    nc      3 490.38595       3   10 152
#> 3.155 616.97988    nc      3 616.97988       3   10 155
#> 3.159 508.78938    nc      3 508.78938       3   10 159
#> 3.151 426.00580    nc      3 426.00580       3   10 151
#> 3.161 391.48706    nc      3 391.48706       3   10 161
#> 4.187 888.16006    sc      1 888.16006       4    4 187
#> 4.181 662.91062    sc      1 662.91062       4    4 181
#> 4.173 959.95646    sc      1 959.95646       4    4 173
#> 4.169 783.51664    sc      1 783.51664       4    4 169
#> 5.223 548.44081    sc      2 548.44081       5    6 223
#> 5.204 991.83105    sc      2 991.83105       5    6 204
#> 5.197 491.80663    sc      2 491.80663       5    6 197
#> 5.213 864.23243    sc      2 864.23243       5    6 213
#> 5.230 836.37106    sc      2 836.37106       5    6 230
#> 5.207 423.71597    sc      2 423.71597       5    6 207

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0

# The same data as in Example 1
# the method is 'systematic' (unequal probability, without replacement)
# the selection probabilities are computed using the variable 'income'
s <- Strata(m,c("region", "state"), size=c(10, 5, 10, 4, 6),
            method="systematic", pik=m$income)

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>           income state region   income.1 stratum size  id
#> 1.67  941.573442    nc      1 941.573442       1   10  67
#> 1.74  550.599515    nc      1 550.599515       1   10  74
#> 1.50  207.669702    nc      1 207.669702       1   10  50
#> 1.60  169.762922    nc      1 169.762922       1   10  60
#> 1.13  379.228159    nc      1 379.228159       1   10  13
#> 1.72  944.806864    nc      1 944.806864       1   10  72
#> 1.93  778.014447    nc      1 778.014447       1   10  93
#> 1.17   52.493456    nc      1  52.493456       1   10  17
#> 1.11  854.247657    nc      1 854.247657       1   10  11
#> 1.97  211.988364    nc      1 211.988364       1   10  97
#> 2.150 948.044730    nc      2 948.044730       2    5 150
#> 2.137 887.681510    nc      2 887.681510       2    5 137
#> 2.130 718.016017    nc      2 718.016017       2    5 130
#> 2.119 785.784168    nc      2 785.784168       2    5 119
#> 2.146 348.991587    nc      2 348.991587       2    5 146
#> 3.155 616.979882    nc      3 616.979882       3   10 155
#> 3.153 507.131356    nc      3 507.131356       3   10 153
#> 3.164 388.503158    nc      3 388.503158       3   10 164
#> 3.152 490.385954    nc      3 490.385954       3   10 152
#> 3.163 135.610908    nc      3 135.610908       3   10 163
#> 3.162 352.577260    nc      3 352.577260       3   10 162
#> 3.159 508.789378    nc      3 508.789378       3   10 159
#> 3.151 426.005796    nc      3 426.005796       3   10 151
#> 3.161 391.487057    nc      3 391.487057       3   10 161
#> 3.165 172.973646    nc      3 172.973646       3   10 165
#> 4.189 483.593212    sc      1 483.593212       4    4 189
#> 4.173 959.956460    sc      1 959.956460       4    4 173
#> 4.171 633.772600    sc      1 633.772600       4    4 171
#> 4.175 982.188314    sc      1 982.188314       4    4 175
#> 5.235   7.926157    sc      2   7.926157       5    6 235
#> 5.196 192.331680    sc      2 192.331680       5    6 196
#> 5.211 368.005813    sc      2 368.005813       5    6 211
#> 5.204 991.831050    sc      2 991.831050       5    6 204
#> 5.227 588.872587    sc      2 588.872587       5    6 227
#> 5.210 580.408316    sc      2 580.408316       5    6 210

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0