Comparison Table For Linear Models

Collect the coefficients and some qualifying statistics of linear models and organize it in a table for comparison and reporting. The function supports linear and general linear models.

TMod(..., FUN = NULL)

ModSummary(x, ...)

# S3 method for lm
ModSummary(x, conf.level=0.95, ...)
# S3 method for glm
ModSummary(x, conf.level=0.95, ...)

Arguments

x
a (general) linear model object.

...
a list of (general) linear models.

conf.level
the level for the confidence intervals.

FUN
function with arguments est, se, tval, pval, lci, uci to display the coefficients. The default function will display the coefficient and significance stars for the p-values.

Details

In order to compare the coefficients of linear models, the user is left to his own devices. R offers no support in this respect. TMod jumps into the breach and displays the coefficients of several models in tabular form. For this purpose, different quality indicators for the models are displayed, so that a comprehensive comparison of the models is possible. In particular, it is easy to see the effect that adding or omitting variables has on forecast quality.

A plot function for a TMod object will produce a dotchart with the coefficients and their confidence intervals.

Value

character table

See also

Examples

r.full <- lm(Fertility ~ . , swiss)
r.nox <- lm(Fertility ~ . -Examination - Catholic, swiss)
r.grp <- lm(Fertility ~ . -Education - Catholic + CutQ(Catholic), swiss)
r.gam <- glm(Fertility ~ . , swiss, family=Gamma(link="identity"))
r.gama <- glm(Fertility ~ .- Agriculture , swiss, family=Gamma(link="identity"))
r.gaml <- glm(Fertility ~ . , swiss, family=Gamma(link="log"))

TMod(r.full, r.nox, r.grp, r.gam, r.gama, r.gaml)

#> Waiting for profiling to be done...

#> Waiting for profiling to be done...

#> Waiting for profiling to be done...

#>                coef     r.full      r.nox      r.grp      r.gam     r.gama    
#> 1       (Intercept) 67 ***         51 ***     53 ***     64 ***     49 ***    
#> 2       Agriculture -0 *           -0         -0         -0 *            -    
#> 3       Examination -0                  -     -1 ***     -0         0         
#> 4         Education -1 ***         -1 ***          -     -1 ***     -1 ***    
#> 5          Catholic 0 **                -          -     0 **       0 *       
#> 6  Infant.Mortality 1 **           1 **       2 ***      1 **       1 **      
#> 7  CutQ(Catholic)Q2      -              -     -0              -          -    
#> 8  CutQ(Catholic)Q3      -              -     -6              -          -    
#> 9  CutQ(Catholic)Q4      -              -     6               -          -    
#> 10              ---                                                           
#> 11        r.squared      0.707          0.566      0.611      -          -    
#> 12    adj.r.squared      0.671          0.536      0.552      -          -    
#> 13            sigma      7.165          8.511      8.359      -          -    
#> 14           logLik   -156.036       -165.243   -162.699   -156.506   -159.674
#> 15          logLik0      -              -          -       -186.992   -186.992
#> 16               G2      -              -          -         60.971     54.636
#> 17         deviance   2105.043       3114.635   2795.124      -          -    
#> 18              AIC    326.072        340.485    341.398    327.012    331.347
#> 19              BIC    339.023        349.736    356.199    339.963    342.448
#> 20            numdf      5              3          6          -          -    
#> 21            dendf     41             43         40          -          -    
#> 22                N     47             47         47         47         47    
#> 23                F     19.761         18.699     10.454      -          -    
#> 24                p      5.594e-10      0.000      0.000      -          -    
#> 25              MAE      5.321          6.826      6.079      5.330      5.869
#> 26             MAPE      0.079          0.098      0.095      0.078      0.086
#> 27              MSE     44.788         66.269     59.471     45.342     52.015
#> 28             RMSE      6.692          8.141      7.712      6.734      7.212
#> 29         McFadden      -              -          -          0.163      0.146
#> 30      McFaddenAdj      -              -          -          0.131      0.119
#> 31       Nagelkerke      -              -          -          0.727      0.688
#> 32         CoxSnell      -              -          -          0.727      0.687
#>    r.gaml    
#> 1  4 ***     
#> 2  -0 *      
#> 3  -0        
#> 4  -0 ***    
#> 5  0 *       
#> 6  0 **      
#> 7       -    
#> 8       -    
#> 9       -    
#> 10           
#> 11      -    
#> 12      -    
#> 13      -    
#> 14   -157.224
#> 15   -186.992
#> 16     59.534
#> 17      -    
#> 18    328.449
#> 19    341.400
#> 20      -    
#> 21      -    
#> 22     47    
#> 23      -    
#> 24      -    
#> 25      5.458
#> 26      0.080
#> 27     47.361
#> 28      6.882
#> 29      0.159
#> 30      0.127
#> 31      0.718
#> 32      0.718

# display confidence intervals
TMod(r.full, r.nox, r.gam, FUN = function(est, se, tval, pval, lci, uci){
  gettextf("%s [%s, %s]",
           Format(est, fmt=Fmt("num")),
           Format(lci, digits=3),
           Format(uci, digits=2)
           )
})

#> Waiting for profiling to be done...

#>                coef         r.full              r.nox              r.gam    
#> 1       (Intercept) 67 [45.294, 88.54] 51 [28.928, 73.27] 64 [44.294, 83.44]
#> 2       Agriculture -0 [-0.314, -0.03] -0 [-0.173, 0.12]  -0 [-0.318, -0.04]
#> 3       Examination -0 [-0.771, 0.25]       -             -0 [-0.626, 0.36] 
#> 4         Education -1 [-1.241, -0.50] -1 [-1.205, -0.51] -1 [-1.252, -0.63]
#> 5          Catholic   0 [0.033, 0.18]       -               0 [0.035, 0.18] 
#> 6  Infant.Mortality   1 [0.306, 1.85]    1 [0.608, 2.38]    1 [0.479, 1.84] 
#> 7               ---                                                         
#> 8         r.squared      0.707              0.566              -            
#> 9     adj.r.squared      0.671              0.536              -            
#> 10            sigma      7.165              8.511              -            
#> 11           logLik   -156.036           -165.243           -156.506        
#> 12          logLik0      -                  -               -186.992        
#> 13               G2      -                  -                 60.971        
#> 14         deviance   2105.043           3114.635              -            
#> 15              AIC    326.072            340.485            327.012        
#> 16              BIC    339.023            349.736            339.963        
#> 17            numdf      5                  3                  -            
#> 18            dendf     41                 43                  -            
#> 19                N     47                 47                 47            
#> 20                F     19.761             18.699              -            
#> 21                p      5.594e-10          0.000              -            
#> 22              MAE      5.321              6.826              5.330        
#> 23             MAPE      0.079              0.098              0.078        
#> 24              MSE     44.788             66.269             45.342        
#> 25             RMSE      6.692              8.141              6.734        
#> 26         McFadden      -                  -                  0.163        
#> 27      McFaddenAdj      -                  -                  0.131        
#> 28       Nagelkerke      -                  -                  0.727        
#> 29         CoxSnell      -                  -                  0.727        

# cbind interface is not supported!!
# d.titanic <- reshape(as.data.frame(Titanic),
#                       idvar = c("Class","Sex","Age"),
#                       timevar="Survived",
#                       direction = "wide")
#
# r.glm0 <- glm(cbind(Freq.Yes, Freq.No) ~ 1, data=d.titanic, family="binomial")
# r.glm1 <- glm(cbind(Freq.Yes, Freq.No) ~ Class, data=d.titanic, family="binomial")
# r.glm2 <- glm(cbind(Freq.Yes, Freq.No) ~ ., data=d.titanic, family="binomial")

d.titanic <- Untable(Titanic)

r.glm0 <- glm(Survived ~ 1, data=d.titanic, family="binomial")
r.glm1 <- glm(Survived ~ Class, data=d.titanic, family="binomial")
r.glm2 <- glm(Survived ~ ., data=d.titanic, family="binomial")

TMod(r.glm0, r.glm1, r.glm2)

#> Waiting for profiling to be done...

#> Error in is.data.frame(data): object 'd.titanic' not found

# plot OddsRatios
d.pima <- MASS::Pima.tr2

r.a <- glm(type ~ npreg + bp + skin + bmi + ped + age, data=d.pima, family=binomial)
r.b <- glm(type ~ npreg + glu + bp + skin, data=d.pima, family=binomial)
r.c <- glm(type ~ npreg + age, data=d.pima, family=binomial)

or.a <- OddsRatio(r.a)

#> Waiting for profiling to be done...

#> Error in is.data.frame(data): object 'd.pima' not found

or.b <- OddsRatio(r.b)

#> Waiting for profiling to be done...

#> Error in is.data.frame(data): object 'd.pima' not found

or.c <- OddsRatio(r.c)

#> Waiting for profiling to be done...

#> Error in is.data.frame(data): object 'd.pima' not found

# create the model table
tm <- TMod(m_A=or.a, m_B=or.b, m_C=or.c)

#> Error in TMod(m_A = or.a, m_B = or.b, m_C = or.c): object 'or.a' not found

# .. and plotit
plot(tm, main="ORs for Models A, B, C",
     pch=15, col=c(hred, hblue), xlim=c(0.5, 1.5), panel.first=abline(v=1, col="grey"))

#> Error in plot(tm, main = "ORs for Models A, B, C", pch = 15, col = c(hred,     hblue), xlim = c(0.5, 1.5), panel.first = abline(v = 1, col = "grey")): object 'tm' not found