Create Table One Describing Baseline Characteristics

Create a table summarizing continuous, categorical and dichotomous variables, optionally stratified by one or more variables, while performing adequate statistical tests.

TOne(x, grp = NA, add.length = TRUE, colnames = NULL, vnames = NULL, total = TRUE,
     align = "\\l", FUN = NULL, TEST = NULL, intref = "high",
     fmt.pval = as.fmt(fmt = "*", na.form = "   "))

Arguments

x
a data.frame containing all the variables to be included in the table.

grp
the grouping variable.

add.length
logical. If set to TRUE (default), a row with the lengths of the group will be inserted as first row of the table.

colnames
a vector of columnnames for the result table.

vnames
a vector of variablenames to be placed in the first column instead of the real names.

total
logical, TRUE if totals should be added.

align
the character on whose position the strings will be aligned. Left alignment can be requested by setting sep = "\\l", right alignment by "\\r" and center alignment by "\\c". Mind the backslashes, as if they are omitted, strings would be aligned to the character l, r or c respectively. Default value is "\\l", thus left alignment.

FUN
the function to be used as location and dispersion measure (typically mean/sd or alternatively median/IQR). See examples.

TEST
a list of functions to be used to test the variables. Must be named as "num", "cat" and "dich" and be defined as function with arguments (x, g), generating something similar to a p.value. (See examples.)

intref
one out of "high" (default) or "low", defining which value of a dichotomous numeric or logical variable should be reported. Usually this will be 1 or TRUE. Setting it to "low" will report the lower value 0 or FALSE.

fmt.pval
format code or function for the p-values of the tests, default is as.fmt(fmt = "p", na.form = " ")

Details

There are three types of variables currently implemented: numeric, factor and variables with only two levels.

Value

a character matrix

See also

Examples

options(scipen = 8)
opt <- DescToolsOptions()

# define some special formats for count data, percentages and numeric results
# (those will be supported by TOne)
Fmt(abs = as.fmt(digits = 0, big.mark = "'"))

#> $abs
#> Format name:    abs (default)
#>  Description:   Number format for counts
#>  Definition:    digits=0, big.mark="'"
#>  Example:       314'159
#> 
#> $per
#> Format name:    per (default)
#>  Description:   Percentage number format
#>  Definition:    digits=1, fmt='%'
#>  Example:       31415926.5%
#> 
#> $num
#> Format name:    num (default)
#>  Description:   Number format for floats
#>  Definition:    digits=3, big.mark="'"
#>  Example:       314'159.265
#> 
#> $nob
#> Format name:    nob
#>  Description:   Number format
#>  Definition:    digits=5, na.form='nodat'
#>  Example:       314159.26536
#> 

Fmt(per = as.fmt(digits = 1, fmt = "%"))

#> $abs
#> Format name:    abs
#>  Description:   Number format
#>  Definition:    digits=0, big.mark="'"
#>  Example:       314'159
#> 
#> $per
#> Format name:    per (default)
#>  Description:   Percentage number format
#>  Definition:    digits=1, fmt='%'
#>  Example:       31415926.5%
#> 
#> $num
#> Format name:    num (default)
#>  Description:   Number format for floats
#>  Definition:    digits=3, big.mark="'"
#>  Example:       314'159.265
#> 
#> $nob
#> Format name:    nob
#>  Description:   Number format
#>  Definition:    digits=5, na.form='nodat'
#>  Example:       314159.26536
#> 

Fmt(num = as.fmt(digits = 1, big.mark = "'"))

#> $abs
#> Format name:    abs
#>  Description:   Number format
#>  Definition:    digits=0, big.mark="'"
#>  Example:       314'159
#> 
#> $per
#> Format name:    per
#>  Description:   Number format
#>  Definition:    digits=1, fmt='%'
#>  Example:       31415926.5%
#> 
#> $num
#> Format name:    num (default)
#>  Description:   Number format for floats
#>  Definition:    digits=3, big.mark="'"
#>  Example:       314'159.265
#> 
#> $nob
#> Format name:    nob
#>  Description:   Number format
#>  Definition:    digits=5, na.form='nodat'
#>  Example:       314159.26536
#> 

TOne(x = d.pizza[, c("temperature", "delivery_min", "driver", "wine_ordered")],
  grp = d.pizza$quality)

#> var                total              low                medium             high                                 
#> n                  1'008              156 (15.5%)        356 (35.3%)        496 (49.2%)                          
#> temperature        47.9 (9.9)         32.9 (7.8)         45.6 (7.4)         53.6 (6.5)         *** '             
#> delivery_min       25.7 (10.8)        33.9 (11.7)        26.5 (10.1)        22.6 (9.5)         *** '             
#> driver                                                                                         *** ""            
#>   Butcher          79 (8.0%)          10 (6.5%)          36 (10.1%)         33 (6.7%)                            
#>   Carpenter        225 (22.6%)        59 (38.1%)         90 (25.4%)         76 (15.4%)                           
#>   Carter           196 (19.4%)        11 (7.1%)          72 (20.3%)         113 (22.9%)                          
#>   Farmer           94 (9.7%)          10 (6.5%)          26 (7.3%)          58 (11.7%)                           
#>   Hunter           130 (13.0%)        8 (5.2%)           43 (12.1%)         79 (16.0%)                           
#>   Miller           109 (10.4%)        16 (10.3%)         35 (9.9%)          58 (11.7%)                           
#>   Taylor           171 (16.9%)        41 (26.5%)         53 (14.9%)         77 (15.6%)                           
#> wine_ordered (= 1) 161 (16.1%)        32 (20.8%)         63 (17.9%)         66 (13.4%)         .   ""            
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

# define median/IQR as describing functions for the numeric variables
TOne(iris[, -5], iris[, 5],
  FUN = function(x) {
    gettextf("%s / %s",
      Format(median(x, na.rm = TRUE), digits = 1),
      Format(IQR(x, na.rm = TRUE), digits = 3))
  }
)

#> var          total        setosa       versicolor   virginica                
#> n            150          50 (33.3%)   50 (33.3%)   50 (33.3%)               
#> Sepal.Length 5.8 / 1.300  5.0 / 0.400  5.9 / 0.700  6.5 / 0.675  *** '       
#> Sepal.Width  3.0 / 0.500  3.4 / 0.475  2.8 / 0.475  3.0 / 0.375  *** '       
#> Petal.Length 4.3 / 3.500  1.5 / 0.175  4.3 / 0.600  5.5 / 0.775  *** '       
#> Petal.Width  1.3 / 1.500  0.2 / 0.100  1.3 / 0.300  2.0 / 0.500  *** '       
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

# replace kruskal.test by ANOVA and report the p.value
# Change tests for all the types
TOne(x = iris[, -5], grp = iris[, 5],
     FUN = function(x) gettextf("%s / %s",
            Format(mean(x, na.rm = TRUE), digits = 1),
            Format(sd(x, na.rm = TRUE), digits = 3)),

     TEST = list(
       num = list(fun = function(x, g){summary(aov(x ~ g))[[1]][1, "Pr(>F)"]},
                        lbl = "ANOVA"),
               cat = list(fun = function(x, g){chisq.test(table(x, g))$p.val},
                        lbl = "Chi-Square test"),
               dich = list(fun = function(x, g){fisher.test(table(x, g))$p.val},
                         lbl = "Fisher exact test")),
     fmt.pval = as.fmt(fmt = "p", na.form = "   ")
)

#> var          total        setosa       versicolor   virginica                
#> n            150          50 (33.3%)   50 (33.3%)   50 (33.3%)               
#> Sepal.Length 5.8 / 0.828  5.0 / 0.352  5.9 / 0.516  6.6 / 0.636  < 2.2e-16 ' 
#> Sepal.Width  3.1 / 0.436  3.4 / 0.379  2.8 / 0.314  3.0 / 0.322  < 2.2e-16 ' 
#> Petal.Length 3.8 / 1.765  1.5 / 0.174  4.3 / 0.470  5.6 / 0.552  < 2.2e-16 ' 
#> Petal.Width  1.2 / 0.762  0.2 / 0.105  1.3 / 0.198  2.0 / 0.275  < 2.2e-16 ' 
#> ---
#> ') ANOVA, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

# dichotomous integer or logical values can be reported by the high or low value
x <- sample(x = c(0, 1), size = 100, prob = c(0.3, 0.7), replace = TRUE)
y <- sample(x = c(0, 1), size = 100, prob = c(0.3, 0.7), replace = TRUE) == 1
z <- factor(sample(x = c(0, 1), size = 100, prob = c(0.3, 0.7), replace = TRUE))
g <- sample(x = letters[1:4], size = 100, replace = TRUE)
d.set <- data.frame(x = x, y = y, z = z, g = g)

TOne(d.set[1:3], d.set$g, intref = "low")

#> Warning: Chi-squared approximation may be incorrect

#> var         total       a           b           c           d                      
#> n           100         28 (28.0%)  23 (23.0%)  19 (19.0%)  30 (30.0%)             
#> x (= 0)     28 (28.0%)  10 (35.7%)  5 (21.7%)   4 (21.1%)   9 (30.0%)       ""     
#> y (= FALSE) 39 (39.0%)  12 (42.9%)  8 (34.8%)   9 (47.4%)   10 (33.3%)      ""     
#> z (= 0)     26 (26.0%)  10 (35.7%)  5 (21.7%)   4 (21.1%)   7 (23.3%)       ""     
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

TOne(d.set[1:3], d.set$g, intref = "high")

#> Warning: Chi-squared approximation may be incorrect

#> var        total      a          b          c          d                    
#> n          100        28 (28.0%) 23 (23.0%) 19 (19.0%) 30 (30.0%)           
#> x (= 1)    72 (72.0%) 18 (64.3%) 18 (78.3%) 15 (78.9%) 21 (70.0%)     ""    
#> y (= TRUE) 61 (61.0%) 16 (57.1%) 15 (65.2%) 10 (52.6%) 20 (66.7%)     ""    
#> z (= 0)    26 (26.0%) 10 (35.7%) 5 (21.7%)  4 (21.1%)  7 (23.3%)      ""    
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

# intref would not control factors, use relevel to change reported value
TOne(data.frame(z = relevel(z, "1")), g)

#> Warning: Chi-squared approximation may be incorrect

#> var        total      a          b          c          d                    
#> n          100        28 (28.0%) 23 (23.0%) 19 (19.0%) 30 (30.0%)           
#> z (= 1)    74 (74.0%) 18 (64.3%) 18 (78.3%) 15 (78.9%) 23 (76.7%)     ""    
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

TOne(data.frame(z = z), g)

#> Warning: Chi-squared approximation may be incorrect

#> var        total      a          b          c          d                    
#> n          100        28 (28.0%) 23 (23.0%) 19 (19.0%) 30 (30.0%)           
#> z (= 0)    26 (26.0%) 10 (35.7%) 5 (21.7%)  4 (21.1%)  7 (23.3%)      ""    
#> ---
#> ') Kruskal-Wallis test, ") Fisher exact test, "") Chi-Square test
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1 
#> 

options(opt)

if (FALSE) {

# Send the whole stuff to Word
wrd <- GetNewWrd()
ToWrd(
  TOne(x = d.pizza[, c("temperature", "delivery_min", "driver", "wine_ordered")],
       grp = d.pizza$quality)
)

}