Vigenere Cypher

Implements a Vigenere cypher, both encryption and decryption. The function handle keys and text of unequal length and discards non-alphabetic characters.

Vigenere(x, key = NULL, decrypt = FALSE)

Arguments

x
the text to be encrypted

key
the key to be used. If this remains to NULL the PasswordDlg will be presented and the key can be entered there.

decrypt
boolean defining if the text should be encrypted or decrypted.

Details

All characters beside charlist = c(LETTERS, letters, 0:9) will be discarded from the text and from the key.

Value

the encrypted, resp. decrypted text

Author

Andri Signorell <andri@signorell.net>
strongly based on code found at http://rosettacode.org/wiki/Vigen%C3%A8re_cipher#R (credits to the unknown soldier)

Examples

key <- "My FavoriteKey452"
(xenc <- Vigenere("Beware the Jabberwock, my son! The jaws that bite, the claws that catch!", key))

#> [1] "NS10cIaFNnk5PYmo0QpCjWVLCBoDOqnltOy1TXLRQ8mFOqnltOy2LXJF"

Vigenere(xenc, key, decrypt = TRUE)

#> [1] "BewaretheJabberwockmysonThejawsthatbitetheclawsthatcatch"

# note that everything besides the characters in the list will be discarded