Woolf Test For Homogeneity in 2x2xk Tables

Test for homogeneity on \(2 \times 2 \times k\) tables over strata (i.e., whether the log odds ratios are the same in all strata).

WoolfTest(x)

Arguments

x
a \(2 \times 2 \times k\) table, where the last dimension refers to the strata.

Value

A list of class "htest" containing the following components:

statistic

the chi-squared test statistic.

parameter

degrees of freedom of the approximate chi-squared distribution of the test statistic.

p.value

\(p\)-value for the test.

method

a character string indicating the type of test performed.

data.name

a character string giving the name(s) of the data.

observed

the observed counts.

expected

the expected counts under the null hypothesis.

See also

Note

This function was previously published as woolf_test() in the vcd package and has been integrated here without logical changes.

Author

David Meyer, Achim Zeileis, Kurt Hornik, Michael Friendly

References

Woolf, B. 1955: On estimating the relation between blood group and disease. Ann. Human Genet. (London) 19, 251-253.

Examples

migraine <- xtabs(freq ~ .,
            cbind(expand.grid(treatment=c("active","placebo"),
                               response=c("better","same"),
                               gender=c("female","male")),
                  freq=c(16,5,11,20,12,7,16,19))
            )

WoolfTest(migraine)

#> 
#> 	Woolf Test on Homogeneity of Odds Ratios (no 3-Way assoc.)
#> 
#> data:  migraine
#> X-squared = 1, df = 1, p-value = 0.2
#>