ISO 3166-1 Country Codes

Country codes published by the International Organization for Standardization (ISO) define codes for the names of countries, dependent territories, and special areas of geographical interest.

data("d.countries")

Format

A data frame with 249 observations on the following 4 variables.

name

a character vector, the name of the country.

a2

a character vector, two-letter country codes (aka alpha-2) which are the most widely used of the three, and used most prominently for the Internet's country code top-level domains (with a few exceptions).

a3

a character vector, three-letter country codes (aka alpha-3) which allow a better visual association between the codes and the country names than the alpha-2 codes.

code

a numeric vector, three-digit country codes which are identical to those developed and maintained by the United Nations Statistics Division, with the advantage of script (writing system) independence, and hence useful for people or systems using non-Latin scripts.

region

the region of the country. One of "East Asia & Pacific" (35), "Europe & Central Asia" (52), "Latin America & Caribbean" (41), "Middle East & North Africa" (20), "North America" (3), "South Asia" (8), "Sub-Saharan Africa" (47)

pop2012

the population in 2012

gcpi2012

the gross national income (per capita) in dollars per country in 2012.

latitude

geographic coordinate that specifies the north–south position of a point on the Earth's surface. Latitude is an angle (defined below) which ranges from 0° at the Equator to 90° (North or South) at the poles.

longitude

geographic coordinate that specifies the east–west position of a point on the Earth's surface, or the surface of a celestial body

References

https://en.wikipedia.org/wiki/ISO_3166-1
https://datacatalog.worldbank.org/search/dataset/0037652

Examples

head(d.countries)
#>                   name a2  a3 code                     region  pop2012 gcpi2012
#> 1              Andorra AD AND   20                       <NA>       NA       NA
#> 2 United Arab Emirates AE ARE  784 Middle East & North Africa  9205651    38620
#> 3          Afghanistan AF AFG    4                 South Asia 29824536      690
#> 4  Antigua and Barbuda AG ATG   28  Latin America & Caribbean    89069    12720
#> 5             Anguilla AI AIA  660                       <NA>       NA       NA
#> 6              Albania AL ALB    8      Europe & Central Asia  2801681     4520
#>   latitude longitude
#> 1     42.5       1.6
#> 2     23.4      53.8
#> 3     33.9      67.7
#> 4     17.1     -61.8
#> 5     18.2     -63.1
#> 6     41.2      20.2