Data diamonds

As I suppose, an artificial dataset

data(d.diamonds)

Format

A data frame with 440 observations on the following 10 variables.

index

a numeric vector

carat

a numeric vector

colour

a factor with levels D E F G H I J K L

clarity

an ordered factor with levels I2 < I1 < SI3 < SI2 < SI1 < VS2 < VS1 < VVS2 < VVS1

cut

an ordered factor with levels F < G < V < X < I

certification

a factor with levels AGS DOW EGL GIA IGI

polish

an ordered factor with levels F < G < V < X < I

symmetry

an ordered factor with levels F < G < V < X < I

price

a numeric vector

wholesaler

a factor with levels A B C

Details

P Poor F Fair G Good V Very good X Excellent I Ideal

Source

somewhere from the net...

Examples

data(d.diamonds)
str(d.diamonds)

#> 'data.frame':	440 obs. of  10 variables:
#>  $ index        : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ carat        : num  0.92 0.92 0.82 0.81 0.9 0.87 0.8 0.84 0.8 0.8 ...
#>  $ colour       : Factor w/ 9 levels "D","E","F","G",..: 6 6 3 4 7 3 1 3 1 1 ...
#>  $ clarity      : Ord.factor w/ 9 levels "I2"<"I1"<"SI3"<..: 4 4 4 5 6 4 4 5 4 4 ...
#>  $ cut          : Ord.factor w/ 5 levels "F"<"G"<"V"<"X"<..: 2 3 5 5 3 5 5 2 3 3 ...
#>  $ certification: Factor w/ 5 levels "AGS","DOW","EGL",..: 1 1 4 4 4 1 4 4 4 4 ...
#>  $ polish       : Ord.factor w/ 5 levels "F"<"G"<"V"<"X"<..: 3 2 4 4 3 2 3 3 3 3 ...
#>  $ symmetry     : Ord.factor w/ 5 levels "F"<"G"<"V"<"X"<..: 3 2 4 3 3 3 3 3 3 4 ...
#>  $ price        : int  3000 3000 3004 3004 3006 3007 3008 3010 3012 3012 ...
#>  $ wholesaler   : Factor w/ 3 levels "A","B","C": 1 1 1 1 1 1 1 1 1 1 ...