Data pizza

An artificial dataset inspired by a similar dataset pizza.sav in Arbeitsbuch zur deskriptiven und induktiven Statistik by Toutenburg et.al.
The dataset contains data of a pizza delivery service in London, delivering pizzas to three areas. Every record defines one order/delivery and the according properties. A pizza is supposed to taste good, if its temperature is high enough, say 45 Celsius. So it might be interesting for the pizza delivery service to minimize the delivery time.
The dataset is designed to be as evil as possible. As far as the description is concerned, it should pose the same difficulties that we have to deal with in everyday life. It contains the most used datatypes as numerics, factors, ordered factors, integers, logicals and a date. NAs are scattered everywhere partly systematically, partly randomly (except in the index).

data(d.pizza)

Format

A data frame with 1209 observations on the following 17 variables.

index

a numeric vector, indexing the records (no missings here).

date

Date, the delivery date

week

integer, the weeknumber

weekday

integer, the weekday

area

factor, the three London districts: Brent, Camden, Westminster

count

integer, the number of pizzas delivered

rabate

logical, TRUE if a rabate has been given

price

numeric, the total price of delivered pizza(s)

operator

a factor with levels Allanah Maria Rhonda

driver

a factor with levels Carpenter Carter Taylor Butcher Hunter Miller Farmer

delivery_min

numeric, the delivery time in minutes (decimal)

temperature

numeric, the temperature of the pizza in degrees Celsius when delivered to the customer

wine_ordered

integer, 1 if wine was ordered, 0 if not

wine_delivered

integer, 1 if wine was delivered, 0 if not

wrongpizza

logical, TRUE if a wrong pizza was delivered

quality

ordered factor with levels low < medium < high, defining the quality of the pizza when delivered

Details

The dataset contains NAs randomly scattered.

References

Toutenburg H, Schomaker M, Wissmann M, Heumann C (2009): Arbeitsbuch zur deskriptiven und induktiven Statistik Springer, Berlin Heidelberg

Examples

str(d.pizza)
#> 'data.frame':	1209 obs. of  16 variables:
#>  $ index         : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ date          : Date, format: "2014-03-01" "2014-03-01" ...
#>  $ week          : num  9 9 9 9 9 9 9 9 9 9 ...
#>  $ weekday       : num  6 6 6 6 6 6 6 6 6 6 ...
#>  $ area          : Factor w/ 3 levels "Brent","Camden",..: 2 3 3 1 1 2 2 1 3 1 ...
#>  $ count         : int  5 2 3 2 5 1 4 NA 3 6 ...
#>  $ rabate        : logi  TRUE FALSE FALSE FALSE TRUE FALSE ...
#>  $ price         : num  65.7 27 41 26 57.6 ...
#>  $ operator      : Factor w/ 3 levels "Allanah","Maria",..: 3 3 1 1 3 1 3 1 1 3 ...
#>  $ driver        : Factor w/ 7 levels "Butcher","Carpenter",..: 7 1 1 7 3 7 7 7 7 3 ...
#>  $ delivery_min  : num  20 19.6 17.8 37.3 21.8 48.7 49.3 25.6 26.4 24.3 ...
#>  $ temperature   : num  53 56.4 36.5 NA 50 27 33.9 54.8 48 54.4 ...
#>  $ wine_ordered  : int  0 0 0 0 0 0 1 NA 0 1 ...
#>  $ wine_delivered: int  0 0 0 0 0 0 1 NA 0 1 ...
#>  $ wrongpizza    : logi  FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ quality       : Ord.factor w/ 3 levels "low"<"medium"<..: 2 3 NA NA 2 1 1 3 3 2 ...
head(d.pizza)
#>   index       date week weekday        area count rabate price operator  driver
#> 1     1 2014-03-01    9       6      Camden     5   TRUE  65.7   Rhonda  Taylor
#> 2     2 2014-03-01    9       6 Westminster     2  FALSE  27.0   Rhonda Butcher
#> 3     3 2014-03-01    9       6 Westminster     3  FALSE  41.0  Allanah Butcher
#> 4     4 2014-03-01    9       6       Brent     2  FALSE  26.0  Allanah  Taylor
#> 5     5 2014-03-01    9       6       Brent     5   TRUE  57.6   Rhonda  Carter
#> 6     6 2014-03-01    9       6      Camden     1  FALSE  14.0  Allanah  Taylor
#>   delivery_min temperature wine_ordered wine_delivered wrongpizza quality
#> 1         20.0        53.0            0              0      FALSE  medium
#> 2         19.6        56.4            0              0      FALSE    high
#> 3         17.8        36.5            0              0      FALSE    <NA>
#> 4         37.3          NA            0              0      FALSE    <NA>
#> 5         21.8        50.0            0              0      FALSE  medium
#> 6         48.7        27.0            0              0      FALSE     low

Desc(d.pizza)
#> Error in .Call("_DescTools_n_pow_sum", PACKAGE = "DescTools", x): "_DescTools_n_pow_sum" not available for .Call() for package "DescTools"