Power Calculations for ChiSquared Tests

Compute power of test or determine parameters to obtain target power (same as power.anova.test).

power.chisq.test(n = NULL, w = NULL, df = NULL, sig.level = 0.05, power = NULL)

Arguments

n

total number of observations.

w

effect size.

df

degree of freedom (depends on the chosen test.

sig.level

Significance level (Type I error probability).

power

Power of test (1 minus Type II error probability).

Details

Exactly one of the parameters w, n, power or sig.level must be passed as NULL, and this parameter is determined from the others. Note that the last one has non-NULL default, so NULL must be explicitly passed, if you want to compute it.

Value

Object of class "power.htest", a list of the arguments (including the computed one) augmented with 'method' and 'note' elements.

References

Cohen, J. (1988) Statistical power analysis for the behavioral sciences (2nd ed.) Hillsdale, NJ: Lawrence Erlbaum.

Author

Stephane Champely <champely@univ-lyon1.fr>
but this is a mere copy of Peter Dalgaard's work on power.t.test

Note

uniroot is used to solve power equation for unknowns, so you may see errors from it, notably about inability to bracket the root when invalid arguments are given.

See also

Examples

## Exercise 7.1 P. 249 from Cohen (1988) 
power.chisq.test(w=0.289, df=(4-1), n=100, sig.level=0.05)
#> 
#>      Chi squared power calculation 
#> 
#>               w = 0.289
#>               n = 100
#>              df = 3
#>       sig.level = 0.05
#>           power = 0.675
#> 
#> NOTE: n is the number of observations
#> 

## Exercise 7.3 p. 251
power.chisq.test(w=0.346, df=(2-1)*(3-1), n=140, sig.level=0.01)
#> 
#>      Chi squared power calculation 
#> 
#>               w = 0.346
#>               n = 140
#>              df = 2
#>       sig.level = 0.01
#>           power = 0.885
#> 
#> NOTE: n is the number of observations
#> 

## Exercise 7.8 p. 270
power.chisq.test(w=0.1, df=(5-1)*(6-1), power=0.80, sig.level=0.05)
#> 
#>      Chi squared power calculation 
#> 
#>               w = 0.1
#>               n = 2096
#>              df = 20
#>       sig.level = 0.05
#>           power = 0.8
#> 
#> NOTE: n is the number of observations
#>