Formula Interface for Split

Implementation of a simple formula interface for the split function.

# S3 method for formula
split (x, f, drop = FALSE, data = NULL, ...)

Arguments

x
a formula of the form y ~ x.

f
a 'factor' in the sense that as.factor(f) defines the grouping, or a list of such factors in which case their interaction is used for the grouping.

drop
logical indicating if levels that do not occur should be dropped (if f is a factor or a list). Defaults to FALSE.

data
the data frame from which the formula should be evaluated.

...
other arguments to be passed to split.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

split(extra ~ group, data = sleep)

#> $`1`
#>  [1]  0.7 -1.6 -0.2 -1.2 -0.1  3.4  3.7  0.8  0.0  2.0
#> 
#> $`2`
#>  [1]  1.9  0.8  1.1  0.1 -0.1  4.4  5.5  1.6  4.6  3.4
#>