Cochran's Q test

Perform the Cochran's Q test for unreplicated randomized block design experiments with a binary response variable and paired data.

CochranQTest(y, ...)

# S3 method for default
CochranQTest(y, groups, blocks, ...)

# S3 method for formula
CochranQTest(formula, data, subset, na.action, ...)

Arguments

y
either a numeric vector of data values, or a data matrix.

groups
a vector giving the group for the corresponding elements of y if this is a vector; ignored if y is a matrix. If not a factor object, it is coerced to one.

blocks
a vector giving the block for the corresponding elements of y if this is a vector; ignored if y is a matrix. If not a factor object, it is coerced to one.

formula
a formula of the form y ~ groups | blocks.

data
an optional matrix or data frame (or similar: see model.frame) containing the variables in the formula. By default the variables are taken from environment(formula).

subset
an optional vector specifying a subset of observations to be used.

na.action
a function which indicates what should happen when the data contain NAs. Defaults to getOption("na.action").

...
further arguments to be passed to or from methods.

Details

CochranQTest() can be used for analyzing unreplicated complete block designs (i.e., there is exactly one binary observation in y for each combination of levels of groups and blocks) where the normality assumption may be violated.

The null hypothesis is that apart from an effect of blocks, the location parameter of y is the same in each of the groups.

If y is a matrix, groups and blocks are obtained from the column and row indices, respectively. NA's are not allowed in groups or blocks; if y contains NA's, corresponding blocks are removed.

Note that Cochran's Q Test is analogue to the Friedman test with 0, 1 coded response. This is used here for a simple implementation.

Value

A list with class htest containing the following components:

statistic

the value of Cochran's chi-squared statistic.

parameter

the degrees of freedom of the approximate chi-squared distribution of the test statistic.

p.value

the p-value of the test.

method

the character string "Cochran's Q-Test".

data.name

a character string giving the names of the data.

References

Cochran, W.G. (1950) The Comparison of Percentages in Matched Samples. Biometrika. 37 (3/4): 256-266. doi:10.1093/biomet/37.3-4.256. JSTOR 2332378.

Author

Andri Signorell <andri@signorell.net>

Examples

# example in: 
# http://support.sas.com/documentation/cdl/en/statugfreq/63124/PDF/default/statugfreq.pdf
# pp. S. 1824

# use expand.grid, xtabs and Untable to create the dataset
d.frm <- Untable(xtabs(c(6,2,2,6,16,4,4,6) ~ ., 
    expand.grid(rep(list(c("F","U")), times=3))), 
    colnames = LETTERS[1:3])

# rearrange to long shape    
d.long <- reshape(d.frm, varying=1:3, times=names(d.frm)[c(1:3)], 
                  v.names="resp", direction="long")

# after having done the hard work of data organisation, performing the test is a piece of cake....
CochranQTest(resp ~ time | id, data=d.long)

#> 
#> 	Cochran's Q test
#> 
#> data:  resp and time and id
#> Q = 8.4706, df = 2, p-value = 0.01448
#> 

# and let's perform a post hoc analysis using mcnemar's test
z <- split(d.long, f=d.long$time)
pairwise.table(function(i, j) { 
    mcnemar.test(z[[i]]$resp, z[[j]]$resp, correct=FALSE)$p.value
  }, 
  level.names = names(z), 
  p.adjust.method = "fdr"
)

#>           A         B
#> B 1.0000000        NA
#> C 0.0350133 0.0350133