List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BinToDec"              
#>  [46] "BinTree"                "BlankIfNA"              "BootCI"                
#>  [49] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [52] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [55] "BrierScore"             "BubbleLegend"           "Canvas"                
#>  [58] "CartToPol"              "CartToSph"              "CatTable"              
#>  [61] "CCC"                    "CharToAsc"              "Clockwise"             
#>  [64] "Closest"                "CmToPts"                "CmykToCmy"             
#>  [67] "CmykToRgb"              "CmyToCmyk"              "Coalesce"              
#>  [70] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [73] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [76] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [79] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [82] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [85] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [88] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [91] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [94] "ConvUnit"               "Cor"                    "CorCI"                 
#>  [97] "CorPart"                "CorPolychor"            "CountCompCases"        
#> [100] "CountWorkDays"          "CourseData"             "Cov"                   
#> [103] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [106] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [109] "Cstat"                  "CutQ"                   "Day"                   
#> [112] "Day<-"                  "DB"                     "dBenf"                 
#> [115] "DecToBin"               "DecToHex"               "DecToOct"              
#> [118] "DegToRad"               "DenseRank"              "Desc"                  
#> [121] "DescToolsOptions"       "dExtrVal"               "dFrechet"              
#> [124] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [127] "dGumbel"                "DiffDays360"            "DigitSum"              
#> [130] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [133] "dNegWeibull"            "DoBy"                   "DoCall"                
#> [136] "dOrder"                 "Dot"                    "DrawArc"               
#> [139] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [142] "DrawEllipse"            "DrawRegPolygon"         "dRevGumbel"            
#> [145] "dRevWeibull"            "Dummy"                  "DunnettTest"           
#> [148] "DunnTest"               "DurbinWatsonTest"       "Entropy"               
#> [151] "Eps"                    "ErrBars"                "EtaSq"                 
#> [154] "ExpFreq"                "Factorize"              "FctArgs"               
#> [157] "Fibonacci"              "FindColor"              "FindCorr"              
#> [160] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [163] "FixToTable"             "Flags"                  "Fmt"                   
#> [166] "Format"                 "Frac"                   "Freq"                  
#> [169] "Freq2D"                 "GCD"                    "GeomSn"                
#> [172] "GeomTrans"              "GetCurrPP"              "GetCurrWrd"            
#> [175] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [178] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [181] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [184] "Gsd"                    "GTest"                  "Herfindahl"            
#> [187] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [190] "HighLow"                "Hmean"                  "HmsToSec"              
#> [193] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [196] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [199] "HunterGaston"           "ICC"                    "identify.formula"      
#> [202] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [205] "InDots"                 "Interval"               "IPMT"                  
#> [208] "IQRw"                   "IRR"                    "IsDate"                
#> [211] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [214] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [217] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [220] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [223] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [226] "KendallW"               "Keywords"               "KrippAlpha"            
#> [229] "Kurt"                   "Label"                  "Label<-"               
#> [232] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [235] "Lc"                     "LCM"                    "LehmacherTest"         
#> [238] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [241] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [244] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [247] "LOCF"                   "LOF"                    "Logit"                 
#> [250] "LogitInv"               "LogSt"                  "LogStInv"              
#> [253] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [256] "MAD"                    "MADCI"                  "MAE"                   
#> [259] "MAPE"                   "Mar"                    "Margins"               
#> [262] "MaxDigits"              "Mean"                   "MeanAD"                
#> [265] "MeanCI"                 "MeanDiffCI"             "MeanSE"                
#> [268] "Median"                 "MedianCI"               "Mgp"                   
#> [271] "MHChisqTest"            "Midx"                   "Minute"                
#> [274] "MixColor"               "Mode"                   "ModSummary"            
#> [277] "Month"                  "MosesTest"              "MoveAvg"               
#> [280] "MSE"                    "MultinomCI"             "MultMerge"             
#> [283] "MutInf"                 "N"                      "NAIfBlank"             
#> [286] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [289] "NMAE"                   "NMSE"                   "Now"                   
#> [292] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [295] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [298] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [301] "PageTest"               "PairApply"              "Pal"                   
#> [304] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [307] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [310] "PearsonTest"            "PercentRank"            "PercTable"             
#> [313] "Permn"                  "pExtrVal"               "pFrechet"              
#> [316] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [319] "pGumbel"                "Phi"                    "Phrase"                
#> [322] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [325] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [328] "PlotACF"                "PlotArea"               "PlotBag"               
#> [331] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [334] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [337] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [340] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [343] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [346] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [349] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [352] "PlotPairs"              "PlotPolar"              "PlotPyramid"           
#> [355] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [358] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [361] "PMT"                    "pNegWeibull"            "PoissonCI"             
#> [364] "PolarGrid"              "PolToCart"              "pOrder"                
#> [367] "PostHocTest"            "power.chisq.test"       "PpAddSlide"            
#> [370] "PPMT"                   "PpPlot"                 "PpText"                
#> [373] "Prec"                   "pRevGumbel"             "pRevWeibull"           
#> [376] "Primes"                 "print.Assocs"           "print.Conf"            
#> [379] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [382] "print.HoeffD"           "print.ICC"              "print.mtest"           
#> [385] "print.PercTable"        "print.PostHocTest"      "PseudoR2"              
#> [388] "PtInPoly"               "PtsToCm"                "qBenf"                 
#> [391] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [394] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [397] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [400] "qRevWeibull"            "Quantile"               "QuantileCI"            
#> [403] "Quarter"                "Quot"                   "RadToDeg"              
#> [406] "Range"                  "RBAL"                   "rBenf"                 
#> [409] "Recode"                 "Recycle"                "RelRisk"               
#> [412] "Rename"                 "reorder.factor"         "Rev"                   
#> [415] "RevCode"                "rExtrVal"               "rFrechet"              
#> [418] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [421] "RgbToLong"              "rGenExtrVal"            "rGenPareto"            
#> [424] "rGompertz"              "rGumbel"                "RMSE"                  
#> [427] "RndPairs"               "RndWord"                "rNegWeibull"           
#> [430] "RobScale"               "RomanToInt"             "rOrder"                
#> [433] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [436] "rRevGumbel"             "rRevWeibull"            "RSessionAlive"         
#> [439] "RTempdirAlive"          "RunsTest"               "Sample"                
#> [442] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [445] "SD"                     "Second"                 "SecToHms"              
#> [448] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [451] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [454] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [457] "Skew"                   "SLN"                    "Small"                 
#> [460] "SMAPE"                  "SmoothSpline"           "Some"                  
#> [463] "SomersDelta"            "Sort"                   "SortMixed"             
#> [466] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [469] "split.formula"          "SplitAt"                "SplitPath"             
#> [472] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [475] "StdCoef"                "Str"                    "StrAbbr"               
#> [478] "StrAlign"               "Strata"                 "StrCap"                
#> [481] "StrChop"                "StrCountW"              "StrDist"               
#> [484] "StrExtract"             "StripAttr"              "StrIsNumeric"          
#> [487] "StrLeft"                "StrPad"                 "StrPos"                
#> [490] "StrRev"                 "StrRight"               "StrSpell"              
#> [493] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [496] "StrVal"                 "StuartMaxwellTest"      "StuartTauC"            
#> [499] "SYD"                    "SysInfo"                "TextContrastColor"     
#> [502] "TextToTable"            "TheilU"                 "Timezone"              
#> [505] "TitleRect"              "TMod"                   "Today"                 
#> [508] "ToLong"                 "TOne"                   "ToWide"                
#> [511] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [514] "ToXL"                   "Trim"                   "TschuprowT"            
#> [517] "TTestA"                 "TukeyBiweight"          "TwoGroups"             
#> [520] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [523] "Unit<-"                 "Untable"                "Unwhich"               
#> [526] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [529] "VarTest"                "VecRot"                 "VecShift"              
#> [532] "VIF"                    "Vigenere"               "VonNeumannTest"        
#> [535] "Week"                   "Weekday"                "Winsorize"             
#> [538] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [541] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [544] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [547] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [550] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [553] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [556] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [559] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [562] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [565] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [568] "XLSaveAs"               "XLView"                 "Year"                  
#> [571] "YearDay"                "YearMonth"              "YTM"                   
#> [574] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [577] "ZeroIfNA"               "Zodiac"                 "ZTest"