List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%:%"                    "%()%"                   "%(]%"                  
#>   [4] "%)(%"                   "%)[%"                   "%[)%"                  
#>   [7] "%[]%"                   "%](%"                   "%][%"                  
#>  [10] "%^%"                    "%c%"                    "%like any%"            
#>  [13] "%like%"                 "%nin%"                  "%overlaps%"            
#>  [16] "Abind"                  "Abstract"               "AddMonths"             
#>  [19] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [22] "AllIdentical"           "AndersonDarlingTest"    "aovlDetails"           
#>  [25] "aovlErrorTerms"         "Append"                 "Arrow"                 
#>  [28] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#>  [31] "AscToChar"              "Asp"                    "Assocs"                
#>  [34] "Atkinson"               "AUC"                    "AxisBreak"             
#>  [37] "axTicks.Date"           "axTicks.POSIXct"        "BarnardTest"           
#>  [40] "BartelsRankTest"        "BarText"                "BhapkarTest"           
#>  [43] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [46] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [49] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [52] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [55] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [58] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [61] "CartToSph"              "CatTable"               "CCC"                   
#>  [64] "CharToAsc"              "Clockwise"              "Closest"               
#>  [67] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [70] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [73] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [76] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [79] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [82] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [85] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [88] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [91] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [94] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [97] "Cor"                    "CorCI"                  "CorPart"               
#> [100] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [103] "CourseData"             "Cov"                    "CramerV"               
#> [106] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [109] "Cross"                  "CrossN"                 "Cstat"                 
#> [112] "CutQ"                   "Day"                    "Day<-"                 
#> [115] "DB"                     "dBenf"                  "DecToBin"              
#> [118] "DecToHex"               "DecToOct"               "DegToRad"              
#> [121] "DenseRank"              "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanDiffCI"            
#> [271] "MeanSE"                 "Median"                 "MedianCI"              
#> [274] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [277] "Minute"                 "MixColor"               "Mode"                  
#> [280] "ModSummary"             "Month"                  "MosesTest"             
#> [283] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [286] "MultMerge"              "MutInf"                 "N"                     
#> [289] "NAIfBlank"              "NAIfZero"               "Ndec"                  
#> [292] "NemenyiTest"            "NMAE"                   "NMSE"                  
#> [295] "Now"                    "NPV"                    "NPVFixBond"            
#> [298] "OctToDec"               "OddsRatio"              "OPR"                   
#> [301] "OrderMixed"             "ORToRelRisk"            "Outlier"               
#> [304] "Overlap"                "PageTest"               "PairApply"             
#> [307] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [310] "PartialSD"              "PasswordDlg"            "pBenf"                 
#> [313] "PDFManual"              "PearsonTest"            "PercentRank"           
#> [316] "PercTable"              "Permn"                  "pExtrVal"              
#> [319] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [322] "pGompertz"              "pGumbel"                "Phi"                   
#> [325] "Phrase"                 "plot.bagplot"           "plot.Conf"             
#> [328] "plot.Lc"                "plot.Lclist"            "plot.palette"          
#> [331] "plot.PostHocTest"       "PlotACF"                "PlotArea"              
#> [334] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [337] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [340] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [343] "PlotDot"                "PlotECDF"               "PlotFaces"             
#> [346] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [349] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [352] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [355] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [358] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [361] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [364] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [367] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [370] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [373] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [376] "PpText"                 "Prec"                   "pRevGumbel"            
#> [379] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [382] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [385] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [388] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [391] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [394] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [397] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [400] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [403] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [406] "QuantileCI"             "Quarter"                "Quot"                  
#> [409] "RadToDeg"               "Range"                  "RBAL"                  
#> [412] "rBenf"                  "Recode"                 "Recycle"               
#> [415] "RelRisk"                "Rename"                 "reorder.factor"        
#> [418] "Rev"                    "RevCode"                "rExtrVal"              
#> [421] "rFrechet"               "RgbToCmy"               "RgbToCol"              
#> [424] "RgbToHex"               "RgbToLong"              "rGenExtrVal"           
#> [427] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [430] "RMSE"                   "RndPairs"               "RndWord"               
#> [433] "rNegWeibull"            "RobScale"               "RomanToInt"            
#> [436] "rOrder"                 "Rosenbluth"             "Rotate"                
#> [439] "RoundTo"                "rRevGumbel"             "rRevWeibull"           
#> [442] "RSessionAlive"          "RTempdirAlive"          "RunsTest"              
#> [445] "Sample"                 "SampleTwins"            "SaveAs"                
#> [448] "ScheffeTest"            "SD"                     "Second"                
#> [451] "SecToHms"               "SendOutlookMail"        "Sens"                  
#> [454] "SetAlpha"               "SetNames"               "Shade"                 
#> [457] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [460] "SignTest"               "Skew"                   "SLN"                   
#> [463] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [466] "Some"                   "SomersDelta"            "Sort"                  
#> [469] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [472] "SphToCart"              "split.formula"          "SplitAt"               
#> [475] "SplitPath"              "SplitToCol"             "SpreadOut"             
#> [478] "Stamp"                  "StdCoef"                "Str"                   
#> [481] "StrAbbr"                "StrAlign"               "Strata"                
#> [484] "StrCap"                 "StrChop"                "StrCountW"             
#> [487] "StrDist"                "StrExtract"             "StrExtractBetween"     
#> [490] "StripAttr"              "StrIsNumeric"           "StrLeft"               
#> [493] "StrPad"                 "StrPos"                 "StrRev"                
#> [496] "StrRight"               "StrSpell"               "StrSplit"              
#> [499] "StrTrim"                "StrTrunc"               "StrVal"                
#> [502] "StuartMaxwellTest"      "StuartTauC"             "SYD"                   
#> [505] "SysInfo"                "TextContrastColor"      "TextToTable"           
#> [508] "TheilU"                 "Timezone"               "TitleRect"             
#> [511] "TMod"                   "Today"                  "ToLong"                
#> [514] "TOne"                   "ToWide"                 "ToWrd"                 
#> [517] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [520] "Trim"                   "TschuprowT"             "TTestA"                
#> [523] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [526] "UnirootAll"             "Unit"                   "Unit<-"                
#> [529] "Untable"                "Unwhich"                "VanWaerdenTest"        
#> [532] "Var"                    "VarCI"                  "VarTest"               
#> [535] "VecRot"                 "VecShift"               "VIF"                   
#> [538] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [541] "Weekday"                "Winsorize"              "WithOptions"           
#> [544] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [547] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [550] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [553] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [556] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [559] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [562] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [565] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [568] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [571] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [574] "XLView"                 "Year"                   "YearDay"               
#> [577] "YearMonth"              "YTM"                    "YuenTTest"             
#> [580] "YuleQ"                  "YuleY"                  "ZeroIfNA"              
#> [583] "Zodiac"                 "ZTest"