List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AndersonDarlingTest"   
#>  [22] "aovlDetails"            "aovlErrorTerms"         "Append"                
#>  [25] "Arrow"                  "as.CDateFmt"            "as.fmt"                
#>  [28] "as.matrix.xtabs"        "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AUC"                   
#>  [34] "AxisBreak"              "axTicks.Date"           "axTicks.POSIXct"       
#>  [37] "BarnardTest"            "BartelsRankTest"        "BarText"               
#>  [40] "BinomCI"                "BinomDiffCI"            "BinomRatioCI"          
#>  [43] "BinToDec"               "BinTree"                "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [58] "CatTable"               "CCC"                    "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "Coalesce"              
#>  [64] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [67] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [70] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [73] "ColToHex"               "ColToHsv"               "ColToRgb"              
#>  [76] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [79] "CombSet"                "ConDisPairs"            "Conf"                  
#>  [82] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [85] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [88] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [91] "CountCompCases"         "CourseData"             "Cov"                   
#>  [94] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#>  [97] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [100] "Cstat"                  "CutQ"                   "Day"                   
#> [103] "Day<-"                  "DB"                     "dBenf"                 
#> [106] "DecToBin"               "DecToHex"               "DecToOct"              
#> [109] "DegToRad"               "DenseRank"              "Desc"                  
#> [112] "DescToolsOptions"       "dExtrVal"               "dFrechet"              
#> [115] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [118] "dGumbel"                "DiffDays360"            "DigitSum"              
#> [121] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [124] "dNegWeibull"            "DoBy"                   "DoCall"                
#> [127] "dOrder"                 "Dot"                    "DrawArc"               
#> [130] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [133] "DrawEllipse"            "DrawRegPolygon"         "dRevGumbel"            
#> [136] "dRevWeibull"            "Dummy"                  "DunnettTest"           
#> [139] "DunnTest"               "DurbinWatsonTest"       "Entropy"               
#> [142] "Eps"                    "ErrBars"                "EtaSq"                 
#> [145] "ExpFreq"                "Factorize"              "FctArgs"               
#> [148] "Fibonacci"              "FindColor"              "FindCorr"              
#> [151] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [154] "FixToTable"             "Flags"                  "Fmt"                   
#> [157] "Format"                 "Frac"                   "Freq"                  
#> [160] "Freq2D"                 "GCD"                    "GeomSn"                
#> [163] "GeomTrans"              "GetCurrPP"              "GetCurrWrd"            
#> [166] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [169] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [172] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [175] "Gsd"                    "GTest"                  "Herfindahl"            
#> [178] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [181] "HighLow"                "Hmean"                  "HmsToSec"              
#> [184] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [187] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [190] "ICC"                    "identify.formula"       "IdentifyA"             
#> [193] "Impute"                 "InDots"                 "Interval"              
#> [196] "IPMT"                   "IRR"                    "IsDate"                
#> [199] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [202] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [205] "IsValidWrd"             "IsWeekend"              "IsWhole"               
#> [208] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [211] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [214] "KendallW"               "Keywords"               "KrippAlpha"            
#> [217] "Kurt"                   "Label"                  "Label<-"               
#> [220] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [223] "Lc"                     "LCM"                    "LehmacherTest"         
#> [226] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [229] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [232] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [235] "LOCF"                   "LOF"                    "Logit"                 
#> [238] "LogitInv"               "LogSt"                  "LogStInv"              
#> [241] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [244] "MAD"                    "MAE"                    "MAPE"                  
#> [247] "Mar"                    "Margins"                "MaxDigits"             
#> [250] "Mean"                   "MeanAD"                 "MeanCI"                
#> [253] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [256] "MedianCI"               "MHChisqTest"            "Midx"                  
#> [259] "Minute"                 "MixColor"               "Mode"                  
#> [262] "ModSummary"             "Month"                  "MosesTest"             
#> [265] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [268] "MutInf"                 "N"                      "NAIfBlank"             
#> [271] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [274] "NMAE"                   "NMSE"                   "Now"                   
#> [277] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [280] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [283] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [286] "PageTest"               "PairApply"              "Pal"                   
#> [289] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [292] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [295] "PearsonTest"            "PercentRank"            "PercTable"             
#> [298] "Permn"                  "pExtrVal"               "pFrechet"              
#> [301] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [304] "pGumbel"                "Phi"                    "Phrase"                
#> [307] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [310] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [313] "PlotACF"                "PlotArea"               "PlotBag"               
#> [316] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [319] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [322] "PlotCorr"               "PlotDev"                "PlotDot"               
#> [325] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [328] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [331] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [334] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [337] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [340] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [343] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [346] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [349] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [352] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [355] "PpPlot"                 "PpText"                 "Prec"                  
#> [358] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [361] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [364] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [367] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [370] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [373] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [376] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [379] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [382] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [385] "Quarter"                "RadToDeg"               "Range"                 
#> [388] "RBAL"                   "rBenf"                  "Recode"                
#> [391] "Recycle"                "RelRisk"                "Rename"                
#> [394] "reorder.factor"         "Rev"                    "rExtrVal"              
#> [397] "rFrechet"               "RgbToCol"               "RgbToLong"             
#> [400] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [403] "rGumbel"                "RMSE"                   "RndPairs"              
#> [406] "RndWord"                "rNegWeibull"            "RobScale"              
#> [409] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [412] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [415] "rRevWeibull"            "RunsTest"               "Sample"                
#> [418] "SampleTwins"            "ScheffeTest"            "SD"                    
#> [421] "Second"                 "SecToHms"               "Sens"                  
#> [424] "SetAlpha"               "SetNames"               "Shade"                 
#> [427] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [430] "SignTest"               "Skew"                   "SLN"                   
#> [433] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [436] "Some"                   "SomersDelta"            "Sort"                  
#> [439] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [442] "SphToCart"              "split.formula"          "SplitAt"               
#> [445] "SplitPath"              "SpreadOut"              "Stamp"                 
#> [448] "StdCoef"                "Str"                    "StrAbbr"               
#> [451] "StrAlign"               "Strata"                 "StrCap"                
#> [454] "StrChop"                "StrCountW"              "StrDist"               
#> [457] "StrExtract"             "StrIsNumeric"           "StrLeft"               
#> [460] "StrPad"                 "StrPos"                 "StrRev"                
#> [463] "StrRight"               "StrSpell"               "StrTrim"               
#> [466] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [469] "StuartTauC"             "SYD"                    "SysInfo"               
#> [472] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [475] "Timezone"               "TitleRect"              "TMod"                  
#> [478] "Today"                  "ToLong"                 "TOne"                  
#> [481] "ToWide"                 "ToWrd"                  "ToXL"                  
#> [484] "Trim"                   "TschuprowT"             "TTestA"                
#> [487] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [490] "UnirootAll"             "Unit"                   "Unit<-"                
#> [493] "Untable"                "Unwhich"                "Var"                   
#> [496] "VarCI"                  "VarTest"                "VecRot"                
#> [499] "VecShift"               "VIF"                    "Vigenere"              
#> [502] "VonNeumannTest"         "Week"                   "Weekday"               
#> [505] "Winsorize"              "WoolfTest"              "WrdCaption"            
#> [508] "WrdCellRange"           "WrdFont"                "WrdFont<-"             
#> [511] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [514] "WrdKill"                "WrdMergeCells"          "WrdPageBreak"          
#> [517] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [520] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [523] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [526] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [529] "XLGetWorkbook"          "XLKill"                 "XLView"                
#> [532] "Year"                   "YearDay"                "YearMonth"             
#> [535] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [538] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [541] "ZTest"