List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [25] "AllIdentical"           "AndersonDarlingTest"    "Append"                
#>  [28] "Arrow"                  "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AxisBreak"             
#>  [34] "BarText"                "BarnardTest"            "BartelsRankTest"       
#>  [37] "Bg"                     "BhapkarTest"            "BinToDec"              
#>  [40] "BinTree"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "CCC"                    "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "CmToPts"               
#>  [64] "CmyToCmyk"              "CmykToCmy"              "CmykToRgb"             
#>  [67] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [70] "CoefVar"                "CohenD"                 "CohenKappa"            
#>  [73] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [76] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [79] "CollapseTable"          "ColorLegend"            "ColumnWrap"            
#>  [82] "CombN"                  "CombPairs"              "CombSet"               
#>  [85] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [88] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [91] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [94] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [97] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [100] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CutQ"                   "DB"                    
#> [109] "Day"                    "Day<-"                  "DecToBin"              
#> [112] "DecToHex"               "DecToOct"               "DegToRad"              
#> [115] "Desc"                   "DescToolsOptions"       "DiffDays360"           
#> [118] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [121] "Divisors"               "DoBy"                   "DoCall"                
#> [124] "Dot"                    "DrawArc"                "DrawBand"              
#> [127] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [130] "DrawRegPolygon"         "Dummy"                  "DunnTest"              
#> [133] "DunnettTest"            "DurbinWatsonTest"       "Entropy"               
#> [136] "Eps"                    "ErrBars"                "EtaSq"                 
#> [139] "ExpFreq"                "Factorize"              "FctArgs"               
#> [142] "Fibonacci"              "FindColor"              "FindCorr"              
#> [145] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [148] "FixToTable"             "Flags"                  "Fmt"                   
#> [151] "Format"                 "Frac"                   "Freq"                  
#> [154] "Freq2D"                 "GCD"                    "GTest"                 
#> [157] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [160] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [163] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [166] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [169] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [172] "Gsd"                    "Herfindahl"             "HexToCol"              
#> [175] "HexToDec"               "HexToRgb"               "HighLow"               
#> [178] "Hmean"                  "HmsToMinute"            "HmsToSec"              
#> [181] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [184] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [187] "HunterGaston"           "ICC"                    "IPMT"                  
#> [190] "IQRw"                   "IRR"                    "IdentifyA"             
#> [193] "Impute"                 "ImputeKnn"              "InDots"                
#> [196] "Interval"               "IsDate"                 "IsDichotomous"         
#> [199] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [202] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [205] "IsWeekend"              "IsWhole"                "IsZero"                
#> [208] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [211] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [214] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [217] "LCM"                    "LOCF"                   "LOF"                   
#> [220] "Label"                  "Label<-"                "Lambda"                
#> [223] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [226] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [229] "LinScale"               "LineToUser"             "LogSt"                 
#> [232] "LogStInv"               "Logit"                  "LogitInv"              
#> [235] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [238] "MAD"                    "MADCI"                  "MAE"                   
#> [241] "MAPE"                   "MHChisqTest"            "MSE"                   
#> [244] "Mar"                    "Margins"                "MaxDigits"             
#> [247] "Mean"                   "MeanAD"                 "MeanCI"                
#> [250] "MeanCIn"                "MeanDiffCI"             "MeanSE"                
#> [253] "Median"                 "MedianCI"               "Mgp"                   
#> [256] "Midx"                   "Minute"                 "MixColor"              
#> [259] "ModSummary"             "Mode"                   "Month"                 
#> [262] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [265] "MultMerge"              "MultinomCI"             "MutInf"                
#> [268] "N"                      "NAIf"                   "NAIfBlank"             
#> [271] "NAIfZero"               "NMAE"                   "NMSE"                  
#> [274] "NPV"                    "NPVFixBond"             "Ndec"                  
#> [277] "NemenyiTest"            "Now"                    "OPR"                   
#> [280] "ORToRelRisk"            "OctToDec"               "OddsRatio"             
#> [283] "OrderMixed"             "Outlier"                "Overlap"               
#> [286] "PDFManual"              "PMT"                    "PPMT"                  
#> [289] "PageTest"               "PairApply"              "Pal"                   
#> [292] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [295] "PasswordDlg"            "PearsonTest"            "PercTable"             
#> [298] "PercentRank"            "Permn"                  "Phi"                   
#> [301] "Phrase"                 "PlotACF"                "PlotArea"              
#> [304] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [307] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [310] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [313] "PlotDot"                "PlotDotCI"              "PlotECDF"              
#> [316] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [319] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [322] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [325] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [328] "PlotPolar"              "PlotProbDist"           "PlotPyramid"           
#> [331] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [334] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [337] "PoissonCI"              "PolToCart"              "PolarGrid"             
#> [340] "PostHocTest"            "PpAddSlide"             "PpPlot"                
#> [343] "PpText"                 "Prec"                   "Primes"                
#> [346] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [349] "Quantile"               "QuantileCI"             "Quarter"               
#> [352] "Quot"                   "RBAL"                   "RMSE"                  
#> [355] "RSessionAlive"          "RTempdirAlive"          "RadToDeg"              
#> [358] "Range"                  "Rank"                   "Recode"                
#> [361] "Recycle"                "RelRisk"                "Rename"                
#> [364] "Rev"                    "RevCode"                "RgbToCmy"              
#> [367] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [370] "RndPairs"               "RndWord"                "RobScale"              
#> [373] "RomanToInt"             "Rosenbluth"             "Rotate"                
#> [376] "RoundTo"                "RunsTest"               "SD"                    
#> [379] "SLN"                    "SMAPE"                  "SYD"                   
#> [382] "Sample"                 "SampleTwins"            "SaveAs"                
#> [385] "ScheffeTest"            "SecToHms"               "Second"                
#> [388] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [391] "SetAttr"                "SetNames"               "Shade"                 
#> [394] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [397] "SignTest"               "Skew"                   "Small"                 
#> [400] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [403] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [406] "Spec"                   "SphToCart"              "SplitAt"               
#> [409] "SplitPath"              "SplitToCol"             "SplitToDummy"          
#> [412] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [415] "Str"                    "StrAbbr"                "StrAlign"              
#> [418] "StrCap"                 "StrChop"                "StrCountW"             
#> [421] "StrDist"                "StrExtract"             "StrExtractBetween"     
#> [424] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [427] "StrPos"                 "StrRev"                 "StrRight"              
#> [430] "StrSpell"               "StrSplit"               "StrTrim"               
#> [433] "StrTrunc"               "StrVal"                 "Strata"                
#> [436] "StripAttr"              "StuartMaxwellTest"      "StuartTauC"            
#> [439] "SysInfo"                "TMod"                   "TOne"                  
#> [442] "TTestA"                 "TextContrastColor"      "TextToTable"           
#> [445] "TheilU"                 "Timezone"               "TitleRect"             
#> [448] "ToLong"                 "ToWide"                 "ToWrd"                 
#> [451] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [454] "Today"                  "Trim"                   "TschuprowT"            
#> [457] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [460] "UnirootAll"             "Unit"                   "Unit<-"                
#> [463] "Untable"                "Unwhich"                "VIF"                   
#> [466] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [469] "VarTest"                "VecRot"                 "VecShift"              
#> [472] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [475] "Weekday"                "Winsorize"              "WithOptions"           
#> [478] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [481] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [484] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [487] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [490] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [493] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [496] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [499] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [502] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [505] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [508] "XLView"                 "YTM"                    "Year"                  
#> [511] "YearDay"                "YearDays"               "YearMonth"             
#> [514] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [517] "ZTest"                  "ZeroIfNA"               "Zodiac"                
#> [520] "aovlDetails"            "aovlErrorTerms"         "as.CDateFmt"           
#> [523] "as.fmt"                 "as.matrix.xtabs"        "axTicks.Date"          
#> [526] "axTicks.POSIXct"        "createCOMReference"     "dBenf"                 
#> [529] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [532] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [535] "dNegWeibull"            "dOrder"                 "dRevGumbel"            
#> [538] "dRevWeibull"            "identify.formula"       "lines.Lc"              
#> [541] "lines.SmoothSpline"     "lines.lm"               "lines.loess"           
#> [544] "lines.smooth.spline"    "pBenf"                  "pExtrVal"              
#> [547] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [550] "pGompertz"              "pGumbel"                "pNegWeibull"           
#> [553] "pOrder"                 "pRevGumbel"             "pRevWeibull"           
#> [556] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [559] "plot.PostHocTest"       "plot.bagplot"           "plot.palette"          
#> [562] "power.chisq.test"       "print.Assocs"           "print.Conf"            
#> [565] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [568] "print.HoeffD"           "print.ICC"              "print.PercTable"       
#> [571] "print.PostHocTest"      "print.mtest"            "qBenf"                 
#> [574] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [577] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [580] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [583] "qRevWeibull"            "rBenf"                  "rExtrVal"              
#> [586] "rFrechet"               "rGenExtrVal"            "rGenPareto"            
#> [589] "rGompertz"              "rGumbel"                "rNegWeibull"           
#> [592] "rOrder"                 "rRevGumbel"             "rRevWeibull"           
#> [595] "rSum21"                 "reorder.factor"         "split.formula"