List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BinomCI"                "BinomDiffCI"           
#>  [43] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [46] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [49] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [52] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [55] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CCC"                   
#>  [61] "CentralValue"           "CharToAsc"              "Clockwise"             
#>  [64] "Closest"                "CmToPts"                "Coalesce"              
#>  [67] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [70] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [73] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [76] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [79] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [82] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [85] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [88] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [91] "ConvUnit"               "Cor"                    "CorCI"                 
#>  [94] "CorPart"                "CorPolychor"            "CountCompCases"        
#>  [97] "CountWorkDays"          "CourseData"             "Cov"                   
#> [100] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CutQ"                   "Day"                   
#> [109] "Day<-"                  "DB"                     "dBenf"                 
#> [112] "DecToBin"               "DecToHex"               "DecToOct"              
#> [115] "DegToRad"               "DenseRank"              "Desc"                  
#> [118] "DescToolsOptions"       "dExtrVal"               "dFrechet"              
#> [121] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [124] "dGumbel"                "DiffDays360"            "DigitSum"              
#> [127] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [130] "dNegWeibull"            "DoBy"                   "DoCall"                
#> [133] "dOrder"                 "Dot"                    "DrawArc"               
#> [136] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [139] "DrawEllipse"            "DrawRegPolygon"         "dRevGumbel"            
#> [142] "dRevWeibull"            "Dummy"                  "DunnettTest"           
#> [145] "DunnTest"               "DurbinWatsonTest"       "Entropy"               
#> [148] "Eps"                    "ErrBars"                "EtaSq"                 
#> [151] "ExpFreq"                "Factorize"              "FctArgs"               
#> [154] "Fibonacci"              "FindColor"              "FindCorr"              
#> [157] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [160] "FixToTable"             "Flags"                  "Fmt"                   
#> [163] "Format"                 "Frac"                   "Freq"                  
#> [166] "Freq2D"                 "GCD"                    "GeomSn"                
#> [169] "GeomTrans"              "GetCurrPP"              "GetCurrWrd"            
#> [172] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [175] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [178] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [181] "Gsd"                    "GTest"                  "Herfindahl"            
#> [184] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [187] "HighLow"                "Hmean"                  "HmsToSec"              
#> [190] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [193] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [196] "HunterGaston"           "ICC"                    "identify.formula"      
#> [199] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [202] "InDots"                 "Interval"               "IPMT"                  
#> [205] "IQRw"                   "IRR"                    "IsDate"                
#> [208] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [211] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [214] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [217] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [220] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [223] "KendallW"               "Keywords"               "KrippAlpha"            
#> [226] "Kurt"                   "Label"                  "Label<-"               
#> [229] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [232] "Lc"                     "LCM"                    "LehmacherTest"         
#> [235] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [238] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [241] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [244] "LOCF"                   "LOF"                    "Logit"                 
#> [247] "LogitInv"               "LogSt"                  "LogStInv"              
#> [250] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [253] "MAD"                    "MAE"                    "MAPE"                  
#> [256] "Mar"                    "Margins"                "MaxDigits"             
#> [259] "Mean"                   "MeanAD"                 "MeanCI"                
#> [262] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [265] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [268] "Midx"                   "Minute"                 "MixColor"              
#> [271] "Mode"                   "ModSummary"             "Month"                 
#> [274] "MosesTest"              "MoveAvg"                "MSE"                   
#> [277] "MultinomCI"             "MultMerge"              "MutInf"                
#> [280] "N"                      "NAIfBlank"              "NAIfZero"              
#> [283] "Ndec"                   "NemenyiTest"            "NMAE"                  
#> [286] "NMSE"                   "Now"                    "NPV"                   
#> [289] "NPVFixBond"             "OctToDec"               "OddsRatio"             
#> [292] "OPR"                    "OrderMixed"             "ORToRelRisk"           
#> [295] "Outlier"                "Overlap"                "PageTest"              
#> [298] "PairApply"              "Pal"                    "ParseFormula"          
#> [301] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [304] "pBenf"                  "PDFManual"              "PearsonTest"           
#> [307] "PercentRank"            "PercTable"              "Permn"                 
#> [310] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [313] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [316] "Phi"                    "Phrase"                 "plot.bagplot"          
#> [319] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [322] "plot.palette"           "plot.PostHocTest"       "PlotACF"               
#> [325] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [328] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [331] "PlotCashFlow"           "PlotCirc"               "PlotCorr"              
#> [334] "PlotDot"                "PlotECDF"               "PlotFaces"             
#> [337] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [340] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [343] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [346] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [349] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [352] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [355] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [358] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [361] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [364] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [367] "PpText"                 "Prec"                   "pRevGumbel"            
#> [370] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [373] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [376] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [379] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [382] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [385] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [388] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [391] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [394] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [397] "Quarter"                "Quot"                   "RadToDeg"              
#> [400] "Range"                  "RBAL"                   "rBenf"                 
#> [403] "Recode"                 "Recycle"                "RelRisk"               
#> [406] "Rename"                 "reorder.factor"         "Rev"                   
#> [409] "rExtrVal"               "rFrechet"               "RgbToCol"              
#> [412] "RgbToHex"               "RgbToLong"              "rGenExtrVal"           
#> [415] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [418] "RMSE"                   "RndPairs"               "RndWord"               
#> [421] "rNegWeibull"            "RobScale"               "RomanToInt"            
#> [424] "rOrder"                 "Rosenbluth"             "Rotate"                
#> [427] "RoundTo"                "rRevGumbel"             "rRevWeibull"           
#> [430] "RSessionAlive"          "RTempdirAlive"          "RunsTest"              
#> [433] "Sample"                 "SampleTwins"            "SaveAs"                
#> [436] "ScheffeTest"            "SD"                     "Second"                
#> [439] "SecToHms"               "SendOutlookMail"        "Sens"                  
#> [442] "SetAlpha"               "SetNames"               "Shade"                 
#> [445] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [448] "SignTest"               "Skew"                   "SLN"                   
#> [451] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [454] "Some"                   "SomersDelta"            "Sort"                  
#> [457] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [460] "SphToCart"              "split.formula"          "SplitAt"               
#> [463] "SplitPath"              "SpreadOut"              "Stamp"                 
#> [466] "StdCoef"                "Str"                    "StrAbbr"               
#> [469] "StrAlign"               "Strata"                 "StrCap"                
#> [472] "StrChop"                "StrCountW"              "StrDist"               
#> [475] "StrExtract"             "StrIsNumeric"           "StrLeft"               
#> [478] "StrPad"                 "StrPos"                 "StrRev"                
#> [481] "StrRight"               "StrSpell"               "StrTrim"               
#> [484] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [487] "StuartTauC"             "SYD"                    "SysInfo"               
#> [490] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [493] "Timezone"               "TitleRect"              "TMod"                  
#> [496] "Today"                  "ToLong"                 "TOne"                  
#> [499] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [502] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [505] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [508] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [511] "Unit"                   "Unit<-"                 "Untable"               
#> [514] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [517] "VarCI"                  "VarTest"                "VecRot"                
#> [520] "VecShift"               "VIF"                    "Vigenere"              
#> [523] "VonNeumannTest"         "Week"                   "Weekday"               
#> [526] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [529] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [532] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [535] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [538] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [541] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [544] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [547] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [550] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [553] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [556] "XLSaveAs"               "XLView"                 "Year"                  
#> [559] "YearDay"                "YearMonth"              "YTM"                   
#> [562] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [565] "ZeroIfNA"               "Zodiac"                 "ZTest"