List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BinomCI"                "BinomDiffCI"           
#>  [43] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [46] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [49] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [52] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [55] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CCC"                   
#>  [61] "CharToAsc"              "Clockwise"              "Closest"               
#>  [64] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [67] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [70] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [73] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [76] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [79] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [82] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [85] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [88] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [91] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [94] "Cor"                    "CorCI"                  "CorPart"               
#>  [97] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [100] "CourseData"             "Cov"                    "CramerV"               
#> [103] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [106] "Cross"                  "CrossN"                 "Cstat"                 
#> [109] "CutQ"                   "Day"                    "Day<-"                 
#> [112] "DB"                     "dBenf"                  "DecToBin"              
#> [115] "DecToHex"               "DecToOct"               "DegToRad"              
#> [118] "DenseRank"              "Desc"                   "DescToolsOptions"      
#> [121] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [124] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [127] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [130] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [133] "DoBy"                   "DoCall"                 "dOrder"                
#> [136] "Dot"                    "DrawArc"                "DrawBand"              
#> [139] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [142] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [145] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [148] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [151] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [154] "Factorize"              "FctArgs"                "Fibonacci"             
#> [157] "FindColor"              "FindCorr"               "FindRProfile"          
#> [160] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [163] "Flags"                  "Fmt"                    "Format"                
#> [166] "Frac"                   "Freq"                   "Freq2D"                
#> [169] "GCD"                    "GeomSn"                 "GeomTrans"             
#> [172] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [175] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [178] "Gini"                   "GiniSimpson"            "Gmean"                 
#> [181] "GoodmanKruskalGamma"    "GoodmanKruskalTau"      "Gsd"                   
#> [184] "GTest"                  "Herfindahl"             "HexToCol"              
#> [187] "HexToDec"               "HexToRgb"               "HighLow"               
#> [190] "Hmean"                  "HmsToSec"               "HodgesLehmann"         
#> [193] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [196] "Hour"                   "HuberM"                 "HunterGaston"          
#> [199] "ICC"                    "identify.formula"       "IdentifyA"             
#> [202] "Impute"                 "ImputeKnn"              "InDots"                
#> [205] "Interval"               "IPMT"                   "IQRw"                  
#> [208] "IRR"                    "IsDate"                 "IsDichotomous"         
#> [211] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [214] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [217] "IsWeekend"              "IsWhole"                "IsZero"                
#> [220] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [223] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [226] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [229] "Label"                  "Label<-"                "Lambda"                
#> [232] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [235] "LCM"                    "LehmacherTest"          "LeveneTest"            
#> [238] "LillieTest"             "lines.Lc"               "lines.lm"              
#> [241] "lines.loess"            "lines.smooth.spline"    "lines.SmoothSpline"    
#> [244] "LineToUser"             "LinScale"               "LOCF"                  
#> [247] "LOF"                    "Logit"                  "LogitInv"              
#> [250] "LogSt"                  "LogStInv"               "LongToRgb"             
#> [253] "LsFct"                  "LsObj"                  "MAD"                   
#> [256] "MADCI"                  "MAE"                    "MAPE"                  
#> [259] "Mar"                    "Margins"                "MaxDigits"             
#> [262] "Mean"                   "MeanAD"                 "MeanCI"                
#> [265] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [268] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [271] "Midx"                   "Minute"                 "MixColor"              
#> [274] "Mode"                   "ModSummary"             "Month"                 
#> [277] "MosesTest"              "MoveAvg"                "MSE"                   
#> [280] "MultinomCI"             "MultMerge"              "MutInf"                
#> [283] "N"                      "NAIfBlank"              "NAIfZero"              
#> [286] "Ndec"                   "NemenyiTest"            "NMAE"                  
#> [289] "NMSE"                   "Now"                    "NPV"                   
#> [292] "NPVFixBond"             "OctToDec"               "OddsRatio"             
#> [295] "OPR"                    "OrderMixed"             "ORToRelRisk"           
#> [298] "Outlier"                "Overlap"                "PageTest"              
#> [301] "PairApply"              "Pal"                    "ParseFormula"          
#> [304] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [307] "pBenf"                  "PDFManual"              "PearsonTest"           
#> [310] "PercentRank"            "PercTable"              "Permn"                 
#> [313] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [316] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [319] "Phi"                    "Phrase"                 "plot.bagplot"          
#> [322] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [325] "plot.palette"           "plot.PostHocTest"       "PlotACF"               
#> [328] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [331] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [334] "PlotCashFlow"           "PlotCirc"               "PlotConDens"           
#> [337] "PlotCorr"               "PlotDot"                "PlotECDF"              
#> [340] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [343] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [346] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [349] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [352] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [355] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [358] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [361] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [364] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [367] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [370] "PpPlot"                 "PpText"                 "Prec"                  
#> [373] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [376] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [379] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [382] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [385] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [388] "PtsToCm"                "qBenf"                  "qExtrVal"              
#> [391] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [394] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [397] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [400] "Quantile"               "Quarter"                "Quot"                  
#> [403] "RadToDeg"               "Range"                  "RBAL"                  
#> [406] "rBenf"                  "Recode"                 "Recycle"               
#> [409] "RelRisk"                "Rename"                 "reorder.factor"        
#> [412] "Rev"                    "RevCode"                "rExtrVal"              
#> [415] "rFrechet"               "RgbToCmy"               "RgbToCol"              
#> [418] "RgbToHex"               "RgbToLong"              "rGenExtrVal"           
#> [421] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [424] "RMSE"                   "RndPairs"               "RndWord"               
#> [427] "rNegWeibull"            "RobScale"               "RomanToInt"            
#> [430] "rOrder"                 "Rosenbluth"             "Rotate"                
#> [433] "RoundTo"                "rRevGumbel"             "rRevWeibull"           
#> [436] "RSessionAlive"          "RTempdirAlive"          "RunsTest"              
#> [439] "Sample"                 "SampleTwins"            "SaveAs"                
#> [442] "ScheffeTest"            "SD"                     "Second"                
#> [445] "SecToHms"               "SendOutlookMail"        "Sens"                  
#> [448] "SetAlpha"               "SetNames"               "Shade"                 
#> [451] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [454] "SignTest"               "Skew"                   "SLN"                   
#> [457] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [460] "Some"                   "SomersDelta"            "Sort"                  
#> [463] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [466] "SphToCart"              "split.formula"          "SplitAt"               
#> [469] "SplitPath"              "SplitToCol"             "SpreadOut"             
#> [472] "Stamp"                  "StdCoef"                "Str"                   
#> [475] "StrAbbr"                "StrAlign"               "Strata"                
#> [478] "StrCap"                 "StrChop"                "StrCountW"             
#> [481] "StrDist"                "StrExtract"             "StripAttr"             
#> [484] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [487] "StrPos"                 "StrRev"                 "StrRight"              
#> [490] "StrSpell"               "StrSplit"               "StrTrim"               
#> [493] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [496] "StuartTauC"             "SYD"                    "SysInfo"               
#> [499] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [502] "Timezone"               "TitleRect"              "TMod"                  
#> [505] "Today"                  "ToLong"                 "TOne"                  
#> [508] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [511] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [514] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [517] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [520] "Unit"                   "Unit<-"                 "Untable"               
#> [523] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [526] "VarCI"                  "VarTest"                "VecRot"                
#> [529] "VecShift"               "VIF"                    "Vigenere"              
#> [532] "VonNeumannTest"         "Week"                   "Weekday"               
#> [535] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [538] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [541] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [544] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [547] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [550] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [553] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [556] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [559] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [562] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [565] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [568] "Year"                   "YearDay"                "YearMonth"             
#> [571] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [574] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [577] "ZTest"