List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%::%"                   "%:%"                    "%()%"                  
#>   [4] "%(]%"                   "%)(%"                   "%)[%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "Abind"                  "Abstract"              
#>  [19] "AddMonths"              "AddMonthsYM"            "Agree"                 
#>  [22] "AllDuplicated"          "AllIdentical"           "AndersonDarlingTest"   
#>  [25] "aovlDetails"            "aovlErrorTerms"         "Append"                
#>  [28] "Arrow"                  "as.CDateFmt"            "as.fmt"                
#>  [31] "as.matrix.xtabs"        "AscToChar"              "Asp"                   
#>  [34] "Assocs"                 "Atkinson"               "AUC"                   
#>  [37] "AxisBreak"              "axTicks.Date"           "axTicks.POSIXct"       
#>  [40] "BarnardTest"            "BartelsRankTest"        "BarText"               
#>  [43] "BhapkarTest"            "BinomCI"                "BinomCIn"              
#>  [46] "BinomDiffCI"            "BinomRatioCI"           "BinToDec"              
#>  [49] "BinTree"                "BlankIfNA"              "BootCI"                
#>  [52] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [55] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [58] "BrierScore"             "BubbleLegend"           "Canvas"                
#>  [61] "CartToPol"              "CartToSph"              "CatTable"              
#>  [64] "CCC"                    "CharToAsc"              "Clockwise"             
#>  [67] "Closest"                "CmToPts"                "CmykToCmy"             
#>  [70] "CmykToRgb"              "CmyToCmyk"              "Coalesce"              
#>  [73] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [76] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [79] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [82] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [85] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [88] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [91] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [94] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [97] "ConvUnit"               "Cor"                    "CorCI"                 
#> [100] "CorPart"                "CorPolychor"            "CountCompCases"        
#> [103] "CountWorkDays"          "CourseData"             "Cov"                   
#> [106] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [109] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [112] "Cstat"                  "CutQ"                   "Day"                   
#> [115] "Day<-"                  "DB"                     "dBenf"                 
#> [118] "DecToBin"               "DecToHex"               "DecToOct"              
#> [121] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [271] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [274] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [277] "Midx"                   "Minute"                 "MixColor"              
#> [280] "Mode"                   "ModSummary"             "Month"                 
#> [283] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [286] "MSE"                    "MultinomCI"             "MultMerge"             
#> [289] "MutInf"                 "N"                      "NAIf"                  
#> [292] "NAIfBlank"              "NAIfZero"               "Ndec"                  
#> [295] "NemenyiTest"            "NMAE"                   "NMSE"                  
#> [298] "Now"                    "NPV"                    "NPVFixBond"            
#> [301] "OctToDec"               "OddsRatio"              "OPR"                   
#> [304] "OrderMixed"             "ORToRelRisk"            "Outlier"               
#> [307] "Overlap"                "PageTest"               "PairApply"             
#> [310] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [313] "PartialSD"              "PasswordDlg"            "pBenf"                 
#> [316] "PDFManual"              "PearsonTest"            "PercentRank"           
#> [319] "PercTable"              "Permn"                  "pExtrVal"              
#> [322] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [325] "pGompertz"              "pGumbel"                "Phi"                   
#> [328] "Phrase"                 "plot.bagplot"           "plot.Conf"             
#> [331] "plot.Lc"                "plot.Lclist"            "plot.palette"          
#> [334] "plot.PostHocTest"       "PlotACF"                "PlotArea"              
#> [337] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [340] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [343] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [346] "PlotDot"                "PlotDotCI"              "PlotECDF"              
#> [349] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [352] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [355] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [358] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [361] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [364] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [367] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [370] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [373] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [376] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [379] "PpPlot"                 "PpText"                 "Prec"                  
#> [382] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [385] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [388] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [391] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [394] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [397] "PtsToCm"                "qBenf"                  "qExtrVal"              
#> [400] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [403] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [406] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [409] "Quantile"               "QuantileCI"             "Quarter"               
#> [412] "Quot"                   "RadToDeg"               "Range"                 
#> [415] "Rank"                   "RBAL"                   "rBenf"                 
#> [418] "Recode"                 "Recycle"                "RelRisk"               
#> [421] "Rename"                 "reorder.factor"         "Rev"                   
#> [424] "RevCode"                "rExtrVal"               "rFrechet"              
#> [427] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [430] "RgbToLong"              "rGenExtrVal"            "rGenPareto"            
#> [433] "rGompertz"              "rGumbel"                "RMSE"                  
#> [436] "RndPairs"               "RndWord"                "rNegWeibull"           
#> [439] "RobScale"               "RomanToInt"             "rOrder"                
#> [442] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [445] "rRevGumbel"             "rRevWeibull"            "RSessionAlive"         
#> [448] "RTempdirAlive"          "RunsTest"               "Sample"                
#> [451] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [454] "SD"                     "Second"                 "SecToHms"              
#> [457] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [460] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [463] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [466] "Skew"                   "SLN"                    "Small"                 
#> [469] "SMAPE"                  "SmoothSpline"           "Some"                  
#> [472] "SomersDelta"            "Sort"                   "SortMixed"             
#> [475] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [478] "split.formula"          "SplitAt"                "SplitPath"             
#> [481] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [484] "StdCoef"                "Str"                    "StrAbbr"               
#> [487] "StrAlign"               "Strata"                 "StrCap"                
#> [490] "StrChop"                "StrCountW"              "StrDist"               
#> [493] "StrExtract"             "StrExtractBetween"      "StripAttr"             
#> [496] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [499] "StrPos"                 "StrRev"                 "StrRight"              
#> [502] "StrSpell"               "StrSplit"               "StrTrim"               
#> [505] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [508] "StuartTauC"             "SYD"                    "SysInfo"               
#> [511] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [514] "Timezone"               "TitleRect"              "TMod"                  
#> [517] "Today"                  "ToLong"                 "TOne"                  
#> [520] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [523] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [526] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [529] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [532] "Unit"                   "Unit<-"                 "Untable"               
#> [535] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [538] "VarCI"                  "VarTest"                "VecRot"                
#> [541] "VecShift"               "VIF"                    "Vigenere"              
#> [544] "VonNeumannTest"         "Week"                   "Weekday"               
#> [547] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [550] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [553] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [556] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [559] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [562] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [565] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [568] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [571] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [574] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [577] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [580] "Year"                   "YearDay"                "YearDays"              
#> [583] "YearMonth"              "YTM"                    "YuenTTest"             
#> [586] "YuleQ"                  "YuleY"                  "ZeroIfNA"              
#> [589] "Zodiac"                 "ZTest"