List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BinomCI"                "BinomDiffCI"           
#>  [43] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [46] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [49] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [52] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [55] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CCC"                   
#>  [61] "CharToAsc"              "Clockwise"              "Closest"               
#>  [64] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [67] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [70] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [73] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [76] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [79] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [82] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [85] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [88] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [91] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [94] "Cor"                    "CorCI"                  "CorPart"               
#>  [97] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [100] "CourseData"             "Cov"                    "CramerV"               
#> [103] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [106] "Cross"                  "CrossN"                 "Cstat"                 
#> [109] "CutQ"                   "Day"                    "Day<-"                 
#> [112] "DB"                     "dBenf"                  "DecToBin"              
#> [115] "DecToHex"               "DecToOct"               "DegToRad"              
#> [118] "DenseRank"              "Desc"                   "DescToolsOptions"      
#> [121] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [124] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [127] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [130] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [133] "DoBy"                   "DoCall"                 "dOrder"                
#> [136] "Dot"                    "DrawArc"                "DrawBand"              
#> [139] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [142] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [145] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [148] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [151] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [154] "Factorize"              "FctArgs"                "Fibonacci"             
#> [157] "FindColor"              "FindCorr"               "FindRProfile"          
#> [160] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [163] "Flags"                  "Fmt"                    "Format"                
#> [166] "Frac"                   "Freq"                   "Freq2D"                
#> [169] "GCD"                    "GeomSn"                 "GeomTrans"             
#> [172] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [175] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [178] "Gini"                   "GiniSimpson"            "Gmean"                 
#> [181] "GoodmanKruskalGamma"    "GoodmanKruskalTau"      "Gsd"                   
#> [184] "GTest"                  "Herfindahl"             "HexToCol"              
#> [187] "HexToDec"               "HexToRgb"               "HighLow"               
#> [190] "Hmean"                  "HmsToSec"               "HodgesLehmann"         
#> [193] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [196] "Hour"                   "HuberM"                 "HunterGaston"          
#> [199] "ICC"                    "identify.formula"       "IdentifyA"             
#> [202] "Impute"                 "ImputeKnn"              "InDots"                
#> [205] "Interval"               "IPMT"                   "IQRw"                  
#> [208] "IRR"                    "IsDate"                 "IsDichotomous"         
#> [211] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [214] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [217] "IsWeekend"              "IsWhole"                "IsZero"                
#> [220] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [223] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [226] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [229] "Label"                  "Label<-"                "Lambda"                
#> [232] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [235] "LCM"                    "LehmacherTest"          "LeveneTest"            
#> [238] "LillieTest"             "lines.Lc"               "lines.lm"              
#> [241] "lines.loess"            "lines.smooth.spline"    "lines.SmoothSpline"    
#> [244] "LineToUser"             "LinScale"               "LOCF"                  
#> [247] "LOF"                    "Logit"                  "LogitInv"              
#> [250] "LogSt"                  "LogStInv"               "LongToRgb"             
#> [253] "LsFct"                  "LsObj"                  "MAD"                   
#> [256] "MAE"                    "MAPE"                   "Mar"                   
#> [259] "Margins"                "MaxDigits"              "Mean"                  
#> [262] "MeanAD"                 "MeanCI"                 "MeanDiffCI"            
#> [265] "MeanSE"                 "Median"                 "MedianCI"              
#> [268] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [271] "Minute"                 "MixColor"               "Mode"                  
#> [274] "ModSummary"             "Month"                  "MosesTest"             
#> [277] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [280] "MultMerge"              "MutInf"                 "N"                     
#> [283] "NAIfBlank"              "NAIfZero"               "Ndec"                  
#> [286] "NemenyiTest"            "NMAE"                   "NMSE"                  
#> [289] "Now"                    "NPV"                    "NPVFixBond"            
#> [292] "OctToDec"               "OddsRatio"              "OPR"                   
#> [295] "OrderMixed"             "ORToRelRisk"            "Outlier"               
#> [298] "Overlap"                "PageTest"               "PairApply"             
#> [301] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [304] "PartialSD"              "PasswordDlg"            "pBenf"                 
#> [307] "PDFManual"              "PearsonTest"            "PercentRank"           
#> [310] "PercTable"              "Permn"                  "pExtrVal"              
#> [313] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [316] "pGompertz"              "pGumbel"                "Phi"                   
#> [319] "Phrase"                 "plot.bagplot"           "plot.Conf"             
#> [322] "plot.Lc"                "plot.Lclist"            "plot.palette"          
#> [325] "plot.PostHocTest"       "PlotACF"                "PlotArea"              
#> [328] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [331] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [334] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [337] "PlotDot"                "PlotECDF"               "PlotFaces"             
#> [340] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [343] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [346] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [349] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [352] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [355] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [358] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [361] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [364] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [367] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [370] "PpText"                 "Prec"                   "pRevGumbel"            
#> [373] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [376] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [379] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [382] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [385] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [388] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [391] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [394] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [397] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [400] "Quarter"                "Quot"                   "RadToDeg"              
#> [403] "Range"                  "RBAL"                   "rBenf"                 
#> [406] "Recode"                 "Recycle"                "RelRisk"               
#> [409] "Rename"                 "reorder.factor"         "Rev"                   
#> [412] "rExtrVal"               "rFrechet"               "RgbToCmy"              
#> [415] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [418] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [421] "rGumbel"                "RMSE"                   "RndPairs"              
#> [424] "RndWord"                "rNegWeibull"            "RobScale"              
#> [427] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [430] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [433] "rRevWeibull"            "RSessionAlive"          "RTempdirAlive"         
#> [436] "RunsTest"               "Sample"                 "SampleTwins"           
#> [439] "SaveAs"                 "ScheffeTest"            "SD"                    
#> [442] "Second"                 "SecToHms"               "SendOutlookMail"       
#> [445] "Sens"                   "SetAlpha"               "SetNames"              
#> [448] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [451] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [454] "SLN"                    "Small"                  "SMAPE"                 
#> [457] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [460] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [463] "Spec"                   "SphToCart"              "split.formula"         
#> [466] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [469] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [472] "Str"                    "StrAbbr"                "StrAlign"              
#> [475] "Strata"                 "StrCap"                 "StrChop"               
#> [478] "StrCountW"              "StrDist"                "StrExtract"            
#> [481] "StripAttr"              "StrIsNumeric"           "StrLeft"               
#> [484] "StrPad"                 "StrPos"                 "StrRev"                
#> [487] "StrRight"               "StrSpell"               "StrSplit"              
#> [490] "StrTrim"                "StrTrunc"               "StrVal"                
#> [493] "StuartMaxwellTest"      "StuartTauC"             "SYD"                   
#> [496] "SysInfo"                "TextContrastColor"      "TextToTable"           
#> [499] "TheilU"                 "Timezone"               "TitleRect"             
#> [502] "TMod"                   "Today"                  "ToLong"                
#> [505] "TOne"                   "ToWide"                 "ToWrd"                 
#> [508] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [511] "Trim"                   "TschuprowT"             "TTestA"                
#> [514] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [517] "UnirootAll"             "Unit"                   "Unit<-"                
#> [520] "Untable"                "Unwhich"                "VanWaerdenTest"        
#> [523] "Var"                    "VarCI"                  "VarTest"               
#> [526] "VecRot"                 "VecShift"               "VIF"                   
#> [529] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [532] "Weekday"                "Winsorize"              "WithOptions"           
#> [535] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [538] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [541] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [544] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [547] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [550] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [553] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [556] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [559] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [562] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [565] "XLView"                 "Year"                   "YearDay"               
#> [568] "YearMonth"              "YTM"                    "YuenTTest"             
#> [571] "YuleQ"                  "YuleY"                  "ZeroIfNA"              
#> [574] "Zodiac"                 "ZTest"