List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [25] "AllIdentical"           "AndersonDarlingTest"    "Append"                
#>  [28] "Arrow"                  "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AxisBreak"             
#>  [34] "BarText"                "BarnardTest"            "BartelsRankTest"       
#>  [37] "Bg"                     "BhapkarTest"            "BinToDec"              
#>  [40] "BinTree"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "CCC"                    "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "CmToPts"               
#>  [64] "CmyToCmyk"              "CmykToCmy"              "CmykToRgb"             
#>  [67] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [70] "CoefVar"                "CohenD"                 "CohenKappa"            
#>  [73] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [76] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [79] "CollapseTable"          "ColorLegend"            "ColumnWrap"            
#>  [82] "CombN"                  "CombPairs"              "CombSet"               
#>  [85] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [88] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [91] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [94] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [97] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [100] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CutQ"                   "DB"                    
#> [109] "Day"                    "Day<-"                  "DecToBin"              
#> [112] "DecToHex"               "DecToOct"               "DegToRad"              
#> [115] "Desc"                   "DescToolsOptions"       "DiffDays360"           
#> [118] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [121] "Divisors"               "DoBy"                   "DoCall"                
#> [124] "Dot"                    "DrawArc"                "DrawBand"              
#> [127] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [130] "DrawRegPolygon"         "Dummy"                  "DunnTest"              
#> [133] "DunnettTest"            "DurbinWatsonTest"       "Entropy"               
#> [136] "Eps"                    "ErrBars"                "EtaSq"                 
#> [139] "ExpFreq"                "Factorize"              "FctArgs"               
#> [142] "Fibonacci"              "FindColor"              "FindCorr"              
#> [145] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [148] "FixToTable"             "Flags"                  "Fmt"                   
#> [151] "Format"                 "Frac"                   "Freq"                  
#> [154] "Freq2D"                 "GCD"                    "GTest"                 
#> [157] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [160] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [163] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [166] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [169] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [172] "Gsd"                    "Herfindahl"             "HexToCol"              
#> [175] "HexToDec"               "HexToRgb"               "HighLow"               
#> [178] "Hmean"                  "HmsToSec"               "HodgesLehmann"         
#> [181] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [184] "Hour"                   "HuberM"                 "HunterGaston"          
#> [187] "ICC"                    "IPMT"                   "IQRw"                  
#> [190] "IRR"                    "IdentifyA"              "Impute"                
#> [193] "ImputeKnn"              "InDots"                 "Interval"              
#> [196] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [199] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [202] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [205] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [208] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [211] "KendallTauB"            "KendallW"               "Keywords"              
#> [214] "KrippAlpha"             "Kurt"                   "LCM"                   
#> [217] "LOCF"                   "LOF"                    "Label"                 
#> [220] "Label<-"                "Lambda"                 "Large"                 
#> [223] "LastDayOfMonth"         "Lc"                     "LehmacherTest"         
#> [226] "LeveneTest"             "LillieTest"             "LinScale"              
#> [229] "LineToUser"             "LogSt"                  "LogStInv"              
#> [232] "Logit"                  "LogitInv"               "LongToRgb"             
#> [235] "LsFct"                  "LsObj"                  "MAD"                   
#> [238] "MADCI"                  "MAE"                    "MAPE"                  
#> [241] "MHChisqTest"            "MSE"                    "Mar"                   
#> [244] "Margins"                "MaxDigits"              "Mean"                  
#> [247] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [250] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [253] "MedianCI"               "Mgp"                    "Midx"                  
#> [256] "Minute"                 "MixColor"               "ModSummary"            
#> [259] "Mode"                   "Month"                  "MonthDays"             
#> [262] "MosesTest"              "MoveAvg"                "MultMerge"             
#> [265] "MultinomCI"             "MutInf"                 "N"                     
#> [268] "NAIf"                   "NAIfBlank"              "NAIfZero"              
#> [271] "NMAE"                   "NMSE"                   "NPV"                   
#> [274] "NPVFixBond"             "Ndec"                   "NemenyiTest"           
#> [277] "Now"                    "OPR"                    "ORToRelRisk"           
#> [280] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [283] "Outlier"                "Overlap"                "PDFManual"             
#> [286] "PMT"                    "PPMT"                   "PageTest"              
#> [289] "PairApply"              "Pal"                    "ParseFormula"          
#> [292] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [295] "PearsonTest"            "PercTable"              "PercentRank"           
#> [298] "Permn"                  "Phi"                    "Phrase"                
#> [301] "PlotACF"                "PlotArea"               "PlotBag"               
#> [304] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [307] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [310] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [313] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [316] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [319] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [322] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [325] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [328] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [331] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [334] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [337] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [340] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [343] "Prec"                   "Primes"                 "PseudoR2"              
#> [346] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [349] "QuantileCI"             "Quarter"                "Quot"                  
#> [352] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [355] "RTempdirAlive"          "RadToDeg"               "Range"                 
#> [358] "Rank"                   "Recode"                 "Recycle"               
#> [361] "RelRisk"                "Rename"                 "Rev"                   
#> [364] "RevCode"                "RgbToCmy"               "RgbToCol"              
#> [367] "RgbToHex"               "RgbToLong"              "RndPairs"              
#> [370] "RndWord"                "RobScale"               "RomanToInt"            
#> [373] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [376] "RunsTest"               "SD"                     "SLN"                   
#> [379] "SMAPE"                  "SYD"                    "Sample"                
#> [382] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [385] "SecToHms"               "Second"                 "SendOutlookMail"       
#> [388] "Sens"                   "SetAlpha"               "SetAttr"               
#> [391] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [394] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [397] "Skew"                   "Small"                  "SmoothSpline"          
#> [400] "Some"                   "SomersDelta"            "Sort"                  
#> [403] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [406] "SphToCart"              "SplitAt"                "SplitPath"             
#> [409] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [412] "StdCoef"                "Str"                    "StrAbbr"               
#> [415] "StrAlign"               "StrCap"                 "StrChop"               
#> [418] "StrCountW"              "StrDist"                "StrExtract"            
#> [421] "StrExtractBetween"      "StrIsNumeric"           "StrLeft"               
#> [424] "StrPad"                 "StrPos"                 "StrRev"                
#> [427] "StrRight"               "StrSpell"               "StrSplit"              
#> [430] "StrTrim"                "StrTrunc"               "StrVal"                
#> [433] "Strata"                 "StripAttr"              "StuartMaxwellTest"     
#> [436] "StuartTauC"             "SysInfo"                "TMod"                  
#> [439] "TOne"                   "TTestA"                 "TextContrastColor"     
#> [442] "TextToTable"            "TheilU"                 "Timezone"              
#> [445] "TitleRect"              "ToLong"                 "ToWide"                
#> [448] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [451] "ToXL"                   "Today"                  "Trim"                  
#> [454] "TschuprowT"             "TukeyBiweight"          "TwoGroups"             
#> [457] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [460] "Unit<-"                 "Untable"                "Unwhich"               
#> [463] "VIF"                    "VanWaerdenTest"         "Var"                   
#> [466] "VarCI"                  "VarTest"                "VecRot"                
#> [469] "VecShift"               "Vigenere"               "VonNeumannTest"        
#> [472] "Week"                   "Weekday"                "Winsorize"             
#> [475] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [478] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [481] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [484] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [487] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [490] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [493] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [496] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [499] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [502] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [505] "XLSaveAs"               "XLView"                 "YTM"                   
#> [508] "Year"                   "YearDay"                "YearDays"              
#> [511] "YearMonth"              "YuenTTest"              "YuleQ"                 
#> [514] "YuleY"                  "ZTest"                  "ZeroIfNA"              
#> [517] "Zodiac"                 "aovlDetails"            "aovlErrorTerms"        
#> [520] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#> [523] "axTicks.Date"           "axTicks.POSIXct"        "createCOMReference"    
#> [526] "dBenf"                  "dExtrVal"               "dFrechet"              
#> [529] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [532] "dGumbel"                "dNegWeibull"            "dOrder"                
#> [535] "dRevGumbel"             "dRevWeibull"            "identify.formula"      
#> [538] "lines.Lc"               "lines.SmoothSpline"     "lines.lm"              
#> [541] "lines.loess"            "lines.smooth.spline"    "pBenf"                 
#> [544] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [547] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [550] "pNegWeibull"            "pOrder"                 "pRevGumbel"            
#> [553] "pRevWeibull"            "plot.Conf"              "plot.Lc"               
#> [556] "plot.Lclist"            "plot.PostHocTest"       "plot.bagplot"          
#> [559] "plot.palette"           "power.chisq.test"       "print.Assocs"          
#> [562] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [565] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [568] "print.PercTable"        "print.PostHocTest"      "print.mtest"           
#> [571] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [574] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [577] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [580] "qRevGumbelExp"          "qRevWeibull"            "rBenf"                 
#> [583] "rExtrVal"               "rFrechet"               "rGenExtrVal"           
#> [586] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [589] "rNegWeibull"            "rOrder"                 "rRevGumbel"            
#> [592] "rRevWeibull"            "rSum21"                 "reorder.factor"        
#> [595] "split.formula"