List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "Append"                 "AscToChar"             
#>  [28] "Asp"                    "Assocs"                 "Atkinson"              
#>  [31] "AxisBreak"              "BarText"                "BarnardTest"           
#>  [34] "BartelsRankTest"        "Bg"                     "BhapkarTest"           
#>  [37] "BinToDec"               "BinTree"                "BinomCI"               
#>  [40] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [43] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [46] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [49] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [52] "BrierScoreCI"           "BubbleLegend"           "CCC"                   
#>  [55] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [58] "CatTable"               "CharToAsc"              "Clockwise"             
#>  [61] "Closest"                "CmToPts"                "CmyToCmyk"             
#>  [64] "CmykToCmy"              "CmykToRgb"              "Coalesce"              
#>  [67] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [70] "CoefVarCI"              "CohenD"                 "CohenKappa"            
#>  [73] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [76] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [79] "CollapseTable"          "ColorLegend"            "ColumnWrap"            
#>  [82] "CombN"                  "CombPairs"              "CombSet"               
#>  [85] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [88] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [91] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [94] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [97] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [100] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CstatCI"                "CutAge"                
#> [109] "CutQ"                   "DB"                     "Day"                   
#> [112] "Day<-"                  "DecToBin"               "DecToHex"              
#> [115] "DecToOct"               "DegToRad"               "Desc"                  
#> [118] "DescToolsOptions"       "DiffDays360"            "DigitSum"              
#> [121] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [124] "DoBy"                   "DoCall"                 "Dot"                   
#> [127] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [130] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [133] "Dummy"                  "DunnTest"               "DunnettTest"           
#> [136] "DurbinWatsonTest"       "EX"                     "Entropy"               
#> [139] "Eps"                    "ErrBars"                "EtaSq"                 
#> [142] "ExpFreq"                "Factorize"              "Fade"                  
#> [145] "FctArgs"                "Fibonacci"              "FindColor"             
#> [148] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [151] "FisherZInv"             "FixToTable"             "Flags"                 
#> [154] "Fmt"                    "Format"                 "Frac"                  
#> [157] "Freq"                   "Freq2D"                 "GCD"                   
#> [160] "GTest"                  "GenRandGroups"          "GeomSn"                
#> [163] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [166] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [169] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [172] "GiniDeltas"             "GiniSimpson"            "Gmean"                 
#> [175] "GoodmanKruskalGamma"    "GoodmanKruskalTau"      "Gsd"                   
#> [178] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [181] "HexToRgb"               "HighLow"                "Hmean"                 
#> [184] "HmsToMinute"            "HmsToSec"               "HodgesLehmann"         
#> [187] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [190] "Hour"                   "HuberM"                 "HunterGaston"          
#> [193] "ICC"                    "IPMT"                   "IQRw"                  
#> [196] "IRR"                    "IdentifyA"              "Impute"                
#> [199] "ImputeKnn"              "InDots"                 "Interval"              
#> [202] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [205] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [208] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [211] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [214] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [217] "KendallTauB"            "KendallW"               "Keywords"              
#> [220] "KrippAlpha"             "Kurt"                   "LCM"                   
#> [223] "LOCF"                   "LOF"                    "Label"                 
#> [226] "Label<-"                "Labels"                 "Labels<-"              
#> [229] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [232] "Lc"                     "LehmacherTest"          "LeveneTest"            
#> [235] "LillieTest"             "LinScale"               "LineToUser"            
#> [238] "LogSt"                  "LogStInv"               "Logit"                 
#> [241] "LogitInv"               "LongToRgb"              "LsFct"                 
#> [244] "LsObj"                  "MAD"                    "MADCI"                 
#> [247] "MAE"                    "MAPE"                   "MHChisqTest"           
#> [250] "MSE"                    "Mar"                    "Margins"               
#> [253] "MaxDigits"              "Mean"                   "MeanAD"                
#> [256] "MeanCI"                 "MeanCIn"                "MeanDiffCI"            
#> [259] "MeanSE"                 "Median"                 "MedianCI"              
#> [262] "Mgp"                    "Mgsub"                  "Midx"                  
#> [265] "Minute"                 "MixColor"               "ModSummary"            
#> [268] "Mode"                   "Month"                  "MonthDays"             
#> [271] "MosesTest"              "MoveAvg"                "MultMerge"             
#> [274] "MultinomCI"             "MutInf"                 "N"                     
#> [277] "NAIf"                   "NAIfBlank"              "NAIfZero"              
#> [280] "NALevel"                "NMAE"                   "NMSE"                  
#> [283] "NPV"                    "NPVFixBond"             "NZ"                    
#> [286] "Ndec"                   "NemenyiTest"            "Nf"                    
#> [289] "Now"                    "OPR"                    "ORToRelRisk"           
#> [292] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [295] "Outlier"                "Overlap"                "PDFManual"             
#> [298] "PMT"                    "PPMT"                   "PageTest"              
#> [301] "PairApply"              "Pal"                    "ParseFormula"          
#> [304] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [307] "PearsonTest"            "PercTable"              "PercentRank"           
#> [310] "Permn"                  "Phi"                    "Phrase"                
#> [313] "PlotACF"                "PlotArea"               "PlotBag"               
#> [316] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [319] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [322] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [325] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [328] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [331] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [334] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [337] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [340] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [343] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [346] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [349] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [352] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [355] "Prec"                   "Primes"                 "PseudoR2"              
#> [358] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [361] "QuantileCI"             "Quarter"                "Quot"                  
#> [364] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [367] "RSqCI"                  "RTempdirAlive"          "RadToDeg"              
#> [370] "Range"                  "Rank"                   "ReadSPSS"              
#> [373] "Recode"                 "Recycle"                "RelRisk"               
#> [376] "Rename"                 "Rev"                    "RevCode"               
#> [379] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [382] "RgbToLong"              "RndPairs"               "RndWord"               
#> [385] "RobScale"               "RomanToInt"             "Rosenbluth"            
#> [388] "Rotate"                 "RoundTo"                "RunsTest"              
#> [391] "SD"                     "SDN"                    "SLN"                   
#> [394] "SMAPE"                  "SYD"                    "Sample"                
#> [397] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [400] "SecToHms"               "Second"                 "SendOutlookMail"       
#> [403] "Sens"                   "SetAlpha"               "SetAttr"               
#> [406] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [409] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [412] "Skew"                   "Small"                  "SmoothSpline"          
#> [415] "Some"                   "SomersDelta"            "Sort"                  
#> [418] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [421] "SphToCart"              "SplitAt"                "SplitPath"             
#> [424] "SplitToCol"             "SplitToDummy"           "SpreadOut"             
#> [427] "Stamp"                  "StdCoef"                "Str"                   
#> [430] "StrAbbr"                "StrAlign"               "StrCap"                
#> [433] "StrChop"                "StrCountW"              "StrDist"               
#> [436] "StrExtract"             "StrExtractBetween"      "StrIsNumeric"          
#> [439] "StrLeft"                "StrPad"                 "StrPos"                
#> [442] "StrRev"                 "StrRight"               "StrSpell"              
#> [445] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [448] "StrVal"                 "Strata"                 "StripAttr"             
#> [451] "StuartMaxwellTest"      "StuartTauC"             "SysInfo"               
#> [454] "TMod"                   "TOne"                   "TTestA"                
#> [457] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [460] "Timezone"               "TitleRect"              "ToLong"                
#> [463] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [466] "ToWrdPlot"              "ToXL"                   "Today"                 
#> [469] "Trim"                   "TschuprowT"             "TukeyBiweight"         
#> [472] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [475] "Unit"                   "Unit<-"                 "Untable"               
#> [478] "Unwhich"                "VIF"                    "VanWaerdenTest"        
#> [481] "Var"                    "VarCI"                  "VarN"                  
#> [484] "VarTest"                "VarX"                   "VecRot"                
#> [487] "VecShift"               "Vigenere"               "VonNeumannTest"        
#> [490] "Week"                   "Weekday"                "Winsorize"             
#> [493] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [496] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [499] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [502] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [505] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [508] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [511] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [514] "WrdTable"               "WrdTableBorders"        "WrdTableHeading"       
#> [517] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [520] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [523] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [526] "YTM"                    "Year"                   "YearDay"               
#> [529] "YearDays"               "YearMonth"              "YuenTTest"             
#> [532] "YuleQ"                  "YuleY"                  "ZTest"                 
#> [535] "ZeroIfNA"               "Zodiac"                 "aovlDetails"           
#> [538] "aovlErrorTerms"         "as.CDateFmt"            "as.fmt"                
#> [541] "as.matrix.xtabs"        "as.ym"                  "axTicks.Date"          
#> [544] "axTicks.POSIXct"        "createCOMReference"     "dBenf"                 
#> [547] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [550] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [553] "dNegWeibull"            "dOrder"                 "dRevGumbel"            
#> [556] "dRevWeibull"            "dTri"                   "identify.formula"      
#> [559] "lines.Lc"               "lines.SmoothSpline"     "lines.lm"              
#> [562] "lines.loess"            "lines.smooth.spline"    "pBenf"                 
#> [565] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [568] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [571] "pNegWeibull"            "pOrder"                 "pRevGumbel"            
#> [574] "pRevWeibull"            "pTri"                   "plot.Conf"             
#> [577] "plot.Lc"                "plot.Lclist"            "plot.PostHocTest"      
#> [580] "plot.bagplot"           "plot.palette"           "power.chisq.test"      
#> [583] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [586] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [589] "print.ICC"              "print.PercTable"        "print.PostHocTest"     
#> [592] "print.mtest"            "qBenf"                  "qExtrVal"              
#> [595] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [598] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [601] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [604] "qTri"                   "rBenf"                  "rExtrVal"              
#> [607] "rFrechet"               "rGenExtrVal"            "rGenPareto"            
#> [610] "rGompertz"              "rGumbel"                "rNegWeibull"           
#> [613] "rOrder"                 "rRevGumbel"             "rRevWeibull"           
#> [616] "rSum21"                 "rTri"                   "reorder.factor"        
#> [619] "split.formula"