List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BinToDec"              
#>  [46] "BinTree"                "BlankIfNA"              "BootCI"                
#>  [49] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [52] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [55] "BrierScore"             "BubbleLegend"           "Canvas"                
#>  [58] "CartToPol"              "CartToSph"              "CatTable"              
#>  [61] "CCC"                    "CharToAsc"              "Clockwise"             
#>  [64] "Closest"                "CmToPts"                "CmykToCmy"             
#>  [67] "CmykToRgb"              "CmyToCmyk"              "Coalesce"              
#>  [70] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [73] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [76] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [79] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [82] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [85] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [88] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [91] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [94] "ConvUnit"               "Cor"                    "CorCI"                 
#>  [97] "CorPart"                "CorPolychor"            "CountCompCases"        
#> [100] "CountWorkDays"          "CourseData"             "Cov"                   
#> [103] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [106] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [109] "Cstat"                  "CutQ"                   "Day"                   
#> [112] "Day<-"                  "DB"                     "dBenf"                 
#> [115] "DecToBin"               "DecToHex"               "DecToOct"              
#> [118] "DegToRad"               "DenseRank"              "Desc"                  
#> [121] "DescToolsOptions"       "dExtrVal"               "dFrechet"              
#> [124] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [127] "dGumbel"                "DiffDays360"            "DigitSum"              
#> [130] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [133] "dNegWeibull"            "DoBy"                   "DoCall"                
#> [136] "dOrder"                 "Dot"                    "DrawArc"               
#> [139] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [142] "DrawEllipse"            "DrawRegPolygon"         "dRevGumbel"            
#> [145] "dRevWeibull"            "Dummy"                  "DunnettTest"           
#> [148] "DunnTest"               "DurbinWatsonTest"       "Entropy"               
#> [151] "Eps"                    "ErrBars"                "EtaSq"                 
#> [154] "ExpFreq"                "Factorize"              "FctArgs"               
#> [157] "Fibonacci"              "FindColor"              "FindCorr"              
#> [160] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [163] "FixToTable"             "Flags"                  "Fmt"                   
#> [166] "Format"                 "Frac"                   "Freq"                  
#> [169] "Freq2D"                 "GCD"                    "GeomSn"                
#> [172] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [175] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [178] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [181] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [184] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [187] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [190] "HexToRgb"               "HighLow"                "Hmean"                 
#> [193] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [196] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [199] "HuberM"                 "HunterGaston"           "ICC"                   
#> [202] "identify.formula"       "IdentifyA"              "Impute"                
#> [205] "ImputeKnn"              "InDots"                 "Interval"              
#> [208] "IPMT"                   "IQRw"                   "IRR"                   
#> [211] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [214] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [217] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [220] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [223] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [226] "KendallTauB"            "KendallW"               "Keywords"              
#> [229] "KrippAlpha"             "Kurt"                   "Label"                 
#> [232] "Label<-"                "Lambda"                 "Large"                 
#> [235] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [238] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [241] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [244] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [247] "LinScale"               "LOCF"                   "LOF"                   
#> [250] "Logit"                  "LogitInv"               "LogSt"                 
#> [253] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [256] "LsObj"                  "MAD"                    "MADCI"                 
#> [259] "MAE"                    "MAPE"                   "Mar"                   
#> [262] "Margins"                "MaxDigits"              "Mean"                  
#> [265] "MeanAD"                 "MeanCI"                 "MeanDiffCI"            
#> [268] "MeanSE"                 "Median"                 "MedianCI"              
#> [271] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [274] "Minute"                 "MixColor"               "Mode"                  
#> [277] "ModSummary"             "Month"                  "MosesTest"             
#> [280] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [283] "MultMerge"              "MutInf"                 "N"                     
#> [286] "NAIfBlank"              "NAIfZero"               "Ndec"                  
#> [289] "NemenyiTest"            "NMAE"                   "NMSE"                  
#> [292] "Now"                    "NPV"                    "NPVFixBond"            
#> [295] "OctToDec"               "OddsRatio"              "OPR"                   
#> [298] "OrderMixed"             "ORToRelRisk"            "Outlier"               
#> [301] "Overlap"                "PageTest"               "PairApply"             
#> [304] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [307] "PartialSD"              "PasswordDlg"            "pBenf"                 
#> [310] "PDFManual"              "PearsonTest"            "PercentRank"           
#> [313] "PercTable"              "Permn"                  "pExtrVal"              
#> [316] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [319] "pGompertz"              "pGumbel"                "Phi"                   
#> [322] "Phrase"                 "plot.bagplot"           "plot.Conf"             
#> [325] "plot.Lc"                "plot.Lclist"            "plot.palette"          
#> [328] "plot.PostHocTest"       "PlotACF"                "PlotArea"              
#> [331] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [334] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [337] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [340] "PlotDot"                "PlotECDF"               "PlotFaces"             
#> [343] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [346] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [349] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [352] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [355] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [358] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [361] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [364] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [367] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [370] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [373] "PpText"                 "Prec"                   "pRevGumbel"            
#> [376] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [379] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [382] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [385] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [388] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [391] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [394] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [397] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [400] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [403] "QuantileCI"             "Quarter"                "Quot"                  
#> [406] "RadToDeg"               "Range"                  "RBAL"                  
#> [409] "rBenf"                  "Recode"                 "Recycle"               
#> [412] "RelRisk"                "Rename"                 "reorder.factor"        
#> [415] "Rev"                    "RevCode"                "rExtrVal"              
#> [418] "rFrechet"               "RgbToCmy"               "RgbToCol"              
#> [421] "RgbToHex"               "RgbToLong"              "rGenExtrVal"           
#> [424] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [427] "RMSE"                   "RndPairs"               "RndWord"               
#> [430] "rNegWeibull"            "RobScale"               "RomanToInt"            
#> [433] "rOrder"                 "Rosenbluth"             "Rotate"                
#> [436] "RoundTo"                "rRevGumbel"             "rRevWeibull"           
#> [439] "RSessionAlive"          "RTempdirAlive"          "RunsTest"              
#> [442] "Sample"                 "SampleTwins"            "SaveAs"                
#> [445] "ScheffeTest"            "SD"                     "Second"                
#> [448] "SecToHms"               "SendOutlookMail"        "Sens"                  
#> [451] "SetAlpha"               "SetNames"               "Shade"                 
#> [454] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [457] "SignTest"               "Skew"                   "SLN"                   
#> [460] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [463] "Some"                   "SomersDelta"            "Sort"                  
#> [466] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [469] "SphToCart"              "split.formula"          "SplitAt"               
#> [472] "SplitPath"              "SplitToCol"             "SpreadOut"             
#> [475] "Stamp"                  "StdCoef"                "Str"                   
#> [478] "StrAbbr"                "StrAlign"               "Strata"                
#> [481] "StrCap"                 "StrChop"                "StrCountW"             
#> [484] "StrDist"                "StrExtract"             "StripAttr"             
#> [487] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [490] "StrPos"                 "StrRev"                 "StrRight"              
#> [493] "StrSpell"               "StrSplit"               "StrTrim"               
#> [496] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [499] "StuartTauC"             "SYD"                    "SysInfo"               
#> [502] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [505] "Timezone"               "TitleRect"              "TMod"                  
#> [508] "Today"                  "ToLong"                 "TOne"                  
#> [511] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [514] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [517] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [520] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [523] "Unit"                   "Unit<-"                 "Untable"               
#> [526] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [529] "VarCI"                  "VarTest"                "VecRot"                
#> [532] "VecShift"               "VIF"                    "Vigenere"              
#> [535] "VonNeumannTest"         "Week"                   "Weekday"               
#> [538] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [541] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [544] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [547] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [550] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [553] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [556] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [559] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [562] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [565] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [568] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [571] "Year"                   "YearDay"                "YearMonth"             
#> [574] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [577] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [580] "ZTest"