List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BhapkarTest"            "BinomCI"               
#>  [43] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [46] "BinToDec"               "BinTree"                "BlankIfNA"             
#>  [49] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [52] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [55] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [58] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [61] "CatTable"               "CCC"                    "CharToAsc"             
#>  [64] "Clockwise"              "Closest"                "CmToPts"               
#>  [67] "CmykToCmy"              "CmykToRgb"              "CmyToCmyk"             
#>  [70] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [73] "CoefVar"                "CohenD"                 "CohenKappa"            
#>  [76] "CollapseTable"          "ColorLegend"            "ColToGray"             
#>  [79] "ColToGrey"              "ColToHex"               "ColToHsv"              
#>  [82] "ColToOpaque"            "ColToRgb"               "ColumnWrap"            
#>  [85] "CombN"                  "CombPairs"              "CombSet"               
#>  [88] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [91] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [94] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [97] "CorCI"                  "CorPart"                "CorPolychor"           
#> [100] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [103] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [106] "createCOMReference"     "CronbachAlpha"          "Cross"                 
#> [109] "CrossN"                 "Cstat"                  "CutQ"                  
#> [112] "Day"                    "Day<-"                  "DB"                    
#> [115] "dBenf"                  "DecToBin"               "DecToHex"              
#> [118] "DecToOct"               "DegToRad"               "DenseRank"             
#> [121] "Desc"                   "DescToolsOptions"       "dExtrVal"              
#> [124] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [127] "dGompertz"              "dGumbel"                "DiffDays360"           
#> [130] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [133] "Divisors"               "dNegWeibull"            "DoBy"                  
#> [136] "DoCall"                 "dOrder"                 "Dot"                   
#> [139] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [142] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [145] "dRevGumbel"             "dRevWeibull"            "Dummy"                 
#> [148] "DunnettTest"            "DunnTest"               "DurbinWatsonTest"      
#> [151] "Entropy"                "Eps"                    "ErrBars"               
#> [154] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [157] "FctArgs"                "Fibonacci"              "FindColor"             
#> [160] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [163] "FisherZInv"             "FixToTable"             "Flags"                 
#> [166] "Fmt"                    "Format"                 "Frac"                  
#> [169] "Freq"                   "Freq2D"                 "GCD"                   
#> [172] "GeomSn"                 "GeomTrans"              "GetCalls"              
#> [175] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [178] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [181] "Gini"                   "GiniSimpson"            "Gmean"                 
#> [184] "GoodmanKruskalGamma"    "GoodmanKruskalTau"      "Gsd"                   
#> [187] "GTest"                  "Herfindahl"             "HexToCol"              
#> [190] "HexToDec"               "HexToRgb"               "HighLow"               
#> [193] "Hmean"                  "HmsToSec"               "HodgesLehmann"         
#> [196] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [199] "Hour"                   "HuberM"                 "HunterGaston"          
#> [202] "ICC"                    "identify.formula"       "IdentifyA"             
#> [205] "Impute"                 "ImputeKnn"              "InDots"                
#> [208] "Interval"               "IPMT"                   "IQRw"                  
#> [211] "IRR"                    "IsDate"                 "IsDichotomous"         
#> [214] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [217] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [220] "IsWeekend"              "IsWhole"                "IsZero"                
#> [223] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [226] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [229] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [232] "Label"                  "Label<-"                "Lambda"                
#> [235] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [238] "LCM"                    "LehmacherTest"          "LeveneTest"            
#> [241] "LillieTest"             "lines.Lc"               "lines.lm"              
#> [244] "lines.loess"            "lines.smooth.spline"    "lines.SmoothSpline"    
#> [247] "LineToUser"             "LinScale"               "LOCF"                  
#> [250] "LOF"                    "Logit"                  "LogitInv"              
#> [253] "LogSt"                  "LogStInv"               "LongToRgb"             
#> [256] "LsFct"                  "LsObj"                  "MAD"                   
#> [259] "MADCI"                  "MAE"                    "MAPE"                  
#> [262] "Mar"                    "Margins"                "MaxDigits"             
#> [265] "Mean"                   "MeanAD"                 "MeanCI"                
#> [268] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [271] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [274] "Midx"                   "Minute"                 "MixColor"              
#> [277] "Mode"                   "ModSummary"             "Month"                 
#> [280] "MosesTest"              "MoveAvg"                "MSE"                   
#> [283] "MultinomCI"             "MultMerge"              "MutInf"                
#> [286] "N"                      "NAIfBlank"              "NAIfZero"              
#> [289] "Ndec"                   "NemenyiTest"            "NMAE"                  
#> [292] "NMSE"                   "Now"                    "NPV"                   
#> [295] "NPVFixBond"             "OctToDec"               "OddsRatio"             
#> [298] "OPR"                    "OrderMixed"             "ORToRelRisk"           
#> [301] "Outlier"                "Overlap"                "PageTest"              
#> [304] "PairApply"              "Pal"                    "ParseFormula"          
#> [307] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [310] "pBenf"                  "PDFManual"              "PearsonTest"           
#> [313] "PercentRank"            "PercTable"              "Permn"                 
#> [316] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [319] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [322] "Phi"                    "Phrase"                 "plot.bagplot"          
#> [325] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [328] "plot.palette"           "plot.PostHocTest"       "PlotACF"               
#> [331] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [334] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [337] "PlotCashFlow"           "PlotCirc"               "PlotConDens"           
#> [340] "PlotCorr"               "PlotDot"                "PlotECDF"              
#> [343] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [346] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [349] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [352] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [355] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [358] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [361] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [364] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [367] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [370] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [373] "PpPlot"                 "PpText"                 "Prec"                  
#> [376] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [379] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [382] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [385] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [388] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [391] "PtsToCm"                "qBenf"                  "qExtrVal"              
#> [394] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [397] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [400] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [403] "Quantile"               "QuantileCI"             "Quarter"               
#> [406] "Quot"                   "RadToDeg"               "Range"                 
#> [409] "RBAL"                   "rBenf"                  "Recode"                
#> [412] "Recycle"                "RelRisk"                "Rename"                
#> [415] "reorder.factor"         "Rev"                    "RevCode"               
#> [418] "rExtrVal"               "rFrechet"               "RgbToCmy"              
#> [421] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [424] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [427] "rGumbel"                "RMSE"                   "RndPairs"              
#> [430] "RndWord"                "rNegWeibull"            "RobScale"              
#> [433] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [436] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [439] "rRevWeibull"            "RSessionAlive"          "RTempdirAlive"         
#> [442] "RunsTest"               "Sample"                 "SampleTwins"           
#> [445] "SaveAs"                 "ScheffeTest"            "SD"                    
#> [448] "Second"                 "SecToHms"               "SendOutlookMail"       
#> [451] "Sens"                   "SetAlpha"               "SetNames"              
#> [454] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [457] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [460] "SLN"                    "Small"                  "SMAPE"                 
#> [463] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [466] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [469] "Spec"                   "SphToCart"              "split.formula"         
#> [472] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [475] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [478] "Str"                    "StrAbbr"                "StrAlign"              
#> [481] "Strata"                 "StrCap"                 "StrChop"               
#> [484] "StrCountW"              "StrDist"                "StrExtract"            
#> [487] "StrExtractBetween"      "StripAttr"              "StrIsNumeric"          
#> [490] "StrLeft"                "StrPad"                 "StrPos"                
#> [493] "StrRev"                 "StrRight"               "StrSpell"              
#> [496] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [499] "StrVal"                 "StuartMaxwellTest"      "StuartTauC"            
#> [502] "SYD"                    "SysInfo"                "TextContrastColor"     
#> [505] "TextToTable"            "TheilU"                 "Timezone"              
#> [508] "TitleRect"              "TMod"                   "Today"                 
#> [511] "ToLong"                 "TOne"                   "ToWide"                
#> [514] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [517] "ToXL"                   "Trim"                   "TschuprowT"            
#> [520] "TTestA"                 "TukeyBiweight"          "TwoGroups"             
#> [523] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [526] "Unit<-"                 "Untable"                "Unwhich"               
#> [529] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [532] "VarTest"                "VecRot"                 "VecShift"              
#> [535] "VIF"                    "Vigenere"               "VonNeumannTest"        
#> [538] "Week"                   "Weekday"                "Winsorize"             
#> [541] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [544] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [547] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [550] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [553] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [556] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [559] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [562] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [565] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [568] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [571] "XLSaveAs"               "XLView"                 "Year"                  
#> [574] "YearDay"                "YearMonth"              "YTM"                   
#> [577] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [580] "ZeroIfNA"               "Zodiac"                 "ZTest"