List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%:%"                    "%()%"                   "%(]%"                  
#>   [4] "%)(%"                   "%)[%"                   "%[)%"                  
#>   [7] "%[]%"                   "%](%"                   "%][%"                  
#>  [10] "%^%"                    "%c%"                    "%like any%"            
#>  [13] "%like%"                 "%nin%"                  "%overlaps%"            
#>  [16] "Abind"                  "Abstract"               "AddMonths"             
#>  [19] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [22] "AllIdentical"           "AndersonDarlingTest"    "aovlDetails"           
#>  [25] "aovlErrorTerms"         "Append"                 "Arrow"                 
#>  [28] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#>  [31] "AscToChar"              "Asp"                    "Assocs"                
#>  [34] "Atkinson"               "AUC"                    "AxisBreak"             
#>  [37] "axTicks.Date"           "axTicks.POSIXct"        "BarnardTest"           
#>  [40] "BartelsRankTest"        "BarText"                "BhapkarTest"           
#>  [43] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [46] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [49] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [52] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [55] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [58] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [61] "CartToSph"              "CatTable"               "CCC"                   
#>  [64] "CharToAsc"              "Clockwise"              "Closest"               
#>  [67] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [70] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [73] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [76] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [79] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [82] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [85] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [88] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [91] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [94] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [97] "Cor"                    "CorCI"                  "CorPart"               
#> [100] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [103] "CourseData"             "Cov"                    "CramerV"               
#> [106] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [109] "Cross"                  "CrossN"                 "Cstat"                 
#> [112] "CutQ"                   "Day"                    "Day<-"                 
#> [115] "DB"                     "dBenf"                  "DecToBin"              
#> [118] "DecToHex"               "DecToOct"               "DegToRad"              
#> [121] "DenseRank"              "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanDiffCI"            
#> [271] "MeanSE"                 "Median"                 "MedianCI"              
#> [274] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [277] "Minute"                 "MixColor"               "Mode"                  
#> [280] "ModSummary"             "Month"                  "MosesTest"             
#> [283] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [286] "MultMerge"              "MutInf"                 "N"                     
#> [289] "NAIf"                   "NAIfBlank"              "NAIfZero"              
#> [292] "Ndec"                   "NemenyiTest"            "NMAE"                  
#> [295] "NMSE"                   "Now"                    "NPV"                   
#> [298] "NPVFixBond"             "OctToDec"               "OddsRatio"             
#> [301] "OPR"                    "OrderMixed"             "ORToRelRisk"           
#> [304] "Outlier"                "Overlap"                "PageTest"              
#> [307] "PairApply"              "Pal"                    "ParseFormula"          
#> [310] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [313] "pBenf"                  "PDFManual"              "PearsonTest"           
#> [316] "PercentRank"            "PercTable"              "Permn"                 
#> [319] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [322] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [325] "Phi"                    "Phrase"                 "plot.bagplot"          
#> [328] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [331] "plot.palette"           "plot.PostHocTest"       "PlotACF"               
#> [334] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [337] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [340] "PlotCashFlow"           "PlotCirc"               "PlotConDens"           
#> [343] "PlotCorr"               "PlotDot"                "PlotECDF"              
#> [346] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [349] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [352] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [355] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [358] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [361] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [364] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [367] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [370] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [373] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [376] "PpPlot"                 "PpText"                 "Prec"                  
#> [379] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [382] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [385] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [388] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [391] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [394] "PtsToCm"                "qBenf"                  "qExtrVal"              
#> [397] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [400] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [403] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [406] "Quantile"               "QuantileCI"             "Quarter"               
#> [409] "Quot"                   "RadToDeg"               "Range"                 
#> [412] "RBAL"                   "rBenf"                  "Recode"                
#> [415] "Recycle"                "RelRisk"                "Rename"                
#> [418] "reorder.factor"         "Rev"                    "RevCode"               
#> [421] "rExtrVal"               "rFrechet"               "RgbToCmy"              
#> [424] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [427] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [430] "rGumbel"                "RMSE"                   "RndPairs"              
#> [433] "RndWord"                "rNegWeibull"            "RobScale"              
#> [436] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [439] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [442] "rRevWeibull"            "RSessionAlive"          "RTempdirAlive"         
#> [445] "RunsTest"               "Sample"                 "SampleTwins"           
#> [448] "SaveAs"                 "ScheffeTest"            "SD"                    
#> [451] "Second"                 "SecToHms"               "SendOutlookMail"       
#> [454] "Sens"                   "SetAlpha"               "SetNames"              
#> [457] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [460] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [463] "SLN"                    "Small"                  "SMAPE"                 
#> [466] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [469] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [472] "Spec"                   "SphToCart"              "split.formula"         
#> [475] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [478] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [481] "Str"                    "StrAbbr"                "StrAlign"              
#> [484] "Strata"                 "StrCap"                 "StrChop"               
#> [487] "StrCountW"              "StrDist"                "StrExtract"            
#> [490] "StrExtractBetween"      "StripAttr"              "StrIsNumeric"          
#> [493] "StrLeft"                "StrPad"                 "StrPos"                
#> [496] "StrRev"                 "StrRight"               "StrSpell"              
#> [499] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [502] "StrVal"                 "StuartMaxwellTest"      "StuartTauC"            
#> [505] "SYD"                    "SysInfo"                "TextContrastColor"     
#> [508] "TextToTable"            "TheilU"                 "Timezone"              
#> [511] "TitleRect"              "TMod"                   "Today"                 
#> [514] "ToLong"                 "TOne"                   "ToWide"                
#> [517] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [520] "ToXL"                   "Trim"                   "TschuprowT"            
#> [523] "TTestA"                 "TukeyBiweight"          "TwoGroups"             
#> [526] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [529] "Unit<-"                 "Untable"                "Unwhich"               
#> [532] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [535] "VarTest"                "VecRot"                 "VecShift"              
#> [538] "VIF"                    "Vigenere"               "VonNeumannTest"        
#> [541] "Week"                   "Weekday"                "Winsorize"             
#> [544] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [547] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [550] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [553] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [556] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [559] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [562] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [565] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [568] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [571] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [574] "XLSaveAs"               "XLView"                 "Year"                  
#> [577] "YearDay"                "YearMonth"              "YTM"                   
#> [580] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [583] "ZeroIfNA"               "Zodiac"                 "ZTest"