List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "Append"                 "AscToChar"             
#>  [28] "Asp"                    "Assocs"                 "Atkinson"              
#>  [31] "AxisBreak"              "BarText"                "BarnardTest"           
#>  [34] "BartelsRankTest"        "Bg"                     "BhapkarTest"           
#>  [37] "BinToDec"               "BinTree"                "BinomCI"               
#>  [40] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [43] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [46] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [49] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [52] "BubbleLegend"           "CCC"                    "Canvas"                
#>  [55] "CartToPol"              "CartToSph"              "CatTable"              
#>  [58] "CharToAsc"              "Clockwise"              "Closest"               
#>  [61] "CmToPts"                "CmyToCmyk"              "CmykToCmy"             
#>  [64] "CmykToRgb"              "Coalesce"               "CochranArmitageTest"   
#>  [67] "CochranQTest"           "CoefVar"                "CoefVarCI"             
#>  [70] "CohenD"                 "CohenKappa"             "ColToGray"             
#>  [73] "ColToGrey"              "ColToHex"               "ColToHsv"              
#>  [76] "ColToOpaque"            "ColToRgb"               "CollapseTable"         
#>  [79] "ColorLegend"            "ColumnWrap"             "CombN"                 
#>  [82] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [85] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [88] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [91] "ConvUnit"               "Cor"                    "CorCI"                 
#>  [94] "CorPart"                "CorPolychor"            "CountCompCases"        
#>  [97] "CountWorkDays"          "CourseData"             "Cov"                   
#> [100] "CramerV"                "CramerVonMisesTest"     "CronbachAlpha"         
#> [103] "Cross"                  "CrossN"                 "Cstat"                 
#> [106] "CutAge"                 "CutQ"                   "DB"                    
#> [109] "Day"                    "Day<-"                  "DecToBin"              
#> [112] "DecToHex"               "DecToOct"               "DegToRad"              
#> [115] "Desc"                   "DescToolsOptions"       "DiffDays360"           
#> [118] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [121] "Divisors"               "DoBy"                   "DoCall"                
#> [124] "Dot"                    "DrawArc"                "DrawBand"              
#> [127] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [130] "DrawRegPolygon"         "Dummy"                  "DunnTest"              
#> [133] "DunnettTest"            "DurbinWatsonTest"       "EX"                    
#> [136] "Entropy"                "Eps"                    "ErrBars"               
#> [139] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [142] "Fade"                   "FctArgs"                "Fibonacci"             
#> [145] "FindColor"              "FindCorr"               "FindRProfile"          
#> [148] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [151] "Flags"                  "Fmt"                    "Format"                
#> [154] "Frac"                   "Freq"                   "Freq2D"                
#> [157] "GCD"                    "GTest"                  "GenRandGroups"         
#> [160] "GeomSn"                 "GeomTrans"              "GetCalls"              
#> [163] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [166] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [169] "Gini"                   "GiniDeltas"             "GiniSimpson"           
#> [172] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [175] "Gsd"                    "Herfindahl"             "HexToCol"              
#> [178] "HexToDec"               "HexToRgb"               "HighLow"               
#> [181] "Hmean"                  "HmsToMinute"            "HmsToSec"              
#> [184] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [187] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [190] "HunterGaston"           "ICC"                    "IPMT"                  
#> [193] "IQRw"                   "IRR"                    "IdentifyA"             
#> [196] "Impute"                 "ImputeKnn"              "InDots"                
#> [199] "Interval"               "IsDate"                 "IsDichotomous"         
#> [202] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [205] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [208] "IsWeekend"              "IsWhole"                "IsZero"                
#> [211] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [214] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [217] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [220] "LCM"                    "LOCF"                   "LOF"                   
#> [223] "Label"                  "Label<-"                "Labels"                
#> [226] "Labels<-"               "Lambda"                 "Large"                 
#> [229] "LastDayOfMonth"         "Lc"                     "LehmacherTest"         
#> [232] "LeveneTest"             "LillieTest"             "LinScale"              
#> [235] "LineToUser"             "LogSt"                  "LogStInv"              
#> [238] "Logit"                  "LogitInv"               "LongToRgb"             
#> [241] "LsFct"                  "LsObj"                  "MAD"                   
#> [244] "MADCI"                  "MAE"                    "MAPE"                  
#> [247] "MHChisqTest"            "MSE"                    "Mar"                   
#> [250] "Margins"                "MaxDigits"              "Mean"                  
#> [253] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [256] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [259] "MedianCI"               "Mgp"                    "Mgsub"                 
#> [262] "Midx"                   "Minute"                 "MixColor"              
#> [265] "ModSummary"             "Mode"                   "Month"                 
#> [268] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [271] "MultMerge"              "MultinomCI"             "MutInf"                
#> [274] "N"                      "NAIf"                   "NAIfBlank"             
#> [277] "NAIfZero"               "NALevel"                "NMAE"                  
#> [280] "NMSE"                   "NPV"                    "NPVFixBond"            
#> [283] "NZ"                     "Ndec"                   "NemenyiTest"           
#> [286] "Nf"                     "Now"                    "OPR"                   
#> [289] "ORToRelRisk"            "OctToDec"               "OddsRatio"             
#> [292] "OrderMixed"             "Outlier"                "Overlap"               
#> [295] "PDFManual"              "PMT"                    "PPMT"                  
#> [298] "PageTest"               "PairApply"              "Pal"                   
#> [301] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [304] "PasswordDlg"            "PearsonTest"            "PercTable"             
#> [307] "PercentRank"            "Permn"                  "Phi"                   
#> [310] "Phrase"                 "PlotACF"                "PlotArea"              
#> [313] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [316] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [319] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [322] "PlotDot"                "PlotDotCI"              "PlotECDF"              
#> [325] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [328] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [331] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [334] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [337] "PlotPolar"              "PlotProbDist"           "PlotPyramid"           
#> [340] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [343] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [346] "PoissonCI"              "PolToCart"              "PolarGrid"             
#> [349] "PostHocTest"            "PpAddSlide"             "PpPlot"                
#> [352] "PpText"                 "Prec"                   "Primes"                
#> [355] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [358] "Quantile"               "QuantileCI"             "Quarter"               
#> [361] "Quot"                   "RBAL"                   "RMSE"                  
#> [364] "RSessionAlive"          "RTempdirAlive"          "RadToDeg"              
#> [367] "Range"                  "Rank"                   "Recode"                
#> [370] "Recycle"                "RelRisk"                "Rename"                
#> [373] "Rev"                    "RevCode"                "RgbToCmy"              
#> [376] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [379] "RndPairs"               "RndWord"                "RobScale"              
#> [382] "RomanToInt"             "Rosenbluth"             "Rotate"                
#> [385] "RoundTo"                "RunsTest"               "SD"                    
#> [388] "SDN"                    "SLN"                    "SMAPE"                 
#> [391] "SYD"                    "Sample"                 "SampleTwins"           
#> [394] "SaveAs"                 "ScheffeTest"            "SecToHms"              
#> [397] "Second"                 "SendOutlookMail"        "Sens"                  
#> [400] "SetAlpha"               "SetAttr"                "SetNames"              
#> [403] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [406] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [409] "Small"                  "SmoothSpline"           "Some"                  
#> [412] "SomersDelta"            "Sort"                   "SortMixed"             
#> [415] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [418] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [421] "SplitToDummy"           "SpreadOut"              "Stamp"                 
#> [424] "StdCoef"                "Str"                    "StrAbbr"               
#> [427] "StrAlign"               "StrCap"                 "StrChop"               
#> [430] "StrCountW"              "StrDist"                "StrExtract"            
#> [433] "StrExtractBetween"      "StrIsNumeric"           "StrLeft"               
#> [436] "StrPad"                 "StrPos"                 "StrRev"                
#> [439] "StrRight"               "StrSpell"               "StrSplit"              
#> [442] "StrTrim"                "StrTrunc"               "StrVal"                
#> [445] "Strata"                 "StripAttr"              "StuartMaxwellTest"     
#> [448] "StuartTauC"             "SysInfo"                "TMod"                  
#> [451] "TOne"                   "TTestA"                 "TextContrastColor"     
#> [454] "TextToTable"            "TheilU"                 "Timezone"              
#> [457] "TitleRect"              "ToLong"                 "ToWide"                
#> [460] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [463] "ToXL"                   "Today"                  "Trim"                  
#> [466] "TschuprowT"             "TukeyBiweight"          "TwoGroups"             
#> [469] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [472] "Unit<-"                 "Untable"                "Unwhich"               
#> [475] "VIF"                    "VanWaerdenTest"         "Var"                   
#> [478] "VarCI"                  "VarN"                   "VarTest"               
#> [481] "VarX"                   "VecRot"                 "VecShift"              
#> [484] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [487] "Weekday"                "Winsorize"              "WithOptions"           
#> [490] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [493] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [496] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [499] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [502] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [505] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [508] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [511] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [514] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [517] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [520] "XLView"                 "YTM"                    "Year"                  
#> [523] "YearDay"                "YearDays"               "YearMonth"             
#> [526] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [529] "ZTest"                  "ZeroIfNA"               "Zodiac"                
#> [532] "aovlDetails"            "aovlErrorTerms"         "as.CDateFmt"           
#> [535] "as.fmt"                 "as.matrix.xtabs"        "as.ym"                 
#> [538] "axTicks.Date"           "axTicks.POSIXct"        "createCOMReference"    
#> [541] "dBenf"                  "dExtrVal"               "dFrechet"              
#> [544] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [547] "dGumbel"                "dNegWeibull"            "dOrder"                
#> [550] "dRevGumbel"             "dRevWeibull"            "dTri"                  
#> [553] "identify.formula"       "lines.Lc"               "lines.SmoothSpline"    
#> [556] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [559] "pBenf"                  "pExtrVal"               "pFrechet"              
#> [562] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [565] "pGumbel"                "pNegWeibull"            "pOrder"                
#> [568] "pRevGumbel"             "pRevWeibull"            "pTri"                  
#> [571] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [574] "plot.PostHocTest"       "plot.bagplot"           "plot.palette"          
#> [577] "power.chisq.test"       "print.Assocs"           "print.Conf"            
#> [580] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [583] "print.HoeffD"           "print.ICC"              "print.PercTable"       
#> [586] "print.PostHocTest"      "print.mtest"            "qBenf"                 
#> [589] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [592] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [595] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [598] "qRevWeibull"            "qTri"                   "rBenf"                 
#> [601] "rExtrVal"               "rFrechet"               "rGenExtrVal"           
#> [604] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [607] "rNegWeibull"            "rOrder"                 "rRevGumbel"            
#> [610] "rRevWeibull"            "rSum21"                 "rTri"                  
#> [613] "reorder.factor"         "split.formula"