List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [25] "AllIdentical"           "AndersonDarlingTest"    "Append"                
#>  [28] "Arrow"                  "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AxisBreak"             
#>  [34] "BarText"                "BarnardTest"            "BartelsRankTest"       
#>  [37] "Bg"                     "BhapkarTest"            "BinToDec"              
#>  [40] "BinTree"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "CCC"                    "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "CmToPts"               
#>  [64] "CmyToCmyk"              "CmykToCmy"              "CmykToRgb"             
#>  [67] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [70] "CoefVar"                "CoefVarCI"              "CohenD"                
#>  [73] "CohenKappa"             "ColToGray"              "ColToGrey"             
#>  [76] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [79] "ColToRgb"               "CollapseTable"          "ColorLegend"           
#>  [82] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [85] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [88] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [91] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [94] "Cor"                    "CorCI"                  "CorPart"               
#>  [97] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [100] "CourseData"             "Cov"                    "CramerV"               
#> [103] "CramerVonMisesTest"     "CronbachAlpha"          "Cross"                 
#> [106] "CrossN"                 "Cstat"                  "CutQ"                  
#> [109] "DB"                     "Day"                    "Day<-"                 
#> [112] "DecToBin"               "DecToHex"               "DecToOct"              
#> [115] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [118] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [121] "DivCoefMax"             "Divisors"               "DoBy"                  
#> [124] "DoCall"                 "Dot"                    "DrawArc"               
#> [127] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [130] "DrawEllipse"            "DrawRegPolygon"         "Dummy"                 
#> [133] "DunnTest"               "DunnettTest"            "DurbinWatsonTest"      
#> [136] "Entropy"                "Eps"                    "ErrBars"               
#> [139] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [142] "FctArgs"                "Fibonacci"              "FindColor"             
#> [145] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [148] "FisherZInv"             "FixToTable"             "Flags"                 
#> [151] "Fmt"                    "Format"                 "Frac"                  
#> [154] "Freq"                   "Freq2D"                 "GCD"                   
#> [157] "GTest"                  "GenRandGroups"          "GeomSn"                
#> [160] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [163] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [166] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [169] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [172] "GoodmanKruskalTau"      "Gsd"                    "Herfindahl"            
#> [175] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [178] "HighLow"                "Hmean"                  "HmsToMinute"           
#> [181] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [184] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [187] "HuberM"                 "HunterGaston"           "ICC"                   
#> [190] "IPMT"                   "IQRw"                   "IRR"                   
#> [193] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [196] "InDots"                 "Interval"               "IsDate"                
#> [199] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [202] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [205] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [208] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [211] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [214] "KendallW"               "Keywords"               "KrippAlpha"            
#> [217] "Kurt"                   "LCM"                    "LOCF"                  
#> [220] "LOF"                    "Label"                  "Label<-"               
#> [223] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [226] "Lc"                     "LehmacherTest"          "LeveneTest"            
#> [229] "LillieTest"             "LinScale"               "LineToUser"            
#> [232] "LogSt"                  "LogStInv"               "Logit"                 
#> [235] "LogitInv"               "LongToRgb"              "LsFct"                 
#> [238] "LsObj"                  "MAD"                    "MADCI"                 
#> [241] "MAE"                    "MAPE"                   "MHChisqTest"           
#> [244] "MSE"                    "Mar"                    "Margins"               
#> [247] "MaxDigits"              "Mean"                   "MeanAD"                
#> [250] "MeanCI"                 "MeanCIn"                "MeanDiffCI"            
#> [253] "MeanSE"                 "Median"                 "MedianCI"              
#> [256] "Mgp"                    "Midx"                   "Minute"                
#> [259] "MixColor"               "ModSummary"             "Mode"                  
#> [262] "Month"                  "MonthDays"              "MosesTest"             
#> [265] "MoveAvg"                "MultMerge"              "MultinomCI"            
#> [268] "MutInf"                 "N"                      "NAIf"                  
#> [271] "NAIfBlank"              "NAIfZero"               "NMAE"                  
#> [274] "NMSE"                   "NPV"                    "NPVFixBond"            
#> [277] "Ndec"                   "NemenyiTest"            "Nf"                    
#> [280] "Now"                    "OPR"                    "ORToRelRisk"           
#> [283] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [286] "Outlier"                "Overlap"                "PDFManual"             
#> [289] "PMT"                    "PPMT"                   "PageTest"              
#> [292] "PairApply"              "Pal"                    "ParseFormula"          
#> [295] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [298] "PearsonTest"            "PercTable"              "PercentRank"           
#> [301] "Permn"                  "Phi"                    "Phrase"                
#> [304] "PlotACF"                "PlotArea"               "PlotBag"               
#> [307] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [310] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [313] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [316] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [319] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [322] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [325] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [328] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [331] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [334] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [337] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [340] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [343] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [346] "Prec"                   "Primes"                 "PseudoR2"              
#> [349] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [352] "QuantileCI"             "Quarter"                "Quot"                  
#> [355] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [358] "RTempdirAlive"          "RadToDeg"               "Range"                 
#> [361] "Rank"                   "Recode"                 "Recycle"               
#> [364] "RelRisk"                "Rename"                 "Rev"                   
#> [367] "RevCode"                "RgbToCmy"               "RgbToCol"              
#> [370] "RgbToHex"               "RgbToLong"              "RndPairs"              
#> [373] "RndWord"                "RobScale"               "RomanToInt"            
#> [376] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [379] "RunsTest"               "SD"                     "SLN"                   
#> [382] "SMAPE"                  "SYD"                    "Sample"                
#> [385] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [388] "SecToHms"               "Second"                 "SendOutlookMail"       
#> [391] "Sens"                   "SetAlpha"               "SetAttr"               
#> [394] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [397] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [400] "Skew"                   "Small"                  "SmoothSpline"          
#> [403] "Some"                   "SomersDelta"            "Sort"                  
#> [406] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [409] "SphToCart"              "SplitAt"                "SplitPath"             
#> [412] "SplitToCol"             "SplitToDummy"           "SpreadOut"             
#> [415] "Stamp"                  "StdCoef"                "Str"                   
#> [418] "StrAbbr"                "StrAlign"               "StrCap"                
#> [421] "StrChop"                "StrCountW"              "StrDist"               
#> [424] "StrExtract"             "StrExtractBetween"      "StrIsNumeric"          
#> [427] "StrLeft"                "StrPad"                 "StrPos"                
#> [430] "StrRev"                 "StrRight"               "StrSpell"              
#> [433] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [436] "StrVal"                 "Strata"                 "StripAttr"             
#> [439] "StuartMaxwellTest"      "StuartTauC"             "SysInfo"               
#> [442] "TMod"                   "TOne"                   "TTestA"                
#> [445] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [448] "Timezone"               "TitleRect"              "ToLong"                
#> [451] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [454] "ToWrdPlot"              "ToXL"                   "Today"                 
#> [457] "Trim"                   "TschuprowT"             "TukeyBiweight"         
#> [460] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [463] "Unit"                   "Unit<-"                 "Untable"               
#> [466] "Unwhich"                "VIF"                    "VanWaerdenTest"        
#> [469] "Var"                    "VarCI"                  "VarTest"               
#> [472] "VecRot"                 "VecShift"               "Vigenere"              
#> [475] "VonNeumannTest"         "Week"                   "Weekday"               
#> [478] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [481] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [484] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [487] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [490] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [493] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [496] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [499] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [502] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [505] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [508] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [511] "YTM"                    "Year"                   "YearDay"               
#> [514] "YearDays"               "YearMonth"              "YuenTTest"             
#> [517] "YuleQ"                  "YuleY"                  "ZTest"                 
#> [520] "ZeroIfNA"               "Zodiac"                 "aovlDetails"           
#> [523] "aovlErrorTerms"         "as.CDateFmt"            "as.fmt"                
#> [526] "as.matrix.xtabs"        "axTicks.Date"           "axTicks.POSIXct"       
#> [529] "createCOMReference"     "dBenf"                  "dExtrVal"              
#> [532] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [535] "dGompertz"              "dGumbel"                "dNegWeibull"           
#> [538] "dOrder"                 "dRevGumbel"             "dRevWeibull"           
#> [541] "identify.formula"       "lines.Lc"               "lines.SmoothSpline"    
#> [544] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [547] "pBenf"                  "pExtrVal"               "pFrechet"              
#> [550] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [553] "pGumbel"                "pNegWeibull"            "pOrder"                
#> [556] "pRevGumbel"             "pRevWeibull"            "plot.Conf"             
#> [559] "plot.Lc"                "plot.Lclist"            "plot.PostHocTest"      
#> [562] "plot.bagplot"           "plot.palette"           "power.chisq.test"      
#> [565] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [568] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [571] "print.ICC"              "print.PercTable"        "print.PostHocTest"     
#> [574] "print.mtest"            "qBenf"                  "qExtrVal"              
#> [577] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [580] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [583] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [586] "rBenf"                  "rExtrVal"               "rFrechet"              
#> [589] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [592] "rGumbel"                "rNegWeibull"            "rOrder"                
#> [595] "rRevGumbel"             "rRevWeibull"            "rSum21"                
#> [598] "reorder.factor"         "split.formula"