List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [25] "AllIdentical"           "AndersonDarlingTest"    "Append"                
#>  [28] "Arrow"                  "AscToChar"              "Asp"                   
#>  [31] "Assocs"                 "Atkinson"               "AxisBreak"             
#>  [34] "BarText"                "BarnardTest"            "BartelsRankTest"       
#>  [37] "Bg"                     "BhapkarTest"            "BinToDec"              
#>  [40] "BinTree"                "BinomCI"                "BinomCIn"              
#>  [43] "BinomDiffCI"            "BinomRatioCI"           "BlankIfNA"             
#>  [46] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [49] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [52] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [55] "CCC"                    "Canvas"                 "CartToPol"             
#>  [58] "CartToSph"              "CatTable"               "CharToAsc"             
#>  [61] "Clockwise"              "Closest"                "CmToPts"               
#>  [64] "CmyToCmyk"              "CmykToCmy"              "CmykToRgb"             
#>  [67] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [70] "CoefVar"                "CoefVarCI"              "CohenD"                
#>  [73] "CohenKappa"             "ColToGray"              "ColToGrey"             
#>  [76] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [79] "ColToRgb"               "CollapseTable"          "ColorLegend"           
#>  [82] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [85] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [88] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [91] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [94] "Cor"                    "CorCI"                  "CorPart"               
#>  [97] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [100] "CourseData"             "Cov"                    "CramerV"               
#> [103] "CramerVonMisesTest"     "CronbachAlpha"          "Cross"                 
#> [106] "CrossN"                 "Cstat"                  "CutQ"                  
#> [109] "DB"                     "Day"                    "Day<-"                 
#> [112] "DecToBin"               "DecToHex"               "DecToOct"              
#> [115] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [118] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [121] "DivCoefMax"             "Divisors"               "DoBy"                  
#> [124] "DoCall"                 "Dot"                    "DrawArc"               
#> [127] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [130] "DrawEllipse"            "DrawRegPolygon"         "Dummy"                 
#> [133] "DunnTest"               "DunnettTest"            "DurbinWatsonTest"      
#> [136] "EX"                     "Entropy"                "Eps"                   
#> [139] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [142] "Factorize"              "FctArgs"                "Fibonacci"             
#> [145] "FindColor"              "FindCorr"               "FindRProfile"          
#> [148] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [151] "Flags"                  "Fmt"                    "Format"                
#> [154] "Frac"                   "Freq"                   "Freq2D"                
#> [157] "GCD"                    "GTest"                  "GenRandGroups"         
#> [160] "GeomSn"                 "GeomTrans"              "GetCalls"              
#> [163] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [166] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [169] "Gini"                   "GiniDeltas"             "GiniSimpson"           
#> [172] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [175] "Gsd"                    "Herfindahl"             "HexToCol"              
#> [178] "HexToDec"               "HexToRgb"               "HighLow"               
#> [181] "Hmean"                  "HmsToMinute"            "HmsToSec"              
#> [184] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [187] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [190] "HunterGaston"           "ICC"                    "IPMT"                  
#> [193] "IQRw"                   "IRR"                    "IdentifyA"             
#> [196] "Impute"                 "ImputeKnn"              "InDots"                
#> [199] "Interval"               "IsDate"                 "IsDichotomous"         
#> [202] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [205] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [208] "IsWeekend"              "IsWhole"                "IsZero"                
#> [211] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [214] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [217] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [220] "LCM"                    "LOCF"                   "LOF"                   
#> [223] "Label"                  "Label<-"                "Labels"                
#> [226] "Labels<-"               "Lambda"                 "Large"                 
#> [229] "LastDayOfMonth"         "Lc"                     "LehmacherTest"         
#> [232] "LeveneTest"             "LillieTest"             "LinScale"              
#> [235] "LineToUser"             "LogSt"                  "LogStInv"              
#> [238] "Logit"                  "LogitInv"               "LongToRgb"             
#> [241] "LsFct"                  "LsObj"                  "MAD"                   
#> [244] "MADCI"                  "MAE"                    "MAPE"                  
#> [247] "MHChisqTest"            "MSE"                    "Mar"                   
#> [250] "Margins"                "MaxDigits"              "Mean"                  
#> [253] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [256] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [259] "MedianCI"               "Mgp"                    "Mgsub"                 
#> [262] "Midx"                   "Minute"                 "MixColor"              
#> [265] "ModSummary"             "Mode"                   "Month"                 
#> [268] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [271] "MultMerge"              "MultinomCI"             "MutInf"                
#> [274] "N"                      "NAIf"                   "NAIfBlank"             
#> [277] "NAIfZero"               "NMAE"                   "NMSE"                  
#> [280] "NPV"                    "NPVFixBond"             "Ndec"                  
#> [283] "NemenyiTest"            "Nf"                     "Now"                   
#> [286] "OPR"                    "ORToRelRisk"            "OctToDec"              
#> [289] "OddsRatio"              "OrderMixed"             "Outlier"               
#> [292] "Overlap"                "PDFManual"              "PMT"                   
#> [295] "PPMT"                   "PageTest"               "PairApply"             
#> [298] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [301] "PartialSD"              "PasswordDlg"            "PearsonTest"           
#> [304] "PercTable"              "PercentRank"            "Permn"                 
#> [307] "Phi"                    "Phrase"                 "PlotACF"               
#> [310] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [313] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [316] "PlotCashFlow"           "PlotCirc"               "PlotConDens"           
#> [319] "PlotCorr"               "PlotDot"                "PlotDotCI"             
#> [322] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [325] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [328] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [331] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [334] "PlotPairs"              "PlotPolar"              "PlotProbDist"          
#> [337] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [340] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [343] "PlotWeb"                "PoissonCI"              "PolToCart"             
#> [346] "PolarGrid"              "PostHocTest"            "PpAddSlide"            
#> [349] "PpPlot"                 "PpText"                 "Prec"                  
#> [352] "Primes"                 "PseudoR2"               "PtInPoly"              
#> [355] "PtsToCm"                "Quantile"               "QuantileCI"            
#> [358] "Quarter"                "Quot"                   "RBAL"                  
#> [361] "RMSE"                   "RSessionAlive"          "RTempdirAlive"         
#> [364] "RadToDeg"               "Range"                  "Rank"                  
#> [367] "Recode"                 "Recycle"                "RelRisk"               
#> [370] "Rename"                 "Rev"                    "RevCode"               
#> [373] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [376] "RgbToLong"              "RndPairs"               "RndWord"               
#> [379] "RobScale"               "RomanToInt"             "Rosenbluth"            
#> [382] "Rotate"                 "RoundTo"                "RunsTest"              
#> [385] "SD"                     "SDN"                    "SLN"                   
#> [388] "SMAPE"                  "SYD"                    "Sample"                
#> [391] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [394] "SecToHms"               "Second"                 "SendOutlookMail"       
#> [397] "Sens"                   "SetAlpha"               "SetAttr"               
#> [400] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [403] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [406] "Skew"                   "Small"                  "SmoothSpline"          
#> [409] "Some"                   "SomersDelta"            "Sort"                  
#> [412] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [415] "SphToCart"              "SplitAt"                "SplitPath"             
#> [418] "SplitToCol"             "SplitToDummy"           "SpreadOut"             
#> [421] "Stamp"                  "StdCoef"                "Str"                   
#> [424] "StrAbbr"                "StrAlign"               "StrCap"                
#> [427] "StrChop"                "StrCountW"              "StrDist"               
#> [430] "StrExtract"             "StrExtractBetween"      "StrIsNumeric"          
#> [433] "StrLeft"                "StrPad"                 "StrPos"                
#> [436] "StrRev"                 "StrRight"               "StrSpell"              
#> [439] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [442] "StrVal"                 "Strata"                 "StripAttr"             
#> [445] "StuartMaxwellTest"      "StuartTauC"             "SysInfo"               
#> [448] "TMod"                   "TOne"                   "TTestA"                
#> [451] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [454] "Timezone"               "TitleRect"              "ToLong"                
#> [457] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [460] "ToWrdPlot"              "ToXL"                   "Today"                 
#> [463] "Trim"                   "TschuprowT"             "TukeyBiweight"         
#> [466] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [469] "Unit"                   "Unit<-"                 "Untable"               
#> [472] "Unwhich"                "VIF"                    "VanWaerdenTest"        
#> [475] "Var"                    "VarCI"                  "VarN"                  
#> [478] "VarTest"                "VarX"                   "VecRot"                
#> [481] "VecShift"               "Vigenere"               "VonNeumannTest"        
#> [484] "Week"                   "Weekday"                "Winsorize"             
#> [487] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [490] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [493] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [496] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [499] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [502] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [505] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [508] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [511] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [514] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [517] "XLSaveAs"               "XLView"                 "YTM"                   
#> [520] "Year"                   "YearDay"                "YearDays"              
#> [523] "YearMonth"              "YuenTTest"              "YuleQ"                 
#> [526] "YuleY"                  "ZTest"                  "ZeroIfNA"              
#> [529] "Zodiac"                 "aovlDetails"            "aovlErrorTerms"        
#> [532] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#> [535] "axTicks.Date"           "axTicks.POSIXct"        "createCOMReference"    
#> [538] "dBenf"                  "dExtrVal"               "dFrechet"              
#> [541] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [544] "dGumbel"                "dNegWeibull"            "dOrder"                
#> [547] "dRevGumbel"             "dRevWeibull"            "identify.formula"      
#> [550] "lines.Lc"               "lines.SmoothSpline"     "lines.lm"              
#> [553] "lines.loess"            "lines.smooth.spline"    "pBenf"                 
#> [556] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [559] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [562] "pNegWeibull"            "pOrder"                 "pRevGumbel"            
#> [565] "pRevWeibull"            "plot.Conf"              "plot.Lc"               
#> [568] "plot.Lclist"            "plot.PostHocTest"       "plot.bagplot"          
#> [571] "plot.palette"           "power.chisq.test"       "print.Assocs"          
#> [574] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [577] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [580] "print.PercTable"        "print.PostHocTest"      "print.mtest"           
#> [583] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [586] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [589] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [592] "qRevGumbelExp"          "qRevWeibull"            "rBenf"                 
#> [595] "rExtrVal"               "rFrechet"               "rGenExtrVal"           
#> [598] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [601] "rNegWeibull"            "rOrder"                 "rRevGumbel"            
#> [604] "rRevWeibull"            "rSum21"                 "reorder.factor"        
#> [607] "split.formula"