List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%:%"                    "%()%"                   "%(]%"                  
#>   [4] "%)(%"                   "%)[%"                   "%[)%"                  
#>   [7] "%[]%"                   "%](%"                   "%][%"                  
#>  [10] "%^%"                    "%c%"                    "%like any%"            
#>  [13] "%like%"                 "%nin%"                  "%overlaps%"            
#>  [16] "Abind"                  "Abstract"               "AddMonths"             
#>  [19] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [22] "AllIdentical"           "AndersonDarlingTest"    "aovlDetails"           
#>  [25] "aovlErrorTerms"         "Append"                 "Arrow"                 
#>  [28] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#>  [31] "AscToChar"              "Asp"                    "Assocs"                
#>  [34] "Atkinson"               "AUC"                    "AxisBreak"             
#>  [37] "axTicks.Date"           "axTicks.POSIXct"        "BarnardTest"           
#>  [40] "BartelsRankTest"        "BarText"                "BhapkarTest"           
#>  [43] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [46] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [49] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [52] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [55] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [58] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [61] "CartToSph"              "CatTable"               "CCC"                   
#>  [64] "CharToAsc"              "Clockwise"              "Closest"               
#>  [67] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [70] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [73] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [76] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [79] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [82] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [85] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [88] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [91] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [94] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [97] "Cor"                    "CorCI"                  "CorPart"               
#> [100] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [103] "CourseData"             "Cov"                    "CramerV"               
#> [106] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [109] "Cross"                  "CrossN"                 "Cstat"                 
#> [112] "CutQ"                   "Day"                    "Day<-"                 
#> [115] "DB"                     "dBenf"                  "DecToBin"              
#> [118] "DecToHex"               "DecToOct"               "DegToRad"              
#> [121] "Desc"                   "DescToolsOptions"       "dExtrVal"              
#> [124] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [127] "dGompertz"              "dGumbel"                "DiffDays360"           
#> [130] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [133] "Divisors"               "dNegWeibull"            "DoBy"                  
#> [136] "DoCall"                 "dOrder"                 "Dot"                   
#> [139] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [142] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [145] "dRevGumbel"             "dRevWeibull"            "Dummy"                 
#> [148] "DunnettTest"            "DunnTest"               "DurbinWatsonTest"      
#> [151] "Entropy"                "Eps"                    "ErrBars"               
#> [154] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [157] "FctArgs"                "Fibonacci"              "FindColor"             
#> [160] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [163] "FisherZInv"             "FixToTable"             "Flags"                 
#> [166] "Fmt"                    "Format"                 "Frac"                  
#> [169] "Freq"                   "Freq2D"                 "GCD"                   
#> [172] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [175] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [178] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [181] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [184] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [187] "Gsd"                    "GTest"                  "Herfindahl"            
#> [190] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [193] "HighLow"                "Hmean"                  "HmsToSec"              
#> [196] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [199] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [202] "HunterGaston"           "ICC"                    "identify.formula"      
#> [205] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [208] "InDots"                 "Interval"               "IPMT"                  
#> [211] "IQRw"                   "IRR"                    "IsDate"                
#> [214] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [217] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [220] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [223] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [226] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [229] "KendallW"               "Keywords"               "KrippAlpha"            
#> [232] "Kurt"                   "Label"                  "Label<-"               
#> [235] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [238] "Lc"                     "LCM"                    "LehmacherTest"         
#> [241] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [244] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [247] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [250] "LOCF"                   "LOF"                    "Logit"                 
#> [253] "LogitInv"               "LogSt"                  "LogStInv"              
#> [256] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [259] "MAD"                    "MADCI"                  "MAE"                   
#> [262] "MAPE"                   "Mar"                    "Margins"               
#> [265] "MaxDigits"              "Mean"                   "MeanAD"                
#> [268] "MeanCI"                 "MeanCIn"                "MeanDiffCI"            
#> [271] "MeanSE"                 "Median"                 "MedianCI"              
#> [274] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [277] "Minute"                 "MixColor"               "Mode"                  
#> [280] "ModSummary"             "Month"                  "MosesTest"             
#> [283] "MoveAvg"                "MSE"                    "MultinomCI"            
#> [286] "MultMerge"              "MutInf"                 "N"                     
#> [289] "NAIf"                   "NAIfBlank"              "NAIfZero"              
#> [292] "Ndec"                   "NemenyiTest"            "NMAE"                  
#> [295] "NMSE"                   "Now"                    "NPV"                   
#> [298] "NPVFixBond"             "OctToDec"               "OddsRatio"             
#> [301] "OPR"                    "OrderMixed"             "ORToRelRisk"           
#> [304] "Outlier"                "Overlap"                "PageTest"              
#> [307] "PairApply"              "Pal"                    "ParseFormula"          
#> [310] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [313] "pBenf"                  "PDFManual"              "PearsonTest"           
#> [316] "PercentRank"            "PercTable"              "Permn"                 
#> [319] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [322] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [325] "Phi"                    "Phrase"                 "plot.bagplot"          
#> [328] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [331] "plot.palette"           "plot.PostHocTest"       "PlotACF"               
#> [334] "PlotArea"               "PlotBag"                "PlotBagPairs"          
#> [337] "PlotBinTree"            "PlotBubble"             "PlotCandlestick"       
#> [340] "PlotCashFlow"           "PlotCirc"               "PlotConDens"           
#> [343] "PlotCorr"               "PlotDot"                "PlotECDF"              
#> [346] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [349] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [352] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [355] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [358] "PlotPolar"              "PlotPyramid"            "PlotQQ"                
#> [361] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [364] "PlotViolin"             "PlotWeb"                "PMT"                   
#> [367] "pNegWeibull"            "PoissonCI"              "PolarGrid"             
#> [370] "PolToCart"              "pOrder"                 "PostHocTest"           
#> [373] "power.chisq.test"       "PpAddSlide"             "PPMT"                  
#> [376] "PpPlot"                 "PpText"                 "Prec"                  
#> [379] "pRevGumbel"             "pRevWeibull"            "Primes"                
#> [382] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [385] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [388] "print.ICC"              "print.mtest"            "print.PercTable"       
#> [391] "print.PostHocTest"      "PseudoR2"               "PtInPoly"              
#> [394] "PtsToCm"                "qBenf"                  "qExtrVal"              
#> [397] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [400] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [403] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [406] "Quantile"               "QuantileCI"             "Quarter"               
#> [409] "Quot"                   "RadToDeg"               "Range"                 
#> [412] "Rank"                   "RBAL"                   "rBenf"                 
#> [415] "Recode"                 "Recycle"                "RelRisk"               
#> [418] "Rename"                 "reorder.factor"         "Rev"                   
#> [421] "RevCode"                "rExtrVal"               "rFrechet"              
#> [424] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [427] "RgbToLong"              "rGenExtrVal"            "rGenPareto"            
#> [430] "rGompertz"              "rGumbel"                "RMSE"                  
#> [433] "RndPairs"               "RndWord"                "rNegWeibull"           
#> [436] "RobScale"               "RomanToInt"             "rOrder"                
#> [439] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [442] "rRevGumbel"             "rRevWeibull"            "RSessionAlive"         
#> [445] "RTempdirAlive"          "RunsTest"               "Sample"                
#> [448] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [451] "SD"                     "Second"                 "SecToHms"              
#> [454] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [457] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [460] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [463] "Skew"                   "SLN"                    "Small"                 
#> [466] "SMAPE"                  "SmoothSpline"           "Some"                  
#> [469] "SomersDelta"            "Sort"                   "SortMixed"             
#> [472] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [475] "split.formula"          "SplitAt"                "SplitPath"             
#> [478] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [481] "StdCoef"                "Str"                    "StrAbbr"               
#> [484] "StrAlign"               "Strata"                 "StrCap"                
#> [487] "StrChop"                "StrCountW"              "StrDist"               
#> [490] "StrExtract"             "StrExtractBetween"      "StripAttr"             
#> [493] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [496] "StrPos"                 "StrRev"                 "StrRight"              
#> [499] "StrSpell"               "StrSplit"               "StrTrim"               
#> [502] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [505] "StuartTauC"             "SYD"                    "SysInfo"               
#> [508] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [511] "Timezone"               "TitleRect"              "TMod"                  
#> [514] "Today"                  "ToLong"                 "TOne"                  
#> [517] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [520] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [523] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [526] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [529] "Unit"                   "Unit<-"                 "Untable"               
#> [532] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [535] "VarCI"                  "VarTest"                "VecRot"                
#> [538] "VecShift"               "VIF"                    "Vigenere"              
#> [541] "VonNeumannTest"         "Week"                   "Weekday"               
#> [544] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [547] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [550] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [553] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [556] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [559] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [562] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [565] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [568] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [571] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [574] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [577] "Year"                   "YearDay"                "YearMonth"             
#> [580] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [583] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [586] "ZTest"