List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "Append"                 "AscToChar"             
#>  [28] "Asp"                    "Assocs"                 "Atkinson"              
#>  [31] "AxisBreak"              "BarText"                "BarnardTest"           
#>  [34] "BartelsRankTest"        "Bg"                     "BhapkarTest"           
#>  [37] "BinToDec"               "BinTree"                "BinomCI"               
#>  [40] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [43] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [46] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [49] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [52] "BrierScoreCI"           "BubbleLegend"           "CCC"                   
#>  [55] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [58] "CatTable"               "CharToAsc"              "Clockwise"             
#>  [61] "Closest"                "CmToPts"                "CmyToCmyk"             
#>  [64] "CmykToCmy"              "CmykToRgb"              "Coalesce"              
#>  [67] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [70] "CoefVarCI"              "CohenD"                 "CohenKappa"            
#>  [73] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [76] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [79] "CollapseTable"          "ColorLegend"            "ColumnWrap"            
#>  [82] "CombN"                  "CombPairs"              "CombSet"               
#>  [85] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [88] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [91] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [94] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [97] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [100] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CstatCI"                "CutAge"                
#> [109] "CutQ"                   "DB"                     "Day"                   
#> [112] "Day<-"                  "DecToBin"               "DecToHex"              
#> [115] "DecToOct"               "DegToRad"               "Desc"                  
#> [118] "DescToolsOptions"       "DiffDays360"            "DigitSum"              
#> [121] "DivCoef"                "DivCoefMax"             "Divisors"              
#> [124] "DoBy"                   "DoCall"                 "Dot"                   
#> [127] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [130] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [133] "Dummy"                  "DunnTest"               "DunnettTest"           
#> [136] "DurbinWatsonTest"       "EX"                     "Entropy"               
#> [139] "Eps"                    "ErrBars"                "EtaSq"                 
#> [142] "ExpFreq"                "Factorize"              "Fade"                  
#> [145] "FctArgs"                "Fibonacci"              "FindColor"             
#> [148] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [151] "FisherZInv"             "FixToTable"             "Flags"                 
#> [154] "Fmt"                    "Format"                 "Frac"                  
#> [157] "Freq"                   "Freq2D"                 "GCD"                   
#> [160] "GTest"                  "GenRandGroups"          "GeomSn"                
#> [163] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [166] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [169] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [172] "GiniDeltas"             "GiniSimpson"            "Gmean"                 
#> [175] "GoodmanKruskalGamma"    "GoodmanKruskalTau"      "Gsd"                   
#> [178] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [181] "HexToRgb"               "HighLow"                "Hmean"                 
#> [184] "HmsToMinute"            "HmsToSec"               "HodgesLehmann"         
#> [187] "HoeffD"                 "HosmerLemeshowTest"     "HotellingsT2Test"      
#> [190] "Hour"                   "HuberM"                 "HunterGaston"          
#> [193] "ICC"                    "IPMT"                   "IQRw"                  
#> [196] "IRR"                    "IdentifyA"              "Impute"                
#> [199] "ImputeKnn"              "InDots"                 "Interval"              
#> [202] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [205] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [208] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [211] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [214] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [217] "KendallTauB"            "KendallW"               "Keywords"              
#> [220] "KrippAlpha"             "Kurt"                   "LCM"                   
#> [223] "LOCF"                   "LOF"                    "Label"                 
#> [226] "Label<-"                "Labels"                 "Labels<-"              
#> [229] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [232] "Lc"                     "LehmacherTest"          "LeveneTest"            
#> [235] "LillieTest"             "LinScale"               "LineToUser"            
#> [238] "LogSt"                  "LogStInv"               "Logit"                 
#> [241] "LogitInv"               "LongToRgb"              "LsFct"                 
#> [244] "LsObj"                  "MAD"                    "MADCI"                 
#> [247] "MAE"                    "MAPE"                   "MHChisqTest"           
#> [250] "MSE"                    "Mar"                    "Margins"               
#> [253] "MaxDigits"              "Mean"                   "MeanAD"                
#> [256] "MeanCI"                 "MeanCIn"                "MeanDiffCI"            
#> [259] "MeanSE"                 "Median"                 "MedianCI"              
#> [262] "Mgp"                    "Mgsub"                  "Midx"                  
#> [265] "Minute"                 "MixColor"               "ModSummary"            
#> [268] "Mode"                   "Month"                  "MonthDays"             
#> [271] "MosesTest"              "MoveAvg"                "MultMerge"             
#> [274] "MultinomCI"             "MutInf"                 "N"                     
#> [277] "NAIf"                   "NAIfBlank"              "NAIfZero"              
#> [280] "NALevel"                "NMAE"                   "NMSE"                  
#> [283] "NPV"                    "NPVFixBond"             "NZ"                    
#> [286] "Ndec"                   "NemenyiTest"            "Nf"                    
#> [289] "Now"                    "OPR"                    "ORToRelRisk"           
#> [292] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [295] "Outlier"                "Overlap"                "PDFManual"             
#> [298] "PMT"                    "PPMT"                   "PageTest"              
#> [301] "PairApply"              "Pal"                    "ParseFormula"          
#> [304] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [307] "PearsonTest"            "PercTable"              "PercentRank"           
#> [310] "Permn"                  "Phi"                    "Phrase"                
#> [313] "PlotACF"                "PlotArea"               "PlotBag"               
#> [316] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [319] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [322] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [325] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [328] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [331] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [334] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [337] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [340] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [343] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [346] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [349] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [352] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [355] "Prec"                   "Primes"                 "PseudoR2"              
#> [358] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [361] "QuantileCI"             "Quarter"                "Quot"                  
#> [364] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [367] "RSqCI"                  "RTempdirAlive"          "RadToDeg"              
#> [370] "Range"                  "Rank"                   "Recode"                
#> [373] "Recycle"                "RelRisk"                "Rename"                
#> [376] "Rev"                    "RevCode"                "RgbToCmy"              
#> [379] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [382] "RndPairs"               "RndWord"                "RobScale"              
#> [385] "RomanToInt"             "Rosenbluth"             "Rotate"                
#> [388] "RoundTo"                "RunsTest"               "SD"                    
#> [391] "SDN"                    "SLN"                    "SMAPE"                 
#> [394] "SYD"                    "Sample"                 "SampleTwins"           
#> [397] "SaveAs"                 "ScheffeTest"            "SecToHms"              
#> [400] "Second"                 "SendOutlookMail"        "Sens"                  
#> [403] "SetAlpha"               "SetAttr"                "SetNames"              
#> [406] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [409] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [412] "Small"                  "SmoothSpline"           "Some"                  
#> [415] "SomersDelta"            "Sort"                   "SortMixed"             
#> [418] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [421] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [424] "SplitToDummy"           "SpreadOut"              "Stamp"                 
#> [427] "StdCoef"                "Str"                    "StrAbbr"               
#> [430] "StrAlign"               "StrCap"                 "StrChop"               
#> [433] "StrCountW"              "StrDist"                "StrExtract"            
#> [436] "StrExtractBetween"      "StrIsNumeric"           "StrLeft"               
#> [439] "StrPad"                 "StrPos"                 "StrRev"                
#> [442] "StrRight"               "StrSpell"               "StrSplit"              
#> [445] "StrTrim"                "StrTrunc"               "StrVal"                
#> [448] "Strata"                 "StripAttr"              "StuartMaxwellTest"     
#> [451] "StuartTauC"             "SysInfo"                "TMod"                  
#> [454] "TOne"                   "TTestA"                 "TextContrastColor"     
#> [457] "TextToTable"            "TheilU"                 "Timezone"              
#> [460] "TitleRect"              "ToLong"                 "ToWide"                
#> [463] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [466] "ToXL"                   "Today"                  "Trim"                  
#> [469] "TschuprowT"             "TukeyBiweight"          "TwoGroups"             
#> [472] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [475] "Unit<-"                 "Untable"                "Unwhich"               
#> [478] "VIF"                    "VanWaerdenTest"         "Var"                   
#> [481] "VarCI"                  "VarN"                   "VarTest"               
#> [484] "VarX"                   "VecRot"                 "VecShift"              
#> [487] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [490] "Weekday"                "Winsorize"              "WithOptions"           
#> [493] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [496] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [499] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [502] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [505] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [508] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [511] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [514] "WrdTableBorders"        "WrdTableHeading"        "WrdUpdateBookmark"     
#> [517] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [520] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [523] "XLSaveAs"               "XLView"                 "YTM"                   
#> [526] "Year"                   "YearDay"                "YearDays"              
#> [529] "YearMonth"              "YuenTTest"              "YuleQ"                 
#> [532] "YuleY"                  "ZTest"                  "ZeroIfNA"              
#> [535] "Zodiac"                 "aovlDetails"            "aovlErrorTerms"        
#> [538] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#> [541] "as.ym"                  "axTicks.Date"           "axTicks.POSIXct"       
#> [544] "createCOMReference"     "dBenf"                  "dExtrVal"              
#> [547] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [550] "dGompertz"              "dGumbel"                "dNegWeibull"           
#> [553] "dOrder"                 "dRevGumbel"             "dRevWeibull"           
#> [556] "dTri"                   "identify.formula"       "lines.Lc"              
#> [559] "lines.SmoothSpline"     "lines.lm"               "lines.loess"           
#> [562] "lines.smooth.spline"    "pBenf"                  "pExtrVal"              
#> [565] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [568] "pGompertz"              "pGumbel"                "pNegWeibull"           
#> [571] "pOrder"                 "pRevGumbel"             "pRevWeibull"           
#> [574] "pTri"                   "plot.Conf"              "plot.Lc"               
#> [577] "plot.Lclist"            "plot.PostHocTest"       "plot.bagplot"          
#> [580] "plot.palette"           "power.chisq.test"       "print.Assocs"          
#> [583] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [586] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [589] "print.PercTable"        "print.PostHocTest"      "print.mtest"           
#> [592] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [595] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [598] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [601] "qRevGumbelExp"          "qRevWeibull"            "qTri"                  
#> [604] "rBenf"                  "rExtrVal"               "rFrechet"              
#> [607] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [610] "rGumbel"                "rNegWeibull"            "rOrder"                
#> [613] "rRevGumbel"             "rRevWeibull"            "rSum21"                
#> [616] "rTri"                   "reorder.factor"         "split.formula"