List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "Append"                 "Arrow"                 
#>  [28] "AscToChar"              "Asp"                    "Assocs"                
#>  [31] "Atkinson"               "AxisBreak"              "BarText"               
#>  [34] "BarnardTest"            "BartelsRankTest"        "Bg"                    
#>  [37] "BhapkarTest"            "BinToDec"               "BinTree"               
#>  [40] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [43] "BinomRatioCI"           "BlankIfNA"              "BootCI"                
#>  [46] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [49] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [52] "BrierScore"             "BubbleLegend"           "CCC"                   
#>  [55] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [58] "CatTable"               "CharToAsc"              "Clockwise"             
#>  [61] "Closest"                "CmToPts"                "CmyToCmyk"             
#>  [64] "CmykToCmy"              "CmykToRgb"              "Coalesce"              
#>  [67] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [70] "CoefVarCI"              "CohenD"                 "CohenKappa"            
#>  [73] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [76] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [79] "CollapseTable"          "ColorLegend"            "ColumnWrap"            
#>  [82] "CombN"                  "CombPairs"              "CombSet"               
#>  [85] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [88] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [91] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [94] "CorCI"                  "CorPart"                "CorPolychor"           
#>  [97] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [100] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [103] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [106] "Cstat"                  "CutAge"                 "CutQ"                  
#> [109] "DB"                     "Day"                    "Day<-"                 
#> [112] "DecToBin"               "DecToHex"               "DecToOct"              
#> [115] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [118] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [121] "DivCoefMax"             "Divisors"               "DoBy"                  
#> [124] "DoCall"                 "Dot"                    "DrawArc"               
#> [127] "DrawBand"               "DrawBezier"             "DrawCircle"            
#> [130] "DrawEllipse"            "DrawRegPolygon"         "Dummy"                 
#> [133] "DunnTest"               "DunnettTest"            "DurbinWatsonTest"      
#> [136] "EX"                     "Entropy"                "Eps"                   
#> [139] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [142] "Factorize"              "Fade"                   "FctArgs"               
#> [145] "Fibonacci"              "FindColor"              "FindCorr"              
#> [148] "FindRProfile"           "FisherZ"                "FisherZInv"            
#> [151] "FixToTable"             "Flags"                  "Fmt"                   
#> [154] "Format"                 "Frac"                   "Freq"                  
#> [157] "Freq2D"                 "GCD"                    "GTest"                 
#> [160] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [163] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [166] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [169] "GetNewXL"               "Gini"                   "GiniDeltas"            
#> [172] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [175] "GoodmanKruskalTau"      "Gsd"                    "Herfindahl"            
#> [178] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [181] "HighLow"                "Hmean"                  "HmsToMinute"           
#> [184] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [187] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [190] "HuberM"                 "HunterGaston"           "ICC"                   
#> [193] "IPMT"                   "IQRw"                   "IRR"                   
#> [196] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [199] "InDots"                 "Interval"               "IsDate"                
#> [202] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [205] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [208] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [211] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [214] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [217] "KendallW"               "Keywords"               "KrippAlpha"            
#> [220] "Kurt"                   "LCM"                    "LOCF"                  
#> [223] "LOF"                    "Label"                  "Label<-"               
#> [226] "Labels"                 "Labels<-"               "Lambda"                
#> [229] "Large"                  "LastDayOfMonth"         "Lc"                    
#> [232] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [235] "LinScale"               "LineToUser"             "LogSt"                 
#> [238] "LogStInv"               "Logit"                  "LogitInv"              
#> [241] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [244] "MAD"                    "MADCI"                  "MAE"                   
#> [247] "MAPE"                   "MHChisqTest"            "MSE"                   
#> [250] "Mar"                    "Margins"                "MaxDigits"             
#> [253] "Mean"                   "MeanAD"                 "MeanCI"                
#> [256] "MeanCIn"                "MeanDiffCI"             "MeanSE"                
#> [259] "Median"                 "MedianCI"               "Mgp"                   
#> [262] "Mgsub"                  "Midx"                   "Minute"                
#> [265] "MixColor"               "ModSummary"             "Mode"                  
#> [268] "Month"                  "MonthDays"              "MosesTest"             
#> [271] "MoveAvg"                "MultMerge"              "MultinomCI"            
#> [274] "MutInf"                 "N"                      "NAIf"                  
#> [277] "NAIfBlank"              "NAIfZero"               "NALevel"               
#> [280] "NMAE"                   "NMSE"                   "NPV"                   
#> [283] "NPVFixBond"             "Ndec"                   "NemenyiTest"           
#> [286] "Nf"                     "Now"                    "OPR"                   
#> [289] "ORToRelRisk"            "OctToDec"               "OddsRatio"             
#> [292] "OrderMixed"             "Outlier"                "Overlap"               
#> [295] "PDFManual"              "PMT"                    "PPMT"                  
#> [298] "PageTest"               "PairApply"              "Pal"                   
#> [301] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [304] "PasswordDlg"            "PearsonTest"            "PercTable"             
#> [307] "PercentRank"            "Permn"                  "Phi"                   
#> [310] "Phrase"                 "PlotACF"                "PlotArea"              
#> [313] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [316] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [319] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [322] "PlotDot"                "PlotDotCI"              "PlotECDF"              
#> [325] "PlotFaces"              "PlotFdist"              "PlotFun"               
#> [328] "PlotGACF"               "PlotLinesA"             "PlotLog"               
#> [331] "PlotMarDens"            "PlotMiss"               "PlotMonth"             
#> [334] "PlotMosaic"             "PlotMultiDens"          "PlotPairs"             
#> [337] "PlotPolar"              "PlotProbDist"           "PlotPyramid"           
#> [340] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [343] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [346] "PoissonCI"              "PolToCart"              "PolarGrid"             
#> [349] "PostHocTest"            "PpAddSlide"             "PpPlot"                
#> [352] "PpText"                 "Prec"                   "Primes"                
#> [355] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [358] "Quantile"               "QuantileCI"             "Quarter"               
#> [361] "Quot"                   "RBAL"                   "RMSE"                  
#> [364] "RSessionAlive"          "RTempdirAlive"          "RadToDeg"              
#> [367] "Range"                  "Rank"                   "Recode"                
#> [370] "Recycle"                "RelRisk"                "Rename"                
#> [373] "Rev"                    "RevCode"                "RgbToCmy"              
#> [376] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [379] "RndPairs"               "RndWord"                "RobScale"              
#> [382] "RomanToInt"             "Rosenbluth"             "Rotate"                
#> [385] "RoundTo"                "RunsTest"               "SD"                    
#> [388] "SDN"                    "SLN"                    "SMAPE"                 
#> [391] "SYD"                    "Sample"                 "SampleTwins"           
#> [394] "SaveAs"                 "ScheffeTest"            "SecToHms"              
#> [397] "Second"                 "SendOutlookMail"        "Sens"                  
#> [400] "SetAlpha"               "SetAttr"                "SetNames"              
#> [403] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [406] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [409] "Small"                  "SmoothSpline"           "Some"                  
#> [412] "SomersDelta"            "Sort"                   "SortMixed"             
#> [415] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [418] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [421] "SplitToDummy"           "SpreadOut"              "Stamp"                 
#> [424] "StdCoef"                "Str"                    "StrAbbr"               
#> [427] "StrAlign"               "StrCap"                 "StrChop"               
#> [430] "StrCountW"              "StrDist"                "StrExtract"            
#> [433] "StrExtractBetween"      "StrIsNumeric"           "StrLeft"               
#> [436] "StrPad"                 "StrPos"                 "StrRev"                
#> [439] "StrRight"               "StrSpell"               "StrSplit"              
#> [442] "StrTrim"                "StrTrunc"               "StrVal"                
#> [445] "Strata"                 "StripAttr"              "StuartMaxwellTest"     
#> [448] "StuartTauC"             "SysInfo"                "TMod"                  
#> [451] "TOne"                   "TTestA"                 "TextContrastColor"     
#> [454] "TextToTable"            "TheilU"                 "Timezone"              
#> [457] "TitleRect"              "ToLong"                 "ToWide"                
#> [460] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [463] "ToXL"                   "Today"                  "Trim"                  
#> [466] "TschuprowT"             "TukeyBiweight"          "TwoGroups"             
#> [469] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [472] "Unit<-"                 "Untable"                "Unwhich"               
#> [475] "VIF"                    "VanWaerdenTest"         "Var"                   
#> [478] "VarCI"                  "VarN"                   "VarTest"               
#> [481] "VarX"                   "VecRot"                 "VecShift"              
#> [484] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [487] "Weekday"                "Winsorize"              "WithOptions"           
#> [490] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [493] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [496] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [499] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [502] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [505] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [508] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [511] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [514] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [517] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [520] "XLView"                 "YTM"                    "Year"                  
#> [523] "YearDay"                "YearDays"               "YearMonth"             
#> [526] "YuenTTest"              "YuleQ"                  "YuleY"                 
#> [529] "ZTest"                  "ZeroIfNA"               "Zodiac"                
#> [532] "aovlDetails"            "aovlErrorTerms"         "as.CDateFmt"           
#> [535] "as.fmt"                 "as.matrix.xtabs"        "as.ym"                 
#> [538] "axTicks.Date"           "axTicks.POSIXct"        "createCOMReference"    
#> [541] "dBenf"                  "dExtrVal"               "dFrechet"              
#> [544] "dGenExtrVal"            "dGenPareto"             "dGompertz"             
#> [547] "dGumbel"                "dNegWeibull"            "dOrder"                
#> [550] "dRevGumbel"             "dRevWeibull"            "dTri"                  
#> [553] "identify.formula"       "lines.Lc"               "lines.SmoothSpline"    
#> [556] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [559] "pBenf"                  "pExtrVal"               "pFrechet"              
#> [562] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [565] "pGumbel"                "pNegWeibull"            "pOrder"                
#> [568] "pRevGumbel"             "pRevWeibull"            "pTri"                  
#> [571] "plot.Conf"              "plot.Lc"                "plot.Lclist"           
#> [574] "plot.PostHocTest"       "plot.bagplot"           "plot.palette"          
#> [577] "power.chisq.test"       "print.Assocs"           "print.Conf"            
#> [580] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [583] "print.HoeffD"           "print.ICC"              "print.PercTable"       
#> [586] "print.PostHocTest"      "print.mtest"            "qBenf"                 
#> [589] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [592] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [595] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [598] "qRevWeibull"            "qTri"                   "rBenf"                 
#> [601] "rExtrVal"               "rFrechet"               "rGenExtrVal"           
#> [604] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [607] "rNegWeibull"            "rOrder"                 "rRevGumbel"            
#> [610] "rRevWeibull"            "rSum21"                 "rTri"                  
#> [613] "reorder.factor"         "split.formula"