List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package

the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")
#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%:%"                    "%::%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "AUC"                   
#>  [19] "Abind"                  "Abstract"               "AddMonths"             
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "Append"                 "AscToChar"             
#>  [28] "Asp"                    "Assocs"                 "Atkinson"              
#>  [31] "AxisBreak"              "BarText"                "BarnardTest"           
#>  [34] "BartelsRankTest"        "Bg"                     "BhapkarTest"           
#>  [37] "BinToDec"               "BinTree"                "BinomCI"               
#>  [40] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [43] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [46] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [49] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [52] "BubbleLegend"           "CCC"                    "Canvas"                
#>  [55] "CartToPol"              "CartToSph"              "CatTable"              
#>  [58] "CharToAsc"              "Clockwise"              "Closest"               
#>  [61] "CmToPts"                "CmyToCmyk"              "CmykToCmy"             
#>  [64] "CmykToRgb"              "Coalesce"               "CochranArmitageTest"   
#>  [67] "CochranQTest"           "CoefVar"                "CoefVarCI"             
#>  [70] "CohenD"                 "CohenKappa"             "ColToGray"             
#>  [73] "ColToGrey"              "ColToHex"               "ColToHsv"              
#>  [76] "ColToOpaque"            "ColToRgb"               "CollapseTable"         
#>  [79] "ColorLegend"            "ColumnWrap"             "CombN"                 
#>  [82] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [85] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [88] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [91] "ConvUnit"               "Cor"                    "CorCI"                 
#>  [94] "CorPart"                "CorPolychor"            "CountCompCases"        
#>  [97] "CountWorkDays"          "CourseData"             "Cov"                   
#> [100] "CramerV"                "CramerVonMisesTest"     "CronbachAlpha"         
#> [103] "Cross"                  "CrossN"                 "Cstat"                 
#> [106] "CutAge"                 "CutQ"                   "DB"                    
#> [109] "Day"                    "Day<-"                  "DecToBin"              
#> [112] "DecToHex"               "DecToOct"               "DegToRad"              
#> [115] "Desc"                   "DescToolsOptions"       "DiffDays360"           
#> [118] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [121] "Divisors"               "DoBy"                   "DoCall"                
#> [124] "Dot"                    "DrawArc"                "DrawBand"              
#> [127] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [130] "DrawRegPolygon"         "Dummy"                  "DunnTest"              
#> [133] "DunnettTest"            "DurbinWatsonTest"       "EX"                    
#> [136] "Entropy"                "Eps"                    "ErrBars"               
#> [139] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [142] "Fade"                   "FctArgs"                "Fibonacci"             
#> [145] "FindColor"              "FindCorr"               "FindRProfile"          
#> [148] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [151] "Flags"                  "Fmt"                    "Format"                
#> [154] "Frac"                   "Freq"                   "Freq2D"                
#> [157] "GCD"                    "GTest"                  "GenRandGroups"         
#> [160] "GeomSn"                 "GeomTrans"              "GetCalls"              
#> [163] "GetCurrPP"              "GetCurrWrd"             "GetCurrXL"             
#> [166] "GetNewPP"               "GetNewWrd"              "GetNewXL"              
#> [169] "Gini"                   "GiniDeltas"             "GiniSimpson"           
#> [172] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [175] "Gsd"                    "Herfindahl"             "HexToCol"              
#> [178] "HexToDec"               "HexToRgb"               "HighLow"               
#> [181] "Hmean"                  "HmsToMinute"            "HmsToSec"              
#> [184] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [187] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [190] "HunterGaston"           "ICC"                    "IPMT"                  
#> [193] "IQRw"                   "IRR"                    "IdentifyA"             
#> [196] "Impute"                 "ImputeKnn"              "InDots"                
#> [199] "Interval"               "IsDate"                 "IsDichotomous"         
#> [202] "IsEuclid"               "IsLeapYear"             "IsNumeric"             
#> [205] "IsOdd"                  "IsPrime"                "IsValidHwnd"           
#> [208] "IsWeekend"              "IsWhole"                "IsZero"                
#> [211] "JarqueBeraTest"         "JonckheereTerpstraTest" "KappaM"                
#> [214] "KendallTauA"            "KendallTauB"            "KendallW"              
#> [217] "Keywords"               "KrippAlpha"             "Kurt"                  
#> [220] "LCM"                    "LOCF"                   "LOF"                   
#> [223] "Label"                  "Label<-"                "Labels"                
#> [226] "Labels<-"               "Lambda"                 "Large"                 
#> [229] "LastDayOfMonth"         "Lc"                     "LehmacherTest"         
#> [232] "LeveneTest"             "LillieTest"             "LinScale"              
#> [235] "LineToUser"             "LogSt"                  "LogStInv"              
#> [238] "Logit"                  "LogitInv"               "LongToRgb"             
#> [241] "LsFct"                  "LsObj"                  "MAD"                   
#> [244] "MADCI"                  "MAE"                    "MAPE"                  
#> [247] "MHChisqTest"            "MSE"                    "Mar"                   
#> [250] "Margins"                "MaxDigits"              "Mean"                  
#> [253] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [256] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [259] "MedianCI"               "Mgp"                    "Mgsub"                 
#> [262] "Midx"                   "Minute"                 "MixColor"              
#> [265] "ModSummary"             "Mode"                   "Month"                 
#> [268] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [271] "MultMerge"              "MultinomCI"             "MutInf"                
#> [274] "N"                      "NAIf"                   "NAIfBlank"             
#> [277] "NAIfZero"               "NALevel"                "NMAE"                  
#> [280] "NMSE"                   "NPV"                    "NPVFixBond"            
#> [283] "Ndec"                   "NemenyiTest"            "Nf"                    
#> [286] "Now"                    "OPR"                    "ORToRelRisk"           
#> [289] "OctToDec"               "OddsRatio"              "OrderMixed"            
#> [292] "Outlier"                "Overlap"                "PDFManual"             
#> [295] "PMT"                    "PPMT"                   "PageTest"              
#> [298] "PairApply"              "Pal"                    "ParseFormula"          
#> [301] "ParseSASDatalines"      "PartialSD"              "PasswordDlg"           
#> [304] "PearsonTest"            "PercTable"              "PercentRank"           
#> [307] "Permn"                  "Phi"                    "Phrase"                
#> [310] "PlotACF"                "PlotArea"               "PlotBag"               
#> [313] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [316] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [319] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [322] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [325] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [328] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [331] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [334] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [337] "PlotProbDist"           "PlotPyramid"            "PlotQQ"                
#> [340] "PlotTernary"            "PlotTreemap"            "PlotVenn"              
#> [343] "PlotViolin"             "PlotWeb"                "PoissonCI"             
#> [346] "PolToCart"              "PolarGrid"              "PostHocTest"           
#> [349] "PpAddSlide"             "PpPlot"                 "PpText"                
#> [352] "Prec"                   "Primes"                 "PseudoR2"              
#> [355] "PtInPoly"               "PtsToCm"                "Quantile"              
#> [358] "QuantileCI"             "Quarter"                "Quot"                  
#> [361] "RBAL"                   "RMSE"                   "RSessionAlive"         
#> [364] "RTempdirAlive"          "RadToDeg"               "Range"                 
#> [367] "Rank"                   "Recode"                 "Recycle"               
#> [370] "RelRisk"                "Rename"                 "Rev"                   
#> [373] "RevCode"                "RgbToCmy"               "RgbToCol"              
#> [376] "RgbToHex"               "RgbToLong"              "RndPairs"              
#> [379] "RndWord"                "RobScale"               "RomanToInt"            
#> [382] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [385] "RunsTest"               "SD"                     "SDN"                   
#> [388] "SLN"                    "SMAPE"                  "SYD"                   
#> [391] "Sample"                 "SampleTwins"            "SaveAs"                
#> [394] "ScheffeTest"            "SecToHms"               "Second"                
#> [397] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [400] "SetAttr"                "SetNames"               "Shade"                 
#> [403] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [406] "SignTest"               "Skew"                   "Small"                 
#> [409] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [412] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [415] "Spec"                   "SphToCart"              "SplitAt"               
#> [418] "SplitPath"              "SplitToCol"             "SplitToDummy"          
#> [421] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [424] "Str"                    "StrAbbr"                "StrAlign"              
#> [427] "StrCap"                 "StrChop"                "StrCountW"             
#> [430] "StrDist"                "StrExtract"             "StrExtractBetween"     
#> [433] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [436] "StrPos"                 "StrRev"                 "StrRight"              
#> [439] "StrSpell"               "StrSplit"               "StrTrim"               
#> [442] "StrTrunc"               "StrVal"                 "Strata"                
#> [445] "StripAttr"              "StuartMaxwellTest"      "StuartTauC"            
#> [448] "SysInfo"                "TMod"                   "TOne"                  
#> [451] "TTestA"                 "TextContrastColor"      "TextToTable"           
#> [454] "TheilU"                 "Timezone"               "TitleRect"             
#> [457] "ToLong"                 "ToWide"                 "ToWrd"                 
#> [460] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [463] "Today"                  "Trim"                   "TschuprowT"            
#> [466] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [469] "UnirootAll"             "Unit"                   "Unit<-"                
#> [472] "Untable"                "Unwhich"                "VIF"                   
#> [475] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [478] "VarN"                   "VarTest"                "VarX"                  
#> [481] "VecRot"                 "VecShift"               "Vigenere"              
#> [484] "VonNeumannTest"         "Week"                   "Weekday"               
#> [487] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [490] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [493] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [496] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [499] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [502] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [505] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [508] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [511] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [514] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [517] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [520] "YTM"                    "Year"                   "YearDay"               
#> [523] "YearDays"               "YearMonth"              "YuenTTest"             
#> [526] "YuleQ"                  "YuleY"                  "ZTest"                 
#> [529] "ZeroIfNA"               "Zodiac"                 "aovlDetails"           
#> [532] "aovlErrorTerms"         "as.CDateFmt"            "as.fmt"                
#> [535] "as.matrix.xtabs"        "as.ym"                  "axTicks.Date"          
#> [538] "axTicks.POSIXct"        "createCOMReference"     "dBenf"                 
#> [541] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [544] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [547] "dNegWeibull"            "dOrder"                 "dRevGumbel"            
#> [550] "dRevWeibull"            "dTri"                   "identify.formula"      
#> [553] "lines.Lc"               "lines.SmoothSpline"     "lines.lm"              
#> [556] "lines.loess"            "lines.smooth.spline"    "pBenf"                 
#> [559] "pExtrVal"               "pFrechet"               "pGenExtrVal"           
#> [562] "pGenPareto"             "pGompertz"              "pGumbel"               
#> [565] "pNegWeibull"            "pOrder"                 "pRevGumbel"            
#> [568] "pRevWeibull"            "pTri"                   "plot.Conf"             
#> [571] "plot.Lc"                "plot.Lclist"            "plot.PostHocTest"      
#> [574] "plot.bagplot"           "plot.palette"           "power.chisq.test"      
#> [577] "print.Assocs"           "print.Conf"             "print.CorPolychor"     
#> [580] "print.DunnTest"         "print.Freq"             "print.HoeffD"          
#> [583] "print.ICC"              "print.PercTable"        "print.PostHocTest"     
#> [586] "print.mtest"            "qBenf"                  "qExtrVal"              
#> [589] "qFrechet"               "qGenExtrVal"            "qGenPareto"            
#> [592] "qGompertz"              "qGumbel"                "qNegWeibull"           
#> [595] "qRevGumbel"             "qRevGumbelExp"          "qRevWeibull"           
#> [598] "qTri"                   "rBenf"                  "rExtrVal"              
#> [601] "rFrechet"               "rGenExtrVal"            "rGenPareto"            
#> [604] "rGompertz"              "rGumbel"                "rNegWeibull"           
#> [607] "rOrder"                 "rRevGumbel"             "rRevWeibull"           
#> [610] "rSum21"                 "rTri"                   "reorder.factor"        
#> [613] "split.formula"