List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%:%"                    "%()%"                   "%(]%"                  
#>   [4] "%)(%"                   "%)[%"                   "%[)%"                  
#>   [7] "%[]%"                   "%](%"                   "%][%"                  
#>  [10] "%^%"                    "%c%"                    "%like any%"            
#>  [13] "%like%"                 "%nin%"                  "%overlaps%"            
#>  [16] "Abind"                  "Abstract"               "AddMonths"             
#>  [19] "AddMonthsYM"            "Agree"                  "AllDuplicated"         
#>  [22] "AllIdentical"           "AndersonDarlingTest"    "aovlDetails"           
#>  [25] "aovlErrorTerms"         "Append"                 "Arrow"                 
#>  [28] "as.CDateFmt"            "as.fmt"                 "as.matrix.xtabs"       
#>  [31] "AscToChar"              "Asp"                    "Assocs"                
#>  [34] "Atkinson"               "AUC"                    "AxisBreak"             
#>  [37] "axTicks.Date"           "axTicks.POSIXct"        "BarnardTest"           
#>  [40] "BartelsRankTest"        "BarText"                "BhapkarTest"           
#>  [43] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [46] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [49] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [52] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [55] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [58] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [61] "CartToSph"              "CatTable"               "CCC"                   
#>  [64] "CharToAsc"              "Clockwise"              "Closest"               
#>  [67] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [70] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [73] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [76] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [79] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [82] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [85] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [88] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [91] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [94] "ContCoef"               "Contrasts"              "ConvUnit"              
#>  [97] "Cor"                    "CorCI"                  "CorPart"               
#> [100] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [103] "CourseData"             "Cov"                    "CramerV"               
#> [106] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [109] "Cross"                  "CrossN"                 "Cstat"                 
#> [112] "CutQ"                   "Day"                    "Day<-"                 
#> [115] "DB"                     "dBenf"                  "DecToBin"              
#> [118] "DecToHex"               "DecToOct"               "DegToRad"              
#> [121] "DenseRank"              "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [271] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [274] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [277] "Midx"                   "Minute"                 "MixColor"              
#> [280] "Mode"                   "ModSummary"             "Month"                 
#> [283] "MosesTest"              "MoveAvg"                "MSE"                   
#> [286] "MultinomCI"             "MultMerge"              "MutInf"                
#> [289] "N"                      "NAIf"                   "NAIfBlank"             
#> [292] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [295] "NMAE"                   "NMSE"                   "Now"                   
#> [298] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [301] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [304] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [307] "PageTest"               "PairApply"              "Pal"                   
#> [310] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [313] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [316] "PearsonTest"            "PercentRank"            "PercTable"             
#> [319] "Permn"                  "pExtrVal"               "pFrechet"              
#> [322] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [325] "pGumbel"                "Phi"                    "Phrase"                
#> [328] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [331] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [334] "PlotACF"                "PlotArea"               "PlotBag"               
#> [337] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [340] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [343] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [346] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [349] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [352] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [355] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [358] "PlotPairs"              "PlotPolar"              "PlotPyramid"           
#> [361] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [364] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [367] "PMT"                    "pNegWeibull"            "PoissonCI"             
#> [370] "PolarGrid"              "PolToCart"              "pOrder"                
#> [373] "PostHocTest"            "power.chisq.test"       "PpAddSlide"            
#> [376] "PPMT"                   "PpPlot"                 "PpText"                
#> [379] "Prec"                   "pRevGumbel"             "pRevWeibull"           
#> [382] "Primes"                 "print.Assocs"           "print.Conf"            
#> [385] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [388] "print.HoeffD"           "print.ICC"              "print.mtest"           
#> [391] "print.PercTable"        "print.PostHocTest"      "PseudoR2"              
#> [394] "PtInPoly"               "PtsToCm"                "qBenf"                 
#> [397] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [400] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [403] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [406] "qRevWeibull"            "Quantile"               "QuantileCI"            
#> [409] "Quarter"                "Quot"                   "RadToDeg"              
#> [412] "Range"                  "RBAL"                   "rBenf"                 
#> [415] "Recode"                 "Recycle"                "RelRisk"               
#> [418] "Rename"                 "reorder.factor"         "Rev"                   
#> [421] "RevCode"                "rExtrVal"               "rFrechet"              
#> [424] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [427] "RgbToLong"              "rGenExtrVal"            "rGenPareto"            
#> [430] "rGompertz"              "rGumbel"                "RMSE"                  
#> [433] "RndPairs"               "RndWord"                "rNegWeibull"           
#> [436] "RobScale"               "RomanToInt"             "rOrder"                
#> [439] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [442] "rRevGumbel"             "rRevWeibull"            "RSessionAlive"         
#> [445] "RTempdirAlive"          "RunsTest"               "Sample"                
#> [448] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [451] "SD"                     "Second"                 "SecToHms"              
#> [454] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [457] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [460] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [463] "Skew"                   "SLN"                    "Small"                 
#> [466] "SMAPE"                  "SmoothSpline"           "Some"                  
#> [469] "SomersDelta"            "Sort"                   "SortMixed"             
#> [472] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [475] "split.formula"          "SplitAt"                "SplitPath"             
#> [478] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [481] "StdCoef"                "Str"                    "StrAbbr"               
#> [484] "StrAlign"               "Strata"                 "StrCap"                
#> [487] "StrChop"                "StrCountW"              "StrDist"               
#> [490] "StrExtract"             "StrExtractBetween"      "StripAttr"             
#> [493] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [496] "StrPos"                 "StrRev"                 "StrRight"              
#> [499] "StrSpell"               "StrSplit"               "StrTrim"               
#> [502] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [505] "StuartTauC"             "SYD"                    "SysInfo"               
#> [508] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [511] "Timezone"               "TitleRect"              "TMod"                  
#> [514] "Today"                  "ToLong"                 "TOne"                  
#> [517] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [520] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [523] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [526] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [529] "Unit"                   "Unit<-"                 "Untable"               
#> [532] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [535] "VarCI"                  "VarTest"                "VecRot"                
#> [538] "VecShift"               "VIF"                    "Vigenere"              
#> [541] "VonNeumannTest"         "Week"                   "Weekday"               
#> [544] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [547] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [550] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [553] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [556] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [559] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [562] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [565] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [568] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [571] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [574] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [577] "Year"                   "YearDay"                "YearMonth"             
#> [580] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [583] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [586] "ZTest"