List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%()%"                   "%(]%"                   "%)(%"                  
#>   [4] "%)[%"                   "%[)%"                   "%[]%"                  
#>   [7] "%](%"                   "%][%"                   "%^%"                   
#>  [10] "%c%"                    "%like any%"             "%like%"                
#>  [13] "%nin%"                  "%overlaps%"             "Abind"                 
#>  [16] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [19] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [22] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [25] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [28] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [31] "Asp"                    "Assocs"                 "Atkinson"              
#>  [34] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [37] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [40] "BarText"                "BhapkarTest"            "BinomCI"               
#>  [43] "BinomCIn"               "BinomDiffCI"            "BinomRatioCI"          
#>  [46] "BinToDec"               "BinTree"                "BlankIfNA"             
#>  [49] "BootCI"                 "BoxCox"                 "BoxCoxInv"             
#>  [52] "BoxCoxLambda"           "BoxedText"              "BreslowDayTest"        
#>  [55] "BreuschGodfreyTest"     "BrierScore"             "BubbleLegend"          
#>  [58] "Canvas"                 "CartToPol"              "CartToSph"             
#>  [61] "CatTable"               "CCC"                    "CharToAsc"             
#>  [64] "Clockwise"              "Closest"                "CmToPts"               
#>  [67] "CmykToCmy"              "CmykToRgb"              "CmyToCmyk"             
#>  [70] "Coalesce"               "CochranArmitageTest"    "CochranQTest"          
#>  [73] "CoefVar"                "CohenD"                 "CohenKappa"            
#>  [76] "CollapseTable"          "ColorLegend"            "ColToGray"             
#>  [79] "ColToGrey"              "ColToHex"               "ColToHsv"              
#>  [82] "ColToOpaque"            "ColToRgb"               "ColumnWrap"            
#>  [85] "CombN"                  "CombPairs"              "CombSet"               
#>  [88] "CompleteColumns"        "ConDisPairs"            "Conf"                  
#>  [91] "ConnLines"              "ConoverTest"            "ContCoef"              
#>  [94] "Contrasts"              "ConvUnit"               "Cor"                   
#>  [97] "CorCI"                  "CorPart"                "CorPolychor"           
#> [100] "CountCompCases"         "CountWorkDays"          "CourseData"            
#> [103] "Cov"                    "CramerV"                "CramerVonMisesTest"    
#> [106] "createCOMReference"     "CronbachAlpha"          "Cross"                 
#> [109] "CrossN"                 "Cstat"                  "CutQ"                  
#> [112] "Day"                    "Day<-"                  "DB"                    
#> [115] "dBenf"                  "DecToBin"               "DecToHex"              
#> [118] "DecToOct"               "DegToRad"               "DenseRank"             
#> [121] "Desc"                   "DescToolsOptions"       "dExtrVal"              
#> [124] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [127] "dGompertz"              "dGumbel"                "DiffDays360"           
#> [130] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [133] "Divisors"               "dNegWeibull"            "DoBy"                  
#> [136] "DoCall"                 "dOrder"                 "Dot"                   
#> [139] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [142] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [145] "dRevGumbel"             "dRevWeibull"            "Dummy"                 
#> [148] "DunnettTest"            "DunnTest"               "DurbinWatsonTest"      
#> [151] "Entropy"                "Eps"                    "ErrBars"               
#> [154] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [157] "FctArgs"                "Fibonacci"              "FindColor"             
#> [160] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [163] "FisherZInv"             "FixToTable"             "Flags"                 
#> [166] "Fmt"                    "Format"                 "Frac"                  
#> [169] "Freq"                   "Freq2D"                 "GCD"                   
#> [172] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [175] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [178] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [181] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [184] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [187] "Gsd"                    "GTest"                  "Herfindahl"            
#> [190] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [193] "HighLow"                "Hmean"                  "HmsToSec"              
#> [196] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [199] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [202] "HunterGaston"           "ICC"                    "identify.formula"      
#> [205] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [208] "InDots"                 "Interval"               "IPMT"                  
#> [211] "IQRw"                   "IRR"                    "IsDate"                
#> [214] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [217] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [220] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [223] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [226] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [229] "KendallW"               "Keywords"               "KrippAlpha"            
#> [232] "Kurt"                   "Label"                  "Label<-"               
#> [235] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [238] "Lc"                     "LCM"                    "LehmacherTest"         
#> [241] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [244] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [247] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [250] "LOCF"                   "LOF"                    "Logit"                 
#> [253] "LogitInv"               "LogSt"                  "LogStInv"              
#> [256] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [259] "MAD"                    "MADCI"                  "MAE"                   
#> [262] "MAPE"                   "Mar"                    "Margins"               
#> [265] "MaxDigits"              "Mean"                   "MeanAD"                
#> [268] "MeanCI"                 "MeanDiffCI"             "MeanSE"                
#> [271] "Median"                 "MedianCI"               "Mgp"                   
#> [274] "MHChisqTest"            "Midx"                   "Minute"                
#> [277] "MixColor"               "Mode"                   "ModSummary"            
#> [280] "Month"                  "MosesTest"              "MoveAvg"               
#> [283] "MSE"                    "MultinomCI"             "MultMerge"             
#> [286] "MutInf"                 "N"                      "NAIfBlank"             
#> [289] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [292] "NMAE"                   "NMSE"                   "Now"                   
#> [295] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [298] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [301] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [304] "PageTest"               "PairApply"              "Pal"                   
#> [307] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [310] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [313] "PearsonTest"            "PercentRank"            "PercTable"             
#> [316] "Permn"                  "pExtrVal"               "pFrechet"              
#> [319] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [322] "pGumbel"                "Phi"                    "Phrase"                
#> [325] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [328] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [331] "PlotACF"                "PlotArea"               "PlotBag"               
#> [334] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [337] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [340] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [343] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [346] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [349] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [352] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [355] "PlotPairs"              "PlotPolar"              "PlotPyramid"           
#> [358] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [361] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [364] "PMT"                    "pNegWeibull"            "PoissonCI"             
#> [367] "PolarGrid"              "PolToCart"              "pOrder"                
#> [370] "PostHocTest"            "power.chisq.test"       "PpAddSlide"            
#> [373] "PPMT"                   "PpPlot"                 "PpText"                
#> [376] "Prec"                   "pRevGumbel"             "pRevWeibull"           
#> [379] "Primes"                 "print.Assocs"           "print.Conf"            
#> [382] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [385] "print.HoeffD"           "print.ICC"              "print.mtest"           
#> [388] "print.PercTable"        "print.PostHocTest"      "PseudoR2"              
#> [391] "PtInPoly"               "PtsToCm"                "qBenf"                 
#> [394] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [397] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [400] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [403] "qRevWeibull"            "Quantile"               "QuantileCI"            
#> [406] "Quarter"                "Quot"                   "RadToDeg"              
#> [409] "Range"                  "RBAL"                   "rBenf"                 
#> [412] "Recode"                 "Recycle"                "RelRisk"               
#> [415] "Rename"                 "reorder.factor"         "Rev"                   
#> [418] "RevCode"                "rExtrVal"               "rFrechet"              
#> [421] "RgbToCmy"               "RgbToCol"               "RgbToHex"              
#> [424] "RgbToLong"              "rGenExtrVal"            "rGenPareto"            
#> [427] "rGompertz"              "rGumbel"                "RMSE"                  
#> [430] "RndPairs"               "RndWord"                "rNegWeibull"           
#> [433] "RobScale"               "RomanToInt"             "rOrder"                
#> [436] "Rosenbluth"             "Rotate"                 "RoundTo"               
#> [439] "rRevGumbel"             "rRevWeibull"            "RSessionAlive"         
#> [442] "RTempdirAlive"          "RunsTest"               "Sample"                
#> [445] "SampleTwins"            "SaveAs"                 "ScheffeTest"           
#> [448] "SD"                     "Second"                 "SecToHms"              
#> [451] "SendOutlookMail"        "Sens"                   "SetAlpha"              
#> [454] "SetNames"               "Shade"                  "ShapiroFranciaTest"    
#> [457] "SiegelTukeyRank"        "SiegelTukeyTest"        "SignTest"              
#> [460] "Skew"                   "SLN"                    "Small"                 
#> [463] "SMAPE"                  "SmoothSpline"           "Some"                  
#> [466] "SomersDelta"            "Sort"                   "SortMixed"             
#> [469] "SpearmanRho"            "Spec"                   "SphToCart"             
#> [472] "split.formula"          "SplitAt"                "SplitPath"             
#> [475] "SplitToCol"             "SpreadOut"              "Stamp"                 
#> [478] "StdCoef"                "Str"                    "StrAbbr"               
#> [481] "StrAlign"               "Strata"                 "StrCap"                
#> [484] "StrChop"                "StrCountW"              "StrDist"               
#> [487] "StrExtract"             "StrExtractBetween"      "StripAttr"             
#> [490] "StrIsNumeric"           "StrLeft"                "StrPad"                
#> [493] "StrPos"                 "StrRev"                 "StrRight"              
#> [496] "StrSpell"               "StrSplit"               "StrTrim"               
#> [499] "StrTrunc"               "StrVal"                 "StuartMaxwellTest"     
#> [502] "StuartTauC"             "SYD"                    "SysInfo"               
#> [505] "TextContrastColor"      "TextToTable"            "TheilU"                
#> [508] "Timezone"               "TitleRect"              "TMod"                  
#> [511] "Today"                  "ToLong"                 "TOne"                  
#> [514] "ToWide"                 "ToWrd"                  "ToWrdB"                
#> [517] "ToWrdPlot"              "ToXL"                   "Trim"                  
#> [520] "TschuprowT"             "TTestA"                 "TukeyBiweight"         
#> [523] "TwoGroups"              "UncertCoef"             "UnirootAll"            
#> [526] "Unit"                   "Unit<-"                 "Untable"               
#> [529] "Unwhich"                "VanWaerdenTest"         "Var"                   
#> [532] "VarCI"                  "VarTest"                "VecRot"                
#> [535] "VecShift"               "VIF"                    "Vigenere"              
#> [538] "VonNeumannTest"         "Week"                   "Weekday"               
#> [541] "Winsorize"              "WithOptions"            "WoolfTest"             
#> [544] "WrdBookmark"            "WrdCaption"             "WrdCellRange"          
#> [547] "WrdDeleteBookmark"      "WrdFont"                "WrdFont<-"             
#> [550] "WrdFormatCells"         "WrdGoto"                "WrdInsertBookmark"     
#> [553] "WrdKill"                "WrdMergeCells"          "WrdOpenFile"           
#> [556] "WrdPageBreak"           "WrdParagraphFormat"     "WrdParagraphFormat<-"  
#> [559] "WrdPlot"                "WrdSaveAs"              "WrdStyle"              
#> [562] "WrdStyle<-"             "WrdTable"               "WrdTableBorders"       
#> [565] "WrdUpdateBookmark"      "XLCurrReg"              "XLDateToPOSIXct"       
#> [568] "XLGetRange"             "XLGetWorkbook"          "XLKill"                
#> [571] "XLNamedReg"             "XLSaveAs"               "XLView"                
#> [574] "Year"                   "YearDay"                "YearMonth"             
#> [577] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [580] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [583] "ZTest"