List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%::%"                   "%:%"                    "%()%"                  
#>   [4] "%(]%"                   "%)(%"                   "%)[%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "Abind"                  "Abstract"              
#>  [19] "AddMonths"              "AddMonthsYM"            "Agree"                 
#>  [22] "AllDuplicated"          "AllIdentical"           "AndersonDarlingTest"   
#>  [25] "aovlDetails"            "aovlErrorTerms"         "Append"                
#>  [28] "Arrow"                  "as.CDateFmt"            "as.fmt"                
#>  [31] "as.matrix.xtabs"        "AscToChar"              "Asp"                   
#>  [34] "Assocs"                 "Atkinson"               "AUC"                   
#>  [37] "AxisBreak"              "axTicks.Date"           "axTicks.POSIXct"       
#>  [40] "BarnardTest"            "BartelsRankTest"        "BarText"               
#>  [43] "BhapkarTest"            "BinomCI"                "BinomCIn"              
#>  [46] "BinomDiffCI"            "BinomRatioCI"           "BinToDec"              
#>  [49] "BinTree"                "BlankIfNA"              "BootCI"                
#>  [52] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [55] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [58] "BrierScore"             "BubbleLegend"           "Canvas"                
#>  [61] "CartToPol"              "CartToSph"              "CatTable"              
#>  [64] "CCC"                    "CharToAsc"              "Clockwise"             
#>  [67] "Closest"                "CmToPts"                "CmykToCmy"             
#>  [70] "CmykToRgb"              "CmyToCmyk"              "Coalesce"              
#>  [73] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [76] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [79] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [82] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [85] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [88] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [91] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [94] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [97] "ConvUnit"               "Cor"                    "CorCI"                 
#> [100] "CorPart"                "CorPolychor"            "CountCompCases"        
#> [103] "CountWorkDays"          "CourseData"             "Cov"                   
#> [106] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [109] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [112] "Cstat"                  "CutQ"                   "Day"                   
#> [115] "Day<-"                  "DB"                     "dBenf"                 
#> [118] "DecToBin"               "DecToHex"               "DecToOct"              
#> [121] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [271] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [274] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [277] "Midx"                   "Minute"                 "MixColor"              
#> [280] "Mode"                   "ModSummary"             "Month"                 
#> [283] "MonthDays"              "MosesTest"              "MoveAvg"               
#> [286] "MSE"                    "MultinomCI"             "MultMerge"             
#> [289] "MutInf"                 "N"                      "NAIf"                  
#> [292] "NAIfBlank"              "NAIfZero"               "Ndec"                  
#> [295] "NemenyiTest"            "NMAE"                   "NMSE"                  
#> [298] "Now"                    "NPV"                    "NPVFixBond"            
#> [301] "OctToDec"               "OddsRatio"              "OPR"                   
#> [304] "OrderMixed"             "ORToRelRisk"            "Outlier"               
#> [307] "Overlap"                "PageTest"               "PairApply"             
#> [310] "Pal"                    "ParseFormula"           "ParseSASDatalines"     
#> [313] "PartialSD"              "PasswordDlg"            "pBenf"                 
#> [316] "PDFManual"              "PearsonTest"            "PercentRank"           
#> [319] "PercTable"              "Permn"                  "pExtrVal"              
#> [322] "pFrechet"               "pGenExtrVal"            "pGenPareto"            
#> [325] "pGompertz"              "pGumbel"                "Phi"                   
#> [328] "Phrase"                 "plot.bagplot"           "plot.Conf"             
#> [331] "plot.Lc"                "plot.Lclist"            "plot.palette"          
#> [334] "plot.PostHocTest"       "PlotACF"                "PlotArea"              
#> [337] "PlotBag"                "PlotBagPairs"           "PlotBinTree"           
#> [340] "PlotBubble"             "PlotCandlestick"        "PlotCashFlow"          
#> [343] "PlotCirc"               "PlotConDens"            "PlotCorr"              
#> [346] "PlotDot"                "PlotECDF"               "PlotFaces"             
#> [349] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [352] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [355] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [358] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [361] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [364] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [367] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [370] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [373] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [376] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [379] "PpText"                 "Prec"                   "pRevGumbel"            
#> [382] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [385] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [388] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [391] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [394] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [397] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [400] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [403] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [406] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [409] "QuantileCI"             "Quarter"                "Quot"                  
#> [412] "RadToDeg"               "Range"                  "Rank"                  
#> [415] "RBAL"                   "rBenf"                  "Recode"                
#> [418] "Recycle"                "RelRisk"                "Rename"                
#> [421] "reorder.factor"         "Rev"                    "RevCode"               
#> [424] "rExtrVal"               "rFrechet"               "RgbToCmy"              
#> [427] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [430] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [433] "rGumbel"                "RMSE"                   "RndPairs"              
#> [436] "RndWord"                "rNegWeibull"            "RobScale"              
#> [439] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [442] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [445] "rRevWeibull"            "RSessionAlive"          "RTempdirAlive"         
#> [448] "RunsTest"               "Sample"                 "SampleTwins"           
#> [451] "SaveAs"                 "ScheffeTest"            "SD"                    
#> [454] "Second"                 "SecToHms"               "SendOutlookMail"       
#> [457] "Sens"                   "SetAlpha"               "SetNames"              
#> [460] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [463] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [466] "SLN"                    "Small"                  "SMAPE"                 
#> [469] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [472] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [475] "Spec"                   "SphToCart"              "split.formula"         
#> [478] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [481] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [484] "Str"                    "StrAbbr"                "StrAlign"              
#> [487] "Strata"                 "StrCap"                 "StrChop"               
#> [490] "StrCountW"              "StrDist"                "StrExtract"            
#> [493] "StrExtractBetween"      "StripAttr"              "StrIsNumeric"          
#> [496] "StrLeft"                "StrPad"                 "StrPos"                
#> [499] "StrRev"                 "StrRight"               "StrSpell"              
#> [502] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [505] "StrVal"                 "StuartMaxwellTest"      "StuartTauC"            
#> [508] "SYD"                    "SysInfo"                "TextContrastColor"     
#> [511] "TextToTable"            "TheilU"                 "Timezone"              
#> [514] "TitleRect"              "TMod"                   "Today"                 
#> [517] "ToLong"                 "TOne"                   "ToWide"                
#> [520] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [523] "ToXL"                   "Trim"                   "TschuprowT"            
#> [526] "TTestA"                 "TukeyBiweight"          "TwoGroups"             
#> [529] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [532] "Unit<-"                 "Untable"                "Unwhich"               
#> [535] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [538] "VarTest"                "VecRot"                 "VecShift"              
#> [541] "VIF"                    "Vigenere"               "VonNeumannTest"        
#> [544] "Week"                   "Weekday"                "Winsorize"             
#> [547] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [550] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [553] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [556] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [559] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [562] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [565] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [568] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [571] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [574] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [577] "XLSaveAs"               "XLView"                 "Year"                  
#> [580] "YearDay"                "YearDays"               "YearMonth"             
#> [583] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [586] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [589] "ZTest"