List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%::%"                   "%:%"                    "%()%"                  
#>   [4] "%(]%"                   "%)(%"                   "%)[%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "ABCCoords"              "Abind"                 
#>  [19] "Abstract"               "AddMonths"              "AddMonthsYM"           
#>  [22] "Agree"                  "AllDuplicated"          "AllIdentical"          
#>  [25] "AndersonDarlingTest"    "aovlDetails"            "aovlErrorTerms"        
#>  [28] "Append"                 "Arrow"                  "as.CDateFmt"           
#>  [31] "as.fmt"                 "as.matrix.xtabs"        "AscToChar"             
#>  [34] "Asp"                    "Assocs"                 "Atkinson"              
#>  [37] "AUC"                    "AxisBreak"              "axTicks.Date"          
#>  [40] "axTicks.POSIXct"        "BarnardTest"            "BartelsRankTest"       
#>  [43] "BarText"                "Bg"                     "BhapkarTest"           
#>  [46] "BinomCI"                "BinomCIn"               "BinomDiffCI"           
#>  [49] "BinomRatioCI"           "BinToDec"               "BinTree"               
#>  [52] "BlankIfNA"              "BootCI"                 "BoxCox"                
#>  [55] "BoxCoxInv"              "BoxCoxLambda"           "BoxedText"             
#>  [58] "BreslowDayTest"         "BreuschGodfreyTest"     "BrierScore"            
#>  [61] "BubbleLegend"           "Canvas"                 "CartToPol"             
#>  [64] "CartToSph"              "CatTable"               "CCC"                   
#>  [67] "CharToAsc"              "Clockwise"              "Closest"               
#>  [70] "CmToPts"                "CmykToCmy"              "CmykToRgb"             
#>  [73] "CmyToCmyk"              "Coalesce"               "CochranArmitageTest"   
#>  [76] "CochranQTest"           "CoefVar"                "CohenD"                
#>  [79] "CohenKappa"             "CollapseTable"          "ColorLegend"           
#>  [82] "ColToGray"              "ColToGrey"              "ColToHex"              
#>  [85] "ColToHsv"               "ColToOpaque"            "ColToRgb"              
#>  [88] "ColumnWrap"             "CombN"                  "CombPairs"             
#>  [91] "CombSet"                "CompleteColumns"        "ConDisPairs"           
#>  [94] "Conf"                   "ConnLines"              "ConoverTest"           
#>  [97] "ContCoef"               "Contrasts"              "ConvUnit"              
#> [100] "Cor"                    "CorCI"                  "CorPart"               
#> [103] "CorPolychor"            "CountCompCases"         "CountWorkDays"         
#> [106] "CourseData"             "Cov"                    "CramerV"               
#> [109] "CramerVonMisesTest"     "createCOMReference"     "CronbachAlpha"         
#> [112] "Cross"                  "CrossN"                 "Cstat"                 
#> [115] "CutQ"                   "Day"                    "Day<-"                 
#> [118] "DB"                     "dBenf"                  "DecToBin"              
#> [121] "DecToHex"               "DecToOct"               "DegToRad"              
#> [124] "Desc"                   "DescToolsOptions"       "dExtrVal"              
#> [127] "dFrechet"               "dGenExtrVal"            "dGenPareto"            
#> [130] "dGompertz"              "dGumbel"                "DiffDays360"           
#> [133] "DigitSum"               "DivCoef"                "DivCoefMax"            
#> [136] "Divisors"               "dNegWeibull"            "DoBy"                  
#> [139] "DoCall"                 "dOrder"                 "Dot"                   
#> [142] "DrawArc"                "DrawBand"               "DrawBezier"            
#> [145] "DrawCircle"             "DrawEllipse"            "DrawRegPolygon"        
#> [148] "dRevGumbel"             "dRevWeibull"            "Dummy"                 
#> [151] "DunnettTest"            "DunnTest"               "DurbinWatsonTest"      
#> [154] "Entropy"                "Eps"                    "ErrBars"               
#> [157] "EtaSq"                  "ExpFreq"                "Factorize"             
#> [160] "FctArgs"                "Fibonacci"              "FindColor"             
#> [163] "FindCorr"               "FindRProfile"           "FisherZ"               
#> [166] "FisherZInv"             "FixToTable"             "Flags"                 
#> [169] "Fmt"                    "Format"                 "Frac"                  
#> [172] "Freq"                   "Freq2D"                 "GCD"                   
#> [175] "GenRandGroups"          "GeomSn"                 "GeomTrans"             
#> [178] "GetCalls"               "GetCurrPP"              "GetCurrWrd"            
#> [181] "GetCurrXL"              "GetNewPP"               "GetNewWrd"             
#> [184] "GetNewXL"               "Gini"                   "GiniSimpson"           
#> [187] "Gmean"                  "GoodmanKruskalGamma"    "GoodmanKruskalTau"     
#> [190] "Gsd"                    "GTest"                  "Herfindahl"            
#> [193] "HexToCol"               "HexToDec"               "HexToRgb"              
#> [196] "HighLow"                "Hmean"                  "HmsToSec"              
#> [199] "HodgesLehmann"          "HoeffD"                 "HosmerLemeshowTest"    
#> [202] "HotellingsT2Test"       "Hour"                   "HuberM"                
#> [205] "HunterGaston"           "ICC"                    "identify.formula"      
#> [208] "IdentifyA"              "Impute"                 "ImputeKnn"             
#> [211] "InDots"                 "Interval"               "IPMT"                  
#> [214] "IQRw"                   "IRR"                    "IsDate"                
#> [217] "IsDichotomous"          "IsEuclid"               "IsLeapYear"            
#> [220] "IsNumeric"              "IsOdd"                  "IsPrime"               
#> [223] "IsValidHwnd"            "IsWeekend"              "IsWhole"               
#> [226] "IsZero"                 "JarqueBeraTest"         "JonckheereTerpstraTest"
#> [229] "KappaM"                 "KendallTauA"            "KendallTauB"           
#> [232] "KendallW"               "Keywords"               "KrippAlpha"            
#> [235] "Kurt"                   "Label"                  "Label<-"               
#> [238] "Lambda"                 "Large"                  "LastDayOfMonth"        
#> [241] "Lc"                     "LCM"                    "LehmacherTest"         
#> [244] "LeveneTest"             "LillieTest"             "lines.Lc"              
#> [247] "lines.lm"               "lines.loess"            "lines.smooth.spline"   
#> [250] "lines.SmoothSpline"     "LineToUser"             "LinScale"              
#> [253] "LOCF"                   "LOF"                    "Logit"                 
#> [256] "LogitInv"               "LogSt"                  "LogStInv"              
#> [259] "LongToRgb"              "LsFct"                  "LsObj"                 
#> [262] "MAD"                    "MADCI"                  "MAE"                   
#> [265] "MAPE"                   "Mar"                    "Margins"               
#> [268] "MaxDigits"              "Mean"                   "MeanAD"                
#> [271] "MeanCI"                 "MeanCIn"                "MeanDiffCI"            
#> [274] "MeanSE"                 "Median"                 "MedianCI"              
#> [277] "Mgp"                    "MHChisqTest"            "Midx"                  
#> [280] "Minute"                 "MixColor"               "Mode"                  
#> [283] "ModSummary"             "Month"                  "MonthDays"             
#> [286] "MosesTest"              "MoveAvg"                "MSE"                   
#> [289] "MultinomCI"             "MultMerge"              "MutInf"                
#> [292] "N"                      "NAIf"                   "NAIfBlank"             
#> [295] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [298] "NMAE"                   "NMSE"                   "Now"                   
#> [301] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [304] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [307] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [310] "PageTest"               "PairApply"              "Pal"                   
#> [313] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [316] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [319] "PearsonTest"            "PercentRank"            "PercTable"             
#> [322] "Permn"                  "pExtrVal"               "pFrechet"              
#> [325] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [328] "pGumbel"                "Phi"                    "Phrase"                
#> [331] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [334] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [337] "PlotACF"                "PlotArea"               "PlotBag"               
#> [340] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [343] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [346] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [349] "PlotDotCI"              "PlotECDF"               "PlotFaces"             
#> [352] "PlotFdist"              "PlotFun"                "PlotGACF"              
#> [355] "PlotLinesA"             "PlotLog"                "PlotMarDens"           
#> [358] "PlotMiss"               "PlotMonth"              "PlotMosaic"            
#> [361] "PlotMultiDens"          "PlotPairs"              "PlotPolar"             
#> [364] "PlotPyramid"            "PlotQQ"                 "PlotTernary"           
#> [367] "PlotTreemap"            "PlotVenn"               "PlotViolin"            
#> [370] "PlotWeb"                "PMT"                    "pNegWeibull"           
#> [373] "PoissonCI"              "PolarGrid"              "PolToCart"             
#> [376] "pOrder"                 "PostHocTest"            "power.chisq.test"      
#> [379] "PpAddSlide"             "PPMT"                   "PpPlot"                
#> [382] "PpText"                 "Prec"                   "pRevGumbel"            
#> [385] "pRevWeibull"            "Primes"                 "print.Assocs"          
#> [388] "print.Conf"             "print.CorPolychor"      "print.DunnTest"        
#> [391] "print.Freq"             "print.HoeffD"           "print.ICC"             
#> [394] "print.mtest"            "print.PercTable"        "print.PostHocTest"     
#> [397] "PseudoR2"               "PtInPoly"               "PtsToCm"               
#> [400] "qBenf"                  "qExtrVal"               "qFrechet"              
#> [403] "qGenExtrVal"            "qGenPareto"             "qGompertz"             
#> [406] "qGumbel"                "qNegWeibull"            "qRevGumbel"            
#> [409] "qRevGumbelExp"          "qRevWeibull"            "Quantile"              
#> [412] "QuantileCI"             "Quarter"                "Quot"                  
#> [415] "RadToDeg"               "Range"                  "Rank"                  
#> [418] "RBAL"                   "rBenf"                  "Recode"                
#> [421] "Recycle"                "RelRisk"                "Rename"                
#> [424] "reorder.factor"         "Rev"                    "RevCode"               
#> [427] "rExtrVal"               "rFrechet"               "RgbToCmy"              
#> [430] "RgbToCol"               "RgbToHex"               "RgbToLong"             
#> [433] "rGenExtrVal"            "rGenPareto"             "rGompertz"             
#> [436] "rGumbel"                "RMSE"                   "RndPairs"              
#> [439] "RndWord"                "rNegWeibull"            "RobScale"              
#> [442] "RomanToInt"             "rOrder"                 "Rosenbluth"            
#> [445] "Rotate"                 "RoundTo"                "rRevGumbel"            
#> [448] "rRevWeibull"            "RSessionAlive"          "RTempdirAlive"         
#> [451] "RunsTest"               "Sample"                 "SampleTwins"           
#> [454] "SaveAs"                 "ScheffeTest"            "SD"                    
#> [457] "Second"                 "SecToHms"               "SendOutlookMail"       
#> [460] "Sens"                   "SetAlpha"               "SetNames"              
#> [463] "Shade"                  "ShapiroFranciaTest"     "SiegelTukeyRank"       
#> [466] "SiegelTukeyTest"        "SignTest"               "Skew"                  
#> [469] "SLN"                    "Small"                  "SMAPE"                 
#> [472] "SmoothSpline"           "Some"                   "SomersDelta"           
#> [475] "Sort"                   "SortMixed"              "SpearmanRho"           
#> [478] "Spec"                   "SphToCart"              "split.formula"         
#> [481] "SplitAt"                "SplitPath"              "SplitToCol"            
#> [484] "SpreadOut"              "Stamp"                  "StdCoef"               
#> [487] "Str"                    "StrAbbr"                "StrAlign"              
#> [490] "Strata"                 "StrCap"                 "StrChop"               
#> [493] "StrCountW"              "StrDist"                "StrExtract"            
#> [496] "StrExtractBetween"      "StripAttr"              "StrIsNumeric"          
#> [499] "StrLeft"                "StrPad"                 "StrPos"                
#> [502] "StrRev"                 "StrRight"               "StrSpell"              
#> [505] "StrSplit"               "StrTrim"                "StrTrunc"              
#> [508] "StrVal"                 "StuartMaxwellTest"      "StuartTauC"            
#> [511] "SYD"                    "SysInfo"                "TextContrastColor"     
#> [514] "TextToTable"            "TheilU"                 "Timezone"              
#> [517] "TitleRect"              "TMod"                   "Today"                 
#> [520] "ToLong"                 "TOne"                   "ToWide"                
#> [523] "ToWrd"                  "ToWrdB"                 "ToWrdPlot"             
#> [526] "ToXL"                   "Trim"                   "TschuprowT"            
#> [529] "TTestA"                 "TukeyBiweight"          "TwoGroups"             
#> [532] "UncertCoef"             "UnirootAll"             "Unit"                  
#> [535] "Unit<-"                 "Untable"                "Unwhich"               
#> [538] "VanWaerdenTest"         "Var"                    "VarCI"                 
#> [541] "VarTest"                "VecRot"                 "VecShift"              
#> [544] "VIF"                    "Vigenere"               "VonNeumannTest"        
#> [547] "Week"                   "Weekday"                "Winsorize"             
#> [550] "WithOptions"            "WoolfTest"              "WrdBookmark"           
#> [553] "WrdCaption"             "WrdCellRange"           "WrdDeleteBookmark"     
#> [556] "WrdFont"                "WrdFont<-"              "WrdFormatCells"        
#> [559] "WrdGoto"                "WrdInsertBookmark"      "WrdKill"               
#> [562] "WrdMergeCells"          "WrdOpenFile"            "WrdPageBreak"          
#> [565] "WrdParagraphFormat"     "WrdParagraphFormat<-"   "WrdPlot"               
#> [568] "WrdSaveAs"              "WrdStyle"               "WrdStyle<-"            
#> [571] "WrdTable"               "WrdTableBorders"        "WrdUpdateBookmark"     
#> [574] "XLCurrReg"              "XLDateToPOSIXct"        "XLGetRange"            
#> [577] "XLGetWorkbook"          "XLKill"                 "XLNamedReg"            
#> [580] "XLSaveAs"               "XLView"                 "Year"                  
#> [583] "YearDay"                "YearDays"               "YearMonth"             
#> [586] "YTM"                    "YuenTTest"              "YuleQ"                 
#> [589] "YuleY"                  "ZeroIfNA"               "Zodiac"                
#> [592] "ZTest"