List Objects, Functions Or Data in a Package

List all the objects, functions or data in a package.

LsObj(package)
LsFct(package)

Arguments

package
the name of the package

Details

This is just a wrapper for ls, ls.str and lsf.str with the appropriate arguments (as I always forgot how to do the trick). LsObj() lists all objects, LsFct() just the functions in a package.

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

LsFct("DescTools")

#>   [1] "%::%"                   "%:%"                    "%()%"                  
#>   [4] "%(]%"                   "%)(%"                   "%)[%"                  
#>   [7] "%[)%"                   "%[]%"                   "%](%"                  
#>  [10] "%][%"                   "%^%"                    "%c%"                   
#>  [13] "%like any%"             "%like%"                 "%nin%"                 
#>  [16] "%overlaps%"             "Abind"                  "Abstract"              
#>  [19] "AddMonths"              "AddMonthsYM"            "Agree"                 
#>  [22] "AllDuplicated"          "AllIdentical"           "AndersonDarlingTest"   
#>  [25] "aovlDetails"            "aovlErrorTerms"         "Append"                
#>  [28] "Arrow"                  "as.CDateFmt"            "as.fmt"                
#>  [31] "as.matrix.xtabs"        "AscToChar"              "Asp"                   
#>  [34] "Assocs"                 "Atkinson"               "AUC"                   
#>  [37] "AxisBreak"              "axTicks.Date"           "axTicks.POSIXct"       
#>  [40] "BarnardTest"            "BartelsRankTest"        "BarText"               
#>  [43] "BhapkarTest"            "BinomCI"                "BinomCIn"              
#>  [46] "BinomDiffCI"            "BinomRatioCI"           "BinToDec"              
#>  [49] "BinTree"                "BlankIfNA"              "BootCI"                
#>  [52] "BoxCox"                 "BoxCoxInv"              "BoxCoxLambda"          
#>  [55] "BoxedText"              "BreslowDayTest"         "BreuschGodfreyTest"    
#>  [58] "BrierScore"             "BubbleLegend"           "Canvas"                
#>  [61] "CartToPol"              "CartToSph"              "CatTable"              
#>  [64] "CCC"                    "CharToAsc"              "Clockwise"             
#>  [67] "Closest"                "CmToPts"                "CmykToCmy"             
#>  [70] "CmykToRgb"              "CmyToCmyk"              "Coalesce"              
#>  [73] "CochranArmitageTest"    "CochranQTest"           "CoefVar"               
#>  [76] "CohenD"                 "CohenKappa"             "CollapseTable"         
#>  [79] "ColorLegend"            "ColToGray"              "ColToGrey"             
#>  [82] "ColToHex"               "ColToHsv"               "ColToOpaque"           
#>  [85] "ColToRgb"               "ColumnWrap"             "CombN"                 
#>  [88] "CombPairs"              "CombSet"                "CompleteColumns"       
#>  [91] "ConDisPairs"            "Conf"                   "ConnLines"             
#>  [94] "ConoverTest"            "ContCoef"               "Contrasts"             
#>  [97] "ConvUnit"               "Cor"                    "CorCI"                 
#> [100] "CorPart"                "CorPolychor"            "CountCompCases"        
#> [103] "CountWorkDays"          "CourseData"             "Cov"                   
#> [106] "CramerV"                "CramerVonMisesTest"     "createCOMReference"    
#> [109] "CronbachAlpha"          "Cross"                  "CrossN"                
#> [112] "Cstat"                  "CutQ"                   "Day"                   
#> [115] "Day<-"                  "DB"                     "dBenf"                 
#> [118] "DecToBin"               "DecToHex"               "DecToOct"              
#> [121] "DegToRad"               "Desc"                   "DescToolsOptions"      
#> [124] "dExtrVal"               "dFrechet"               "dGenExtrVal"           
#> [127] "dGenPareto"             "dGompertz"              "dGumbel"               
#> [130] "DiffDays360"            "DigitSum"               "DivCoef"               
#> [133] "DivCoefMax"             "Divisors"               "dNegWeibull"           
#> [136] "DoBy"                   "DoCall"                 "dOrder"                
#> [139] "Dot"                    "DrawArc"                "DrawBand"              
#> [142] "DrawBezier"             "DrawCircle"             "DrawEllipse"           
#> [145] "DrawRegPolygon"         "dRevGumbel"             "dRevWeibull"           
#> [148] "Dummy"                  "DunnettTest"            "DunnTest"              
#> [151] "DurbinWatsonTest"       "Entropy"                "Eps"                   
#> [154] "ErrBars"                "EtaSq"                  "ExpFreq"               
#> [157] "Factorize"              "FctArgs"                "Fibonacci"             
#> [160] "FindColor"              "FindCorr"               "FindRProfile"          
#> [163] "FisherZ"                "FisherZInv"             "FixToTable"            
#> [166] "Flags"                  "Fmt"                    "Format"                
#> [169] "Frac"                   "Freq"                   "Freq2D"                
#> [172] "GCD"                    "GenRandGroups"          "GeomSn"                
#> [175] "GeomTrans"              "GetCalls"               "GetCurrPP"             
#> [178] "GetCurrWrd"             "GetCurrXL"              "GetNewPP"              
#> [181] "GetNewWrd"              "GetNewXL"               "Gini"                  
#> [184] "GiniSimpson"            "Gmean"                  "GoodmanKruskalGamma"   
#> [187] "GoodmanKruskalTau"      "Gsd"                    "GTest"                 
#> [190] "Herfindahl"             "HexToCol"               "HexToDec"              
#> [193] "HexToRgb"               "HighLow"                "Hmean"                 
#> [196] "HmsToSec"               "HodgesLehmann"          "HoeffD"                
#> [199] "HosmerLemeshowTest"     "HotellingsT2Test"       "Hour"                  
#> [202] "HuberM"                 "HunterGaston"           "ICC"                   
#> [205] "identify.formula"       "IdentifyA"              "Impute"                
#> [208] "ImputeKnn"              "InDots"                 "Interval"              
#> [211] "IPMT"                   "IQRw"                   "IRR"                   
#> [214] "IsDate"                 "IsDichotomous"          "IsEuclid"              
#> [217] "IsLeapYear"             "IsNumeric"              "IsOdd"                 
#> [220] "IsPrime"                "IsValidHwnd"            "IsWeekend"             
#> [223] "IsWhole"                "IsZero"                 "JarqueBeraTest"        
#> [226] "JonckheereTerpstraTest" "KappaM"                 "KendallTauA"           
#> [229] "KendallTauB"            "KendallW"               "Keywords"              
#> [232] "KrippAlpha"             "Kurt"                   "Label"                 
#> [235] "Label<-"                "Lambda"                 "Large"                 
#> [238] "LastDayOfMonth"         "Lc"                     "LCM"                   
#> [241] "LehmacherTest"          "LeveneTest"             "LillieTest"            
#> [244] "lines.Lc"               "lines.lm"               "lines.loess"           
#> [247] "lines.smooth.spline"    "lines.SmoothSpline"     "LineToUser"            
#> [250] "LinScale"               "LOCF"                   "LOF"                   
#> [253] "Logit"                  "LogitInv"               "LogSt"                 
#> [256] "LogStInv"               "LongToRgb"              "LsFct"                 
#> [259] "LsObj"                  "MAD"                    "MADCI"                 
#> [262] "MAE"                    "MAPE"                   "Mar"                   
#> [265] "Margins"                "MaxDigits"              "Mean"                  
#> [268] "MeanAD"                 "MeanCI"                 "MeanCIn"               
#> [271] "MeanDiffCI"             "MeanSE"                 "Median"                
#> [274] "MedianCI"               "Mgp"                    "MHChisqTest"           
#> [277] "Midx"                   "Minute"                 "MixColor"              
#> [280] "Mode"                   "ModSummary"             "Month"                 
#> [283] "MosesTest"              "MoveAvg"                "MSE"                   
#> [286] "MultinomCI"             "MultMerge"              "MutInf"                
#> [289] "N"                      "NAIf"                   "NAIfBlank"             
#> [292] "NAIfZero"               "Ndec"                   "NemenyiTest"           
#> [295] "NMAE"                   "NMSE"                   "Now"                   
#> [298] "NPV"                    "NPVFixBond"             "OctToDec"              
#> [301] "OddsRatio"              "OPR"                    "OrderMixed"            
#> [304] "ORToRelRisk"            "Outlier"                "Overlap"               
#> [307] "PageTest"               "PairApply"              "Pal"                   
#> [310] "ParseFormula"           "ParseSASDatalines"      "PartialSD"             
#> [313] "PasswordDlg"            "pBenf"                  "PDFManual"             
#> [316] "PearsonTest"            "PercentRank"            "PercTable"             
#> [319] "Permn"                  "pExtrVal"               "pFrechet"              
#> [322] "pGenExtrVal"            "pGenPareto"             "pGompertz"             
#> [325] "pGumbel"                "Phi"                    "Phrase"                
#> [328] "plot.bagplot"           "plot.Conf"              "plot.Lc"               
#> [331] "plot.Lclist"            "plot.palette"           "plot.PostHocTest"      
#> [334] "PlotACF"                "PlotArea"               "PlotBag"               
#> [337] "PlotBagPairs"           "PlotBinTree"            "PlotBubble"            
#> [340] "PlotCandlestick"        "PlotCashFlow"           "PlotCirc"              
#> [343] "PlotConDens"            "PlotCorr"               "PlotDot"               
#> [346] "PlotECDF"               "PlotFaces"              "PlotFdist"             
#> [349] "PlotFun"                "PlotGACF"               "PlotLinesA"            
#> [352] "PlotLog"                "PlotMarDens"            "PlotMiss"              
#> [355] "PlotMonth"              "PlotMosaic"             "PlotMultiDens"         
#> [358] "PlotPairs"              "PlotPolar"              "PlotPyramid"           
#> [361] "PlotQQ"                 "PlotTernary"            "PlotTreemap"           
#> [364] "PlotVenn"               "PlotViolin"             "PlotWeb"               
#> [367] "PMT"                    "pNegWeibull"            "PoissonCI"             
#> [370] "PolarGrid"              "PolToCart"              "pOrder"                
#> [373] "PostHocTest"            "power.chisq.test"       "PpAddSlide"            
#> [376] "PPMT"                   "PpPlot"                 "PpText"                
#> [379] "Prec"                   "pRevGumbel"             "pRevWeibull"           
#> [382] "Primes"                 "print.Assocs"           "print.Conf"            
#> [385] "print.CorPolychor"      "print.DunnTest"         "print.Freq"            
#> [388] "print.HoeffD"           "print.ICC"              "print.mtest"           
#> [391] "print.PercTable"        "print.PostHocTest"      "PseudoR2"              
#> [394] "PtInPoly"               "PtsToCm"                "qBenf"                 
#> [397] "qExtrVal"               "qFrechet"               "qGenExtrVal"           
#> [400] "qGenPareto"             "qGompertz"              "qGumbel"               
#> [403] "qNegWeibull"            "qRevGumbel"             "qRevGumbelExp"         
#> [406] "qRevWeibull"            "Quantile"               "QuantileCI"            
#> [409] "Quarter"                "Quot"                   "RadToDeg"              
#> [412] "Range"                  "Rank"                   "RBAL"                  
#> [415] "rBenf"                  "Recode"                 "Recycle"               
#> [418] "RelRisk"                "Rename"                 "reorder.factor"        
#> [421] "Rev"                    "RevCode"                "rExtrVal"              
#> [424] "rFrechet"               "RgbToCmy"               "RgbToCol"              
#> [427] "RgbToHex"               "RgbToLong"              "rGenExtrVal"           
#> [430] "rGenPareto"             "rGompertz"              "rGumbel"               
#> [433] "RMSE"                   "RndPairs"               "RndWord"               
#> [436] "rNegWeibull"            "RobScale"               "RomanToInt"            
#> [439] "rOrder"                 "Rosenbluth"             "Rotate"                
#> [442] "RoundTo"                "rRevGumbel"             "rRevWeibull"           
#> [445] "RSessionAlive"          "RTempdirAlive"          "RunsTest"              
#> [448] "Sample"                 "SampleTwins"            "SaveAs"                
#> [451] "ScheffeTest"            "SD"                     "Second"                
#> [454] "SecToHms"               "SendOutlookMail"        "Sens"                  
#> [457] "SetAlpha"               "SetNames"               "Shade"                 
#> [460] "ShapiroFranciaTest"     "SiegelTukeyRank"        "SiegelTukeyTest"       
#> [463] "SignTest"               "Skew"                   "SLN"                   
#> [466] "Small"                  "SMAPE"                  "SmoothSpline"          
#> [469] "Some"                   "SomersDelta"            "Sort"                  
#> [472] "SortMixed"              "SpearmanRho"            "Spec"                  
#> [475] "SphToCart"              "split.formula"          "SplitAt"               
#> [478] "SplitPath"              "SplitToCol"             "SpreadOut"             
#> [481] "Stamp"                  "StdCoef"                "Str"                   
#> [484] "StrAbbr"                "StrAlign"               "Strata"                
#> [487] "StrCap"                 "StrChop"                "StrCountW"             
#> [490] "StrDist"                "StrExtract"             "StrExtractBetween"     
#> [493] "StripAttr"              "StrIsNumeric"           "StrLeft"               
#> [496] "StrPad"                 "StrPos"                 "StrRev"                
#> [499] "StrRight"               "StrSpell"               "StrSplit"              
#> [502] "StrTrim"                "StrTrunc"               "StrVal"                
#> [505] "StuartMaxwellTest"      "StuartTauC"             "SYD"                   
#> [508] "SysInfo"                "TextContrastColor"      "TextToTable"           
#> [511] "TheilU"                 "Timezone"               "TitleRect"             
#> [514] "TMod"                   "Today"                  "ToLong"                
#> [517] "TOne"                   "ToWide"                 "ToWrd"                 
#> [520] "ToWrdB"                 "ToWrdPlot"              "ToXL"                  
#> [523] "Trim"                   "TschuprowT"             "TTestA"                
#> [526] "TukeyBiweight"          "TwoGroups"              "UncertCoef"            
#> [529] "UnirootAll"             "Unit"                   "Unit<-"                
#> [532] "Untable"                "Unwhich"                "VanWaerdenTest"        
#> [535] "Var"                    "VarCI"                  "VarTest"               
#> [538] "VecRot"                 "VecShift"               "VIF"                   
#> [541] "Vigenere"               "VonNeumannTest"         "Week"                  
#> [544] "Weekday"                "Winsorize"              "WithOptions"           
#> [547] "WoolfTest"              "WrdBookmark"            "WrdCaption"            
#> [550] "WrdCellRange"           "WrdDeleteBookmark"      "WrdFont"               
#> [553] "WrdFont<-"              "WrdFormatCells"         "WrdGoto"               
#> [556] "WrdInsertBookmark"      "WrdKill"                "WrdMergeCells"         
#> [559] "WrdOpenFile"            "WrdPageBreak"           "WrdParagraphFormat"    
#> [562] "WrdParagraphFormat<-"   "WrdPlot"                "WrdSaveAs"             
#> [565] "WrdStyle"               "WrdStyle<-"             "WrdTable"              
#> [568] "WrdTableBorders"        "WrdUpdateBookmark"      "XLCurrReg"             
#> [571] "XLDateToPOSIXct"        "XLGetRange"             "XLGetWorkbook"         
#> [574] "XLKill"                 "XLNamedReg"             "XLSaveAs"              
#> [577] "XLView"                 "Year"                   "YearDay"               
#> [580] "YearMonth"              "YTM"                    "YuenTTest"             
#> [583] "YuleQ"                  "YuleY"                  "ZeroIfNA"              
#> [586] "Zodiac"                 "ZTest"