Confidence Intervals for Multinomial Proportions

Confidence intervals for multinomial proportions are often approximated by single binomial confidence intervals, which might in practice often yield satisfying results, but is properly speaking not correct. This function calculates simultaneous confidence intervals for multinomial proportions either according to the methods of Sison and Glaz, Goodman, Wald, Wald with continuity correction or Wilson.

MultinomCI(x, conf.level = 0.95, sides = c("two.sided", "left", "right"),
           method = c("sisonglaz", "cplus1", "goodman", "wald", "waldcc", "wilson"))

Arguments

x
A vector of positive integers representing the number of occurrences of each class. The total number of samples equals the sum of such elements.

conf.level
confidence level, defaults to 0.95.

sides
a character string specifying the side of the confidence interval, must be one of "two.sided" (default), "left" or "right". You can specify just the initial letter. "left" would be analogue to a hypothesis of "greater" in a t.test.

method
character string specifing which method to use; can be one out of "sisonglaz", "cplus1", "goodman", "wald", "waldcc", "wilson". Method can be abbreviated. See details. Defaults to "sisonglaz".

Author

Pablo J. Villacorta Iglesias <pjvi@decsai.ugr.es>
Department of Computer Science and Artificial Intelligence, University of Granada (Spain) (Sison-Glaz)

Andri Signorell <andri@signorell.net> (Goodman, Wald, Wilson)

Details

Given a vector of observations with the number of samples falling in each class of a multinomial distribution, builds the simultaneous confidence intervals for the multinomial probabilities according to the method proposed by the mentioned authors. The R code for Sison and Glaz (1995) has been translated from thes SAS code written by May and Johnson (2000).
Some approaches for the confidence intervals can potentially yield negative results or values beyond 1. These would be reset such as not to exceed the range of [0, 1].

Value

A matrix with 3 columns:

est

estimate

lwr.ci

lower bound of the confidence interval

upr.ci

upper bound of the confidence interval

The number of rows correspond to the dimension of x.
References

Sison, C.P and Glaz, J. (1995) Simultaneous confidence intervals and sample size determination for multinomial proportions. Journal of the American Statistical Association, 90:366-369.

Glaz, J., Sison, C.P. (1999) Simultaneous confidence intervals for multinomial proportions. Journal of Statistical Planning and Inference 82:251-262.

May, W.L., Johnson, W.D.(2000) Constructing two-sided simultaneous confidence intervals for multinomial proportions for small counts in a large number of cells. Journal of Statistical Software 5(6) . Paper and code available at https://www.jstatsoft.org/v05/i06.

Goodman, L. A. (1965) On Simultaneous Confidence Intervals for Multinomial Proportions Technometrics, 7, 247-254.

Wald, A. Tests of statistical hypotheses concerning several parameters when the number of observations is large, Trans. Am. Math. Soc. 54 (1943) 426-482.

Wilson, E. B. Probable inference, the law of succession and statistical inference, J.Am. Stat. Assoc. 22 (1927) 209-212.

Examples

# Multinomial distribution with 3 classes, from which a sample of 79 elements
# were drawn: 23 of them belong to the first class, 12 to the
# second class and 44 to the third class. Punctual estimations
# of the probabilities from this sample would be 23/79, 12/79
# and 44/79 but we want to build 95% simultaneous confidence intervals
# for the true probabilities

MultinomCI(c(23, 12, 44), conf.level=0.95)

#>            est     lwr.ci    upr.ci
#> [1,] 0.2911392 0.18987342 0.4104183
#> [2,] 0.1518987 0.05063291 0.2711778
#> [3,] 0.5569620 0.45569620 0.6762410

x <- c(35, 74, 22, 69)

MultinomCI(x, method="goodman")

#>        est     lwr.ci    upr.ci
#> [1,] 0.175 0.11253215 0.2619106
#> [2,] 0.370 0.28113643 0.4686407
#> [3,] 0.110 0.06224338 0.1870880
#> [4,] 0.345 0.25846198 0.4431954

MultinomCI(x, method="sisonglaz")

#>        est lwr.ci    upr.ci
#> [1,] 0.175  0.105 0.2512563
#> [2,] 0.370  0.300 0.4462563
#> [3,] 0.110  0.040 0.1862563
#> [4,] 0.345  0.275 0.4212563

MultinomCI(x, method="cplus1")

#>        est lwr.ci upr.ci
#> [1,] 0.175  0.100  0.250
#> [2,] 0.370  0.295  0.445
#> [3,] 0.110  0.035  0.185
#> [4,] 0.345  0.270  0.420

MultinomCI(x, method="wald")

#>        est     lwr.ci    upr.ci
#> [1,] 0.175 0.12234021 0.2276598
#> [2,] 0.370 0.30308797 0.4369120
#> [3,] 0.110 0.06663649 0.1533635
#> [4,] 0.345 0.27911853 0.4108815

MultinomCI(x, method="waldcc")

#>        est     lwr.ci    upr.ci
#> [1,] 0.175 0.11984021 0.2301598
#> [2,] 0.370 0.30058797 0.4394120
#> [3,] 0.110 0.06413649 0.1558635
#> [4,] 0.345 0.27661853 0.4133815

MultinomCI(x, method="wilson")

#>        est     lwr.ci    upr.ci
#> [1,] 0.175 0.12860515 0.2336443
#> [2,] 0.370 0.30612608 0.4387737
#> [3,] 0.110 0.07377244 0.1609269
#> [4,] 0.345 0.28259794 0.4132441

# compare to
BinomCI(x, n=sum(x))

#>       est     lwr.ci    upr.ci
#> x.1 0.175 0.12860515 0.2336443
#> x.2 0.370 0.30612608 0.4387737
#> x.3 0.110 0.07377244 0.1609269
#> x.4 0.345 0.28259794 0.4132441