Distributions of Order Statistics

Density function, distribution function and random generation for a selected Order statistic of a given number of independent variables from a specified distribution.

dOrder(x, densfun, distnfun, ..., distn, mlen = 1, j = 1,
      largest = TRUE, log = FALSE)
pOrder(q, distnfun, ..., distn, mlen = 1, j = 1, largest = TRUE,
      lower.tail = TRUE)
rOrder(n, quantfun, ..., distn, mlen = 1, j = 1, largest = TRUE)

Arguments

x, q
Vector of quantiles.

n
Number of observations.

densfun, distnfun, quantfun
Density, distribution and quantile function of the specified distribution. The density function must have a log argument (a simple wrapper can always be constructed to achieve this).

...
Parameters of the specified distribution.

distn
A character string, optionally specified as an alternative to densfun, distnfun and quantfun such that the density, distribution and quantile functions are formed upon the addition of the prefixes d, p and q respectively.

mlen
The number of independent variables.

j
The Order statistic, taken as the jth largest (default) or smallest of mlen, according to the value of largest.

largest
Logical; if TRUE (default) use the jth largest Order statistic, otherwise use the jth smallest.

log
Logical; if TRUE, the log density is returned.

lower.tail
Logical; if TRUE (default) probabilities are P[X <= x], otherwise P[X > x].

Value

dOrder gives the density function, pOrder gives the distribution function and qOrder gives the quantile function of a selected Order statistic from a sample of size mlen, from a specified distibution. rOrder generates random deviates.

See also

Author

Alec Stephenson <alec_stephenson@hotmail.com>

Examples

dOrder(2:4, dnorm, pnorm, mean = 0.5, sd = 1.2, mlen = 5, j = 2)

#> [1] 0.2300687782 0.0133524232 0.0001663078

dOrder(2:4, distn = "norm", mean = 0.5, sd = 1.2, mlen = 5, j = 2)

#> [1] 0.2300687782 0.0133524232 0.0001663078

dOrder(2:4, distn = "exp", mlen = 2, j = 2)

#> [1] 0.0366312778 0.0049575044 0.0006709253

pOrder(2:4, distn = "exp", rate = 1.2, mlen = 2, j = 2)

#> [1] 0.9917703 0.9992534 0.9999323

rOrder(5, qgamma, shape = 1, mlen = 10, j = 2)

#> [1] 2.249505 1.141041 1.598620 2.612313 1.917155