Reorder the Levels of a Factor

Reorder the levels of a factor

# S3 method for factor
reorder(x, X, FUN, ..., 
        order = is.ordered(x), new.order, sort = SortMixed)

Arguments

x
factor

X
auxillary data vector

FUN
function to be applied to subsets of X determined by x, to determine factor order

...
optional parameters to FUN

order
logical value indicating whether the returned object should be an ordered factor

new.order
a vector of indexes or a vector of label names giving the order of the new factor levels

sort
function to use to sort the factor level names, used only when new.order is missing

Details

This function changes the order of the levels of a factor. It can do so via three different mechanisms, depending on whether, X and FUN, new.order or sort are provided.

If X and Fun are provided: The data in X is grouped by the levels of x and FUN is applied. The groups are then sorted by this value, and the resulting order is used for the new factor level names.

If new.order is provided: For a numeric vector, the new factor level names are constructed by reordering the factor levels according to the numeric values. For vectors, new.order gives the list of new factor level names. In either case levels omitted from new.order will become missing (NA) values.

If sort is provided (as it is by default): The new factor level names are generated by applying the supplied function to the existing factor level names. With sort=mixedsort the factor levels are sorted so that combined numeric and character strings are sorted in according to character rules on the character sections (including ignoring case), and the numeric rules for the numeric sections. See mixedsort for details.

Value

A new factor with reordered levels

Author

Gregory R. Warnes greg@warnes.net

See also

Examples

# Create a 4 level example factor
trt <- factor( sample( c("PLACEBO", "300 MG", "600 MG", "1200 MG"),
              100, replace=TRUE ) )
summary(trt)

#> 1200 MG  300 MG  600 MG PLACEBO 
#>      22      31      22      25 

# Note that the levels are not in a meaningful order.

# Change the order to something useful
# default "mixedsort" ordering
trt2 <- reorder(trt)
summary(trt2)

#>  300 MG  600 MG 1200 MG PLACEBO 
#>      31      22      22      25 

# using indexes:
trt3 <- reorder(trt, new.order=c(4, 2, 3, 1))
summary(trt3)

#> PLACEBO  300 MG  600 MG 1200 MG 
#>      25      31      22      22 

# using label names:
trt4 <- reorder(trt, new.order=c("PLACEBO", "300 MG", "600 MG", "1200 MG"))
summary(trt4)

#> PLACEBO  300 MG  600 MG 1200 MG 
#>      25      31      22      22 

# using frequency
trt5 <- reorder(trt, X=as.numeric(trt), FUN=length)
summary(trt5)

#> 1200 MG  600 MG PLACEBO  300 MG 
#>      22      22      25      31 

# drop out the '300 MG' level
trt6 <- reorder(trt, new.order=c("PLACEBO", "600 MG", "1200 MG"))
summary(trt6)

#> PLACEBO  600 MG 1200 MG    NA's 
#>      25      22      22      31