Find the Nearest Named R-Color to a Given RGB-Color

Converting a RGB-color to a named R-Color means looking for a color in the R-palette, which is nearest to the given RGB-color. This function uses the minimum of squared distance ("euclidean") or the minimum absolute distance ("manhattan") as proximity measure.
RgbToLong() converts a RGB-color to a long integer in numeric format. LongToRGB() does it the other way round.

RgbToCol(col, method = "rgb", metric = "euclidean")

RgbToLong(col)
LongToRgb(col)

Arguments

col
the color in rgb code, say a matrix with the red, green and blue code in the rows.

method
character string specifying the color space to be used. Can be "rgb" (default) or "hsv".

metric
character string specifying the metric to be used for calculating distances between the colors. Available options are "euclidean" (default) and "manhattan". Euclidean distances are root sum-of-squares of differences, and manhattan distances are the sum of absolute differences.

Details

It may not be clear from the start which method, rgb or hsv, yield the more natural results. Trying and comparing is a recommended strategy. Moreover the shortest numerical distance will not always be the best choice, when comparing the colours visually.

Value

the name of the nearest found R color.

Author

Andri Signorell <andri@signorell.net>

See also

Examples

RgbToCol(matrix(c(162,42,42), nrow=3))

#> [1] "brown"

RgbToLong(matrix(c(162,42,42), nrow=3))

#> [1] 2763426