Spread Out a Vector of Numbers To a Minimum Interval

Spread the numbers of a vector so that there is a minimum interval between any two numbers (in ascending or descending order). This is helpful when we want to place textboxes on a plot and ensure, that they do not mutually overlap.

SpreadOut(x, mindist = NULL, cex = 1.0)

Arguments

x
a numeric vector which may contain NAs.

mindist
the minimum interval between any two values. If this is left to NULL (default) the function will check if a plot is open and then use 90% of strheight().

cex
numeric character expansion factor; multiplied by par("cex") yields the final character size; the default NULL is equivalent to 1.

Details

SpreadOut() starts at or near the middle of the vector and increases the intervals between the ordered values. NAs are preserved. SpreadOut() first tries to spread groups of values with intervals less than mindist out neatly away from the mean of the group. If this doesn't entirely succeed, a second pass that forces values away from the middle is performed.

SpreadOut() can also be used to avoid overplotting of axis tick labels where they may be close together.

Value

On success, the spread out values. If there are less than two valid values, the original vector is returned.

Note

This function is based on plotrix::spreadout() and has been integrated here with some minor changes.

See also

Author

Jim Lemon <jim@bitwrit.com.au>
some extensions Andri Signorell <andri@signorell.net>

Examples

SpreadOut(c(1, 3, 3, 3, 3, 5), 0.2)

#> [1] 1.0 2.7 2.9 3.1 3.3 5.0

SpreadOut(c(1, 2.5, 2.5, 3.5, 3.5, 5), 0.2)

#> [1] 1.0 2.4 2.6 3.4 3.6 5.0

SpreadOut(c(5, 2.5, 2.5, NA, 3.5, 1, 3.5, NA), 0.2)

#> [1] 5.0 2.4 2.6  NA 3.4 1.0 3.6  NA

# this will almost always invoke the brute force second pass
SpreadOut(rnorm(10), 0.5)

#>  [1] -2.40347214  2.09652786 -0.90347214  1.09652786  0.59652786 -1.90347214
#>  [7] -1.40347214 -0.40347214  0.09652786  1.59652786