Stratified Sampling

Stratified sampling with equal/unequal probabilities.

Strata(x, stratanames = NULL, size,
       method = c("srswor", "srswr", "poisson", "systematic"),
       pik, description = FALSE)

Arguments

x
a data frame or a matrix; its number of rows is n, the population size.

stratanames
vector of stratification variables.

size
vector of stratum sample sizes (in the order in which the strata are given in the input data set).

method
method to select units; implemented are: a) simple random sampling without replacement ("srswor"), b) simple random sampling with replacement ("srswr"), c) Poisson sampling ("poisson"), d) systematic sampling ("systematic") (default is "srswor").

pik
vector of inclusion probabilities or auxiliary information used to compute them; this argument is only used for unequal probability sampling (Poisson and systematic). If an auxiliary information is provided, the function uses the inclusionprobabilities function for computing these probabilities. If the method is "srswr" and the sample size is larger than the population size, this vector is normalized to one.

description
a message is printed if its value is TRUE; the message gives the number of selected units and the number of the units in the population. By default, the value is FALSE.

Value

The function produces an object, which contains the following information:

id

the identifier of the selected units.

stratum

the unit stratum.

prob

the final unit inclusion probability.

See also

Examples

# Example from An and Watts (New SAS procedures for Analysis of Sample Survey Data)
# generates artificial data (a 235X3 matrix with 3 columns: state, region, income).
# the variable "state" has 2 categories ('nc' and 'sc').
# the variable "region" has 3 categories (1, 2 and 3).
# the sampling frame is stratified by region within state.
# the income variable is randomly generated

m <- rbind(matrix(rep("nc",165), 165, 1, byrow=TRUE),
           matrix(rep("sc", 70), 70, 1, byrow=TRUE))
m <- cbind.data.frame(m, c(rep(1, 100), rep(2,50), rep(3,15),
                      rep(1, 30), rep(2, 40)), 1000 * runif(235))
names(m) <- c("state", "region", "income")

# computes the population stratum sizes
table(m$region, m$state)

#>    
#>      nc  sc
#>   1 100  30
#>   2  50  40
#>   3  15   0

# not run
#     nc  sc
#  1 100  30
#  2  50  40
#  3  15   0
# there are 5 cells with non-zero values
# one draws 5 samples (1 sample in each stratum)
# the sample stratum sizes are 10,5,10,4,6, respectively
# the method is 'srswor' (equal probability, without replacement)

s <- Strata(m, c("region", "state"), size=c(10, 5, 10, 4, 6), method="srswor")

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>          income state region  income.1 stratum size  id
#> 1.100 142.10787    nc      1 142.10787       1   10 100
#> 1.6   118.74966    nc      1 118.74966       1   10   6
#> 1.88  940.18472    nc      1 940.18472       1   10  88
#> 1.37  476.75226    nc      1 476.75226       1   10  37
#> 1.61  116.64964    nc      1 116.64964       1   10  61
#> 1.66  815.06143    nc      1 815.06143       1   10  66
#> 1.19   30.95652    nc      1  30.95652       1   10  19
#> 1.46  269.25720    nc      1 269.25720       1   10  46
#> 1.69  269.63662    nc      1 269.63662       1   10  69
#> 1.34  750.08931    nc      1 750.08931       1   10  34
#> 2.118 933.18496    nc      2 933.18496       2    5 118
#> 2.136 222.53513    nc      2 222.53513       2    5 136
#> 2.115 347.76692    nc      2 347.76692       2    5 115
#> 2.121 202.05867    nc      2 202.05867       2    5 121
#> 2.123  23.17701    nc      2  23.17701       2    5 123
#> 3.158 963.06344    nc      3 963.06344       3   10 158
#> 3.156 414.33498    nc      3 414.33498       3   10 156
#> 3.162 924.67107    nc      3 924.67107       3   10 162
#> 3.163 470.69327    nc      3 470.69327       3   10 163
#> 3.159 312.36488    nc      3 312.36488       3   10 159
#> 3.151 920.57039    nc      3 920.57039       3   10 151
#> 3.161 656.02706    nc      3 656.02706       3   10 161
#> 3.164  62.96757    nc      3  62.96757       3   10 164
#> 3.153 750.30281    nc      3 750.30281       3   10 153
#> 3.160 596.18182    nc      3 596.18182       3   10 160
#> 4.183 292.59138    sc      1 292.59138       4    4 183
#> 4.182 722.01772    sc      1 722.01772       4    4 182
#> 4.195 195.05467    sc      1 195.05467       4    4 195
#> 4.166 742.18845    sc      1 742.18845       4    4 166
#> 5.212 140.30218    sc      2 140.30218       5    6 212
#> 5.199 458.56342    sc      2 458.56342       5    6 199
#> 5.197 580.75133    sc      2 580.75133       5    6 197
#> 5.235 105.96198    sc      2 105.96198       5    6 235
#> 5.220 594.92260    sc      2 594.92260       5    6 220
#> 5.201 718.81226    sc      2 718.81226       5    6 201

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0

# The same data as in Example 1
# the method is 'systematic' (unequal probability, without replacement)
# the selection probabilities are computed using the variable 'income'
s <- Strata(m,c("region", "state"), size=c(10, 5, 10, 4, 6),
            method="systematic", pik=m$income)

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>          income state region  income.1 stratum size  id
#> 1.25  124.96221    nc      1 124.96221       1   10  25
#> 1.59  449.16120    nc      1 449.16120       1   10  59
#> 1.18  656.25815    nc      1 656.25815       1   10  18
#> 1.50  371.04978    nc      1 371.04978       1   10  50
#> 1.24  935.14601    nc      1 935.14601       1   10  24
#> 1.74  187.67062    nc      1 187.67062       1   10  74
#> 1.69  269.63662    nc      1 269.63662       1   10  69
#> 1.16  907.12184    nc      1 907.12184       1   10  16
#> 1.22  785.07720    nc      1 785.07720       1   10  22
#> 1.15  915.77003    nc      1 915.77003       1   10  15
#> 2.148 136.49195    nc      2 136.49195       2    5 148
#> 2.119 759.79573    nc      2 759.79573       2    5 119
#> 2.104 750.36523    nc      2 750.36523       2    5 104
#> 2.122 415.22721    nc      2 415.22721       2    5 122
#> 2.114 240.84917    nc      2 240.84917       2    5 114
#> 3.154  67.95241    nc      3  67.95241       3   10 154
#> 3.151 920.57039    nc      3 920.57039       3   10 151
#> 3.164  62.96757    nc      3  62.96757       3   10 164
#> 3.159 312.36488    nc      3 312.36488       3   10 159
#> 3.163 470.69327    nc      3 470.69327       3   10 163
#> 3.158 963.06344    nc      3 963.06344       3   10 158
#> 3.153 750.30281    nc      3 750.30281       3   10 153
#> 3.157 330.49882    nc      3 330.49882       3   10 157
#> 3.162 924.67107    nc      3 924.67107       3   10 162
#> 3.152 873.39318    nc      3 873.39318       3   10 152
#> 4.171 126.99791    sc      1 126.99791       4    4 171
#> 4.175 585.83823    sc      1 585.83823       4    4 175
#> 4.180 703.93278    sc      1 703.93278       4    4 180
#> 4.179 807.63759    sc      1 807.63759       4    4 179
#> 5.214 561.47174    sc      2 561.47174       5    6 214
#> 5.217 244.49491    sc      2 244.49491       5    6 217
#> 5.211 961.49881    sc      2 961.49881       5    6 211
#> 5.202 707.76068    sc      2 707.76068       5    6 202
#> 5.221 354.30801    sc      2 354.30801       5    6 221
#> 5.219 728.44139    sc      2 728.44139       5    6 219

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0