Stratified Sampling

Stratified sampling with equal/unequal probabilities.

Strata(x, stratanames = NULL, size,
       method = c("srswor", "srswr", "poisson", "systematic"),
       pik, description = FALSE)

Arguments

x
a data frame or a matrix; its number of rows is n, the population size.

stratanames
vector of stratification variables.

size
vector of stratum sample sizes (in the order in which the strata are given in the input data set).

method
method to select units; implemented are: a) simple random sampling without replacement ("srswor"), b) simple random sampling with replacement ("srswr"), c) Poisson sampling ("poisson"), d) systematic sampling ("systematic") (default is "srswor").

pik
vector of inclusion probabilities or auxiliary information used to compute them; this argument is only used for unequal probability sampling (Poisson and systematic). If an auxiliary information is provided, the function uses the inclusionprobabilities function for computing these probabilities. If the method is "srswr" and the sample size is larger than the population size, this vector is normalized to one.

description
a message is printed if its value is TRUE; the message gives the number of selected units and the number of the units in the population. By default, the value is FALSE.

Value

The function produces an object, which contains the following information:

id

the identifier of the selected units.

stratum

the unit stratum.

prob

the final unit inclusion probability.

See also

Examples

# Example from An and Watts (New SAS procedures for Analysis of Sample Survey Data)
# generates artificial data (a 235X3 matrix with 3 columns: state, region, income).
# the variable "state" has 2 categories ('nc' and 'sc').
# the variable "region" has 3 categories (1, 2 and 3).
# the sampling frame is stratified by region within state.
# the income variable is randomly generated

m <- rbind(matrix(rep("nc",165), 165, 1, byrow=TRUE),
           matrix(rep("sc", 70), 70, 1, byrow=TRUE))
m <- cbind.data.frame(m, c(rep(1, 100), rep(2,50), rep(3,15),
                      rep(1, 30), rep(2, 40)), 1000 * runif(235))
names(m) <- c("state", "region", "income")

# computes the population stratum sizes
table(m$region, m$state)

#>    
#>      nc  sc
#>   1 100  30
#>   2  50  40
#>   3  15   0

# not run
#     nc  sc
#  1 100  30
#  2  50  40
#  3  15   0
# there are 5 cells with non-zero values
# one draws 5 samples (1 sample in each stratum)
# the sample stratum sizes are 10,5,10,4,6, respectively
# the method is 'srswor' (equal probability, without replacement)

s <- Strata(m, c("region", "state"), size=c(10, 5, 10, 4, 6), method="srswor")

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>          income state region  income.1 stratum size  id
#> 1.98   80.87036    nc      1  80.87036       1   10  98
#> 1.13   44.13647    nc      1  44.13647       1   10  13
#> 1.9   440.75781    nc      1 440.75781       1   10   9
#> 1.29  655.96816    nc      1 655.96816       1   10  29
#> 1.86   20.54442    nc      1  20.54442       1   10  86
#> 1.16  542.45146    nc      1 542.45146       1   10  16
#> 1.8   482.35179    nc      1 482.35179       1   10   8
#> 1.36  572.70453    nc      1 572.70453       1   10  36
#> 1.92  253.80529    nc      1 253.80529       1   10  92
#> 1.100 533.04155    nc      1 533.04155       1   10 100
#> 2.102 718.01602    nc      2 718.01602       2    5 102
#> 2.118 348.99159    nc      2 348.99159       2    5 118
#> 2.145 959.95646    nc      2 959.95646       2    5 145
#> 2.135 135.61091    nc      2 135.61091       2    5 135
#> 2.112 230.09193    nc      2 230.09193       2    5 112
#> 3.163 388.30586    nc      3 388.30586       3   10 163
#> 3.153 662.91062    nc      3 662.91062       3   10 153
#> 3.157  36.66020    nc      3  36.66020       3   10 157
#> 3.160 555.29766    nc      3 555.29766       3   10 160
#> 3.152 642.68356    nc      3 642.68356       3   10 152
#> 3.156 270.36405    nc      3 270.36405       3   10 156
#> 3.158 836.54641    nc      3 836.54641       3   10 158
#> 3.155 975.16088    nc      3 975.16088       3   10 155
#> 3.161 483.59321    nc      3 483.59321       3   10 161
#> 3.151 394.08056    nc      3 394.08056       3   10 151
#> 4.183 368.00581    sc      1 368.00581       4    4 183
#> 4.176 991.83105    sc      1 991.83105       4    4 176
#> 4.193 894.79894    sc      1 894.79894       4    4 193
#> 4.195 548.44081    sc      1 548.44081       4    4 195
#> 5.232 704.64299    sc      2 704.64299       5    6 232
#> 5.225  80.48474    sc      2  80.48474       5    6 225
#> 5.214 520.96277    sc      2 520.96277       5    6 214
#> 5.200 290.48006    sc      2 290.48006       5    6 200
#> 5.233 569.91573    sc      2 569.91573       5    6 233
#> 5.229 644.60175    sc      2 644.60175       5    6 229

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0

# The same data as in Example 1
# the method is 'systematic' (unequal probability, without replacement)
# the selection probabilities are computed using the variable 'income'
s <- Strata(m,c("region", "state"), size=c(10, 5, 10, 4, 6),
            method="systematic", pik=m$income)

# extracts the observed data
data.frame(income=m[s$id, "income"], s)

#>          income state region  income.1 stratum size  id
#> 1.67  563.19057    nc      1 563.19057       1   10  67
#> 1.93  571.54761    nc      1 571.54761       1   10  93
#> 1.46  550.59951    nc      1 550.59951       1   10  46
#> 1.50  333.65031    nc      1 333.65031       1   10  50
#> 1.10  553.89989    nc      1 553.89989       1   10  10
#> 1.73  336.51893    nc      1 336.51893       1   10  73
#> 1.81  665.81094    nc      1 665.81094       1   10  81
#> 1.3    55.82118    nc      1  55.82118       1   10   3
#> 1.13   44.13647    nc      1  44.13647       1   10  13
#> 1.56  592.34629    nc      1 592.34629       1   10  56
#> 2.108 737.76557    nc      2 737.76557       2    5 108
#> 2.106 232.86548    nc      2 232.86548       2    5 106
#> 2.142 595.64563    nc      2 595.64563       2    5 142
#> 2.140 249.62570    nc      2 249.62570       2    5 140
#> 2.101 307.53812    nc      2 307.53812       2    5 101
#> 3.159 888.16006    nc      3 888.16006       3   10 159
#> 3.157  36.66020    nc      3  36.66020       3   10 157
#> 3.152 642.68356    nc      3 642.68356       3   10 152
#> 3.161 483.59321    nc      3 483.59321       3   10 161
#> 3.151 394.08056    nc      3 394.08056       3   10 151
#> 3.165 773.62990    nc      3 773.62990       3   10 165
#> 3.160 555.29766    nc      3 555.29766       3   10 160
#> 3.154 835.90480    nc      3 835.90480       3   10 154
#> 3.156 270.36405    nc      3 270.36405       3   10 156
#> 3.153 662.91062    nc      3 662.91062       3   10 153
#> 4.182 580.40832    sc      1 580.40832       4    4 182
#> 4.171 152.05055    sc      1 152.05055       4    4 171
#> 4.174 414.26735    sc      1 414.26735       4    4 174
#> 4.191 180.95308    sc      1 180.95308       4    4 191
#> 5.224 135.38640    sc      2 135.38640       5    6 224
#> 5.206 814.83794    sc      2 814.83794       5    6 206
#> 5.231 344.82056    sc      2 344.82056       5    6 231
#> 5.229 644.60175    sc      2 644.60175       5    6 229
#> 5.203 776.36402    sc      2 776.36402       5    6 203
#> 5.210 663.38892    sc      2 663.38892       5    6 210

# see the result using a contigency table
table(s$region, s$state)

#>    
#>     nc sc
#>   1 10  4
#>   2  5  6
#>   3 10  0